% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actuator_api.R
\docType{class}
\name{ActuatorApi}
\alias{ActuatorApi}
\title{Actuator operations}
\format{
An \code{R6Class} generator object
}
\description{
ActuatorApi
}
\details{
SIRIUS Nightsky API

REST API that provides the full functionality of SIRIUS and its web services as background service. It is intended as entry-point for scripting languages and software integration SDKs.This API is exposed by SIRIUS 6

The version of the OpenAPI document: 3.2
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  Health  ####################

library(RSirius)

#Actuator web endpoint 'health'
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$Health(data_file = "result.txt")
result <- api_instance$actuator_api$Health()
dput(result)


####################  Shutdown  ####################

library(RSirius)

#Actuator web endpoint 'shutdown'
api_instance <- rsirius_api$new()

api_instance$actuator_api$Shutdown()


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ActuatorApi-new}{\code{ActuatorApi$new()}}
\item \href{#method-ActuatorApi-Health}{\code{ActuatorApi$Health()}}
\item \href{#method-ActuatorApi-HealthWithHttpInfo}{\code{ActuatorApi$HealthWithHttpInfo()}}
\item \href{#method-ActuatorApi-Shutdown}{\code{ActuatorApi$Shutdown()}}
\item \href{#method-ActuatorApi-ShutdownWithHttpInfo}{\code{ActuatorApi$ShutdownWithHttpInfo()}}
\item \href{#method-ActuatorApi-clone}{\code{ActuatorApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActuatorApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-ActuatorApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ActuatorApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActuatorApi-Health"></a>}}
\if{latex}{\out{\hypertarget{method-ActuatorApi-Health}{}}}
\subsection{Method \code{Health()}}{
Actuator web endpoint 'health'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorApi$Health(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActuatorApi-HealthWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ActuatorApi-HealthWithHttpInfo}{}}}
\subsection{Method \code{HealthWithHttpInfo()}}{
Actuator web endpoint 'health'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorApi$HealthWithHttpInfo(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (object) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActuatorApi-Shutdown"></a>}}
\if{latex}{\out{\hypertarget{method-ActuatorApi-Shutdown}{}}}
\subsection{Method \code{Shutdown()}}{
Actuator web endpoint 'shutdown'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorApi$Shutdown(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActuatorApi-ShutdownWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ActuatorApi-ShutdownWithHttpInfo}{}}}
\subsection{Method \code{ShutdownWithHttpInfo()}}{
Actuator web endpoint 'shutdown'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorApi$ShutdownWithHttpInfo(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActuatorApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ActuatorApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
