% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligned_feature_opt_field.R
\docType{class}
\name{AlignedFeatureOptField}
\alias{AlignedFeatureOptField}
\title{AlignedFeatureOptField}
\format{
An \code{R6Class} generator object
}
\description{
AlignedFeatureOptField Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AlignedFeatureOptField-new}{\code{AlignedFeatureOptField$new()}}
\item \href{#method-AlignedFeatureOptField-toJSON}{\code{AlignedFeatureOptField$toJSON()}}
\item \href{#method-AlignedFeatureOptField-fromJSON}{\code{AlignedFeatureOptField$fromJSON()}}
\item \href{#method-AlignedFeatureOptField-toJSONString}{\code{AlignedFeatureOptField$toJSONString()}}
\item \href{#method-AlignedFeatureOptField-fromJSONString}{\code{AlignedFeatureOptField$fromJSONString()}}
\item \href{#method-AlignedFeatureOptField-clone}{\code{AlignedFeatureOptField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureOptField-new"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureOptField-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AlignedFeatureOptField class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureOptField$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureOptField-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureOptField-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureOptField$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AlignedFeatureOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureOptField-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureOptField-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of AlignedFeatureOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureOptField$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AlignedFeatureOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureOptField-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureOptField-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureOptField$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AlignedFeatureOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureOptField-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureOptField-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of AlignedFeatureOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureOptField$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AlignedFeatureOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureOptField-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureOptField-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureOptField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
