% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligned_feature_quality.R
\docType{class}
\name{AlignedFeatureQuality}
\alias{AlignedFeatureQuality}
\title{AlignedFeatureQuality}
\format{
An \code{R6Class} generator object
}
\description{
AlignedFeatureQuality Class

AlignedFeatureQuality Class
}
\details{
Create a new AlignedFeatureQuality
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alignedFeatureId}}{Id of the feature (aligned over runs) this quality information belongs to. character}

\item{\code{overallQuality}}{\link{DataQuality}}

\item{\code{categories}}{Contains all pre-computation quality information that belong to  this feature (aligned over runs), such as information about the quality of the peak shape, MS2 spectrum etc., named list(\link{Category})}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AlignedFeatureQuality-new}{\code{AlignedFeatureQuality$new()}}
\item \href{#method-AlignedFeatureQuality-toJSON}{\code{AlignedFeatureQuality$toJSON()}}
\item \href{#method-AlignedFeatureQuality-fromJSON}{\code{AlignedFeatureQuality$fromJSON()}}
\item \href{#method-AlignedFeatureQuality-toJSONString}{\code{AlignedFeatureQuality$toJSONString()}}
\item \href{#method-AlignedFeatureQuality-fromJSONString}{\code{AlignedFeatureQuality$fromJSONString()}}
\item \href{#method-AlignedFeatureQuality-validateJSON}{\code{AlignedFeatureQuality$validateJSON()}}
\item \href{#method-AlignedFeatureQuality-toString}{\code{AlignedFeatureQuality$toString()}}
\item \href{#method-AlignedFeatureQuality-isValid}{\code{AlignedFeatureQuality$isValid()}}
\item \href{#method-AlignedFeatureQuality-getInvalidFields}{\code{AlignedFeatureQuality$getInvalidFields()}}
\item \href{#method-AlignedFeatureQuality-print}{\code{AlignedFeatureQuality$print()}}
\item \href{#method-AlignedFeatureQuality-clone}{\code{AlignedFeatureQuality$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureQuality-new"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureQuality-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AlignedFeatureQuality class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureQuality$new(alignedFeatureId, overallQuality, categories, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alignedFeatureId}}{Id of the feature (aligned over runs) this quality information belongs to.}

\item{\code{overallQuality}}{overallQuality}

\item{\code{categories}}{Contains all pre-computation quality information that belong to  this feature (aligned over runs), such as information about the quality of the peak shape, MS2 spectrum etc.,}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureQuality-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureQuality-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureQuality$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AlignedFeatureQuality in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureQuality-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureQuality-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of AlignedFeatureQuality
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureQuality$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AlignedFeatureQuality
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureQuality-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureQuality-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureQuality$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AlignedFeatureQuality in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureQuality-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureQuality-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of AlignedFeatureQuality
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureQuality$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AlignedFeatureQuality
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureQuality-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureQuality-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to AlignedFeatureQuality and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureQuality$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureQuality-toString"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureQuality-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureQuality$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of AlignedFeatureQuality
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureQuality-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureQuality-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureQuality$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureQuality-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureQuality-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureQuality$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureQuality-print"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureQuality-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureQuality$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignedFeatureQuality-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AlignedFeatureQuality-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignedFeatureQuality$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
