% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allowed_features.R
\docType{class}
\name{AllowedFeatures}
\alias{AllowedFeatures}
\title{AllowedFeatures}
\format{
An \code{R6Class} generator object
}
\description{
AllowedFeatures Class

AllowedFeatures Class
}
\details{
Create a new AllowedFeatures
}
\examples{

## ------------------------------------------------
## Method `AllowedFeatures$toList`
## ------------------------------------------------

# convert array of AllowedFeatures (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cli}}{character [optional]}

\item{\code{api}}{character [optional]}

\item{\code{deNovo}}{character [optional]}

\item{\code{importMSRuns}}{character [optional]}

\item{\code{importPeakLists}}{character [optional]}

\item{\code{importCef}}{character [optional]}

\item{\code{transformationProducts}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AllowedFeatures-new}{\code{AllowedFeatures$new()}}
\item \href{#method-AllowedFeatures-toJSON}{\code{AllowedFeatures$toJSON()}}
\item \href{#method-AllowedFeatures-toList}{\code{AllowedFeatures$toList()}}
\item \href{#method-AllowedFeatures-toSimpleType}{\code{AllowedFeatures$toSimpleType()}}
\item \href{#method-AllowedFeatures-fromJSON}{\code{AllowedFeatures$fromJSON()}}
\item \href{#method-AllowedFeatures-toJSONString}{\code{AllowedFeatures$toJSONString()}}
\item \href{#method-AllowedFeatures-fromJSONString}{\code{AllowedFeatures$fromJSONString()}}
\item \href{#method-AllowedFeatures-validateJSON}{\code{AllowedFeatures$validateJSON()}}
\item \href{#method-AllowedFeatures-toString}{\code{AllowedFeatures$toString()}}
\item \href{#method-AllowedFeatures-isValid}{\code{AllowedFeatures$isValid()}}
\item \href{#method-AllowedFeatures-getInvalidFields}{\code{AllowedFeatures$getInvalidFields()}}
\item \href{#method-AllowedFeatures-print}{\code{AllowedFeatures$print()}}
\item \href{#method-AllowedFeatures-clone}{\code{AllowedFeatures$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllowedFeatures-new"></a>}}
\if{latex}{\out{\hypertarget{method-AllowedFeatures-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AllowedFeatures class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllowedFeatures$new(
  cli = NULL,
  api = NULL,
  deNovo = NULL,
  importMSRuns = NULL,
  importPeakLists = NULL,
  importCef = NULL,
  transformationProducts = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cli}}{cli}

\item{\code{api}}{api}

\item{\code{deNovo}}{deNovo}

\item{\code{importMSRuns}}{importMSRuns}

\item{\code{importPeakLists}}{importPeakLists}

\item{\code{importCef}}{importCef}

\item{\code{transformationProducts}}{transformationProducts}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllowedFeatures-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AllowedFeatures-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllowedFeatures$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllowedFeatures-toList"></a>}}
\if{latex}{\out{\hypertarget{method-AllowedFeatures-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllowedFeatures$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AllowedFeatures as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of AllowedFeatures (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllowedFeatures-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-AllowedFeatures-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert AllowedFeatures to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllowedFeatures$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllowedFeatures-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AllowedFeatures-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of AllowedFeatures
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllowedFeatures$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AllowedFeatures
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllowedFeatures-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AllowedFeatures-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllowedFeatures$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AllowedFeatures in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllowedFeatures-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AllowedFeatures-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of AllowedFeatures
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllowedFeatures$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AllowedFeatures
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllowedFeatures-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AllowedFeatures-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to AllowedFeatures and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllowedFeatures$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllowedFeatures-toString"></a>}}
\if{latex}{\out{\hypertarget{method-AllowedFeatures-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllowedFeatures$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of AllowedFeatures
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllowedFeatures-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-AllowedFeatures-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllowedFeatures$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllowedFeatures-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-AllowedFeatures-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllowedFeatures$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllowedFeatures-print"></a>}}
\if{latex}{\out{\hypertarget{method-AllowedFeatures-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllowedFeatures$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllowedFeatures-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AllowedFeatures-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllowedFeatures$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
