% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_peak.R
\docType{class}
\name{AnnotatedPeak}
\alias{AnnotatedPeak}
\title{AnnotatedPeak}
\format{
An \code{R6Class} generator object
}
\description{
AnnotatedPeak Class

AnnotatedPeak Class
}
\details{
Create a new AnnotatedPeak
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mz}}{numeric [optional]}

\item{\code{intensity}}{numeric [optional]}

\item{\code{peakAnnotation}}{\link{PeakAnnotation} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnnotatedPeak-new}{\code{AnnotatedPeak$new()}}
\item \href{#method-AnnotatedPeak-toJSON}{\code{AnnotatedPeak$toJSON()}}
\item \href{#method-AnnotatedPeak-fromJSON}{\code{AnnotatedPeak$fromJSON()}}
\item \href{#method-AnnotatedPeak-toJSONString}{\code{AnnotatedPeak$toJSONString()}}
\item \href{#method-AnnotatedPeak-fromJSONString}{\code{AnnotatedPeak$fromJSONString()}}
\item \href{#method-AnnotatedPeak-validateJSON}{\code{AnnotatedPeak$validateJSON()}}
\item \href{#method-AnnotatedPeak-toString}{\code{AnnotatedPeak$toString()}}
\item \href{#method-AnnotatedPeak-isValid}{\code{AnnotatedPeak$isValid()}}
\item \href{#method-AnnotatedPeak-getInvalidFields}{\code{AnnotatedPeak$getInvalidFields()}}
\item \href{#method-AnnotatedPeak-print}{\code{AnnotatedPeak$print()}}
\item \href{#method-AnnotatedPeak-clone}{\code{AnnotatedPeak$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedPeak-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedPeak-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AnnotatedPeak class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedPeak$new(mz = NULL, intensity = NULL, peakAnnotation = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mz}}{mz}

\item{\code{intensity}}{intensity}

\item{\code{peakAnnotation}}{peakAnnotation}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedPeak-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedPeak-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedPeak$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AnnotatedPeak in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedPeak-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedPeak-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of AnnotatedPeak
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedPeak$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AnnotatedPeak
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedPeak-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedPeak-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedPeak$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AnnotatedPeak in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedPeak-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedPeak-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of AnnotatedPeak
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedPeak$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AnnotatedPeak
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedPeak-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedPeak-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to AnnotatedPeak and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedPeak$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedPeak-toString"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedPeak-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedPeak$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of AnnotatedPeak
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedPeak-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedPeak-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedPeak$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedPeak-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedPeak-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedPeak$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedPeak-print"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedPeak-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedPeak$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedPeak-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedPeak-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedPeak$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
