% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_value.R
\docType{class}
\name{AnyValue}
\alias{AnyValue}
\title{AnyValue}
\format{
An \code{R6Class} generator object
}
\description{
AnyValue Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{any_of}}{a list of object types defined in the anyOf schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnyValue-new}{\code{AnyValue$new()}}
\item \href{#method-AnyValue-fromJSONString}{\code{AnyValue$fromJSONString()}}
\item \href{#method-AnyValue-fromJSON}{\code{AnyValue$fromJSON()}}
\item \href{#method-AnyValue-toJSONString}{\code{AnyValue$toJSONString()}}
\item \href{#method-AnyValue-toJSON}{\code{AnyValue$toJSON()}}
\item \href{#method-AnyValue-validateJSON}{\code{AnyValue$validateJSON()}}
\item \href{#method-AnyValue-toString}{\code{AnyValue$toString()}}
\item \href{#method-AnyValue-print}{\code{AnyValue$print()}}
\item \href{#method-AnyValue-clone}{\code{AnyValue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyValue-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnyValue-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AnyValue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyValue$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the anyOf schemas: "character", "integer", "numeric"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyValue-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnyValue-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of AnyValue.
An alias to the method `fromJSON`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyValue$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of AnyValue.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyValue-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnyValue-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of AnyValue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyValue$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of AnyValue.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyValue-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnyValue-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize AnyValue to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyValue$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the AnyValue.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyValue-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnyValue-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize AnyValue to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyValue$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the AnyValue.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyValue-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnyValue-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to AnyValue and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyValue$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyValue-toString"></a>}}
\if{latex}{\out{\hypertarget{method-AnyValue-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyValue$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyValue-print"></a>}}
\if{latex}{\out{\hypertarget{method-AnyValue-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyValue$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyValue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnyValue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyValue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
