% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_response.R
\docType{class}
\name{ApiResponse}
\alias{ApiResponse}
\title{ApiResponse}
\format{
An \code{R6Class} generator object
}
\description{
ApiResponse Class
}
\details{
SIRIUS Nightsky API

REST API that provides the full functionality of SIRIUS and its web services as background service. It is intended as entry-point for scripting languages and software integration SDKs.This API is exposed by SIRIUS 6

The version of the OpenAPI document: 3.2
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content}}{The deserialized response body.}

\item{\code{response}}{The raw response from the endpoint.}

\item{\code{status_code}}{The HTTP response status code.}

\item{\code{status_code_desc}}{The brief description of the HTTP response status code.}

\item{\code{headers}}{The HTTP response headers.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApiResponse-new}{\code{ApiResponse$new()}}
\item \href{#method-ApiResponse-response_as_text}{\code{ApiResponse$response_as_text()}}
\item \href{#method-ApiResponse-clone}{\code{ApiResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApiResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ApiResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiResponse$new(
  content = NULL,
  response = NULL,
  status_code = NULL,
  status_code_desc = NULL,
  headers = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{The deserialized response body.}

\item{\code{response}}{The raw response from the endpoint.}

\item{\code{status_code}}{The HTTP response status code.}

\item{\code{status_code_desc}}{The brief description of the HTTP response status code.}

\item{\code{headers}}{The HTTP response headers.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiResponse-response_as_text"></a>}}
\if{latex}{\out{\hypertarget{method-ApiResponse-response_as_text}{}}}
\subsection{Method \code{response_as_text()}}{
The response is stored as a raw vector. Use this to access the response after
converting it to text. If the response is not coercible to text NA is returned.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiResponse$response_as_text(from_encoding = "", to_encoding = "UTF-8")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from_encoding}}{The encoding of the raw response.}

\item{\code{to_encoding}}{The target encoding of the return value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApiResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
