% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes.R
\docType{class}
\name{Axes}
\alias{Axes}
\title{Axes}
\format{
An \code{R6Class} generator object
}
\description{
Axes Class

Axes Class
}
\details{
Create a new Axes
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{scanNumber}}{list(integer) [optional]}

\item{\code{scanIds}}{list(character) [optional]}

\item{\code{retentionTimeInSeconds}}{list(numeric) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Axes-new}{\code{Axes$new()}}
\item \href{#method-Axes-toJSON}{\code{Axes$toJSON()}}
\item \href{#method-Axes-fromJSON}{\code{Axes$fromJSON()}}
\item \href{#method-Axes-toJSONString}{\code{Axes$toJSONString()}}
\item \href{#method-Axes-fromJSONString}{\code{Axes$fromJSONString()}}
\item \href{#method-Axes-validateJSON}{\code{Axes$validateJSON()}}
\item \href{#method-Axes-toString}{\code{Axes$toString()}}
\item \href{#method-Axes-isValid}{\code{Axes$isValid()}}
\item \href{#method-Axes-getInvalidFields}{\code{Axes$getInvalidFields()}}
\item \href{#method-Axes-print}{\code{Axes$print()}}
\item \href{#method-Axes-clone}{\code{Axes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axes-new"></a>}}
\if{latex}{\out{\hypertarget{method-Axes-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Axes class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axes$new(scanNumber = NULL, scanIds = NULL, retentionTimeInSeconds = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scanNumber}}{scanNumber}

\item{\code{scanIds}}{scanIds}

\item{\code{retentionTimeInSeconds}}{retentionTimeInSeconds}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axes-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Axes-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axes$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Axes in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axes-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Axes-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Axes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axes$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Axes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axes-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Axes-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axes$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Axes in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axes-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Axes-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Axes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axes$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Axes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axes-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Axes-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Axes and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axes$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axes-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Axes-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axes$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Axes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axes-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Axes-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axes$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axes-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Axes-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axes$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axes-print"></a>}}
\if{latex}{\out{\hypertarget{method-Axes-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axes$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Axes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
