% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_spectrum.R
\docType{class}
\name{BasicSpectrum}
\alias{BasicSpectrum}
\title{BasicSpectrum}
\format{
An \code{R6Class} generator object
}
\description{
BasicSpectrum Class

BasicSpectrum Class
}
\details{
Create a new BasicSpectrum
}
\examples{

## ------------------------------------------------
## Method `BasicSpectrum$toList`
## ------------------------------------------------

# convert array of BasicSpectrum (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Optional display name of this spectrum. character [optional]}

\item{\code{msLevel}}{MS level of the measured spectrum.  Artificial spectra with no msLevel (e.g. Simulated Isotope patterns) use null or zero integer [optional]}

\item{\code{collisionEnergy}}{Collision energy used for MS/MS spectra  Null for spectra where collision energy is not applicable character [optional]}

\item{\code{instrument}}{Instrument information. character [optional]}

\item{\code{precursorMz}}{Precursor m/z of the MS/MS spectrum  Null for spectra where precursor m/z is not applicable numeric [optional]}

\item{\code{scanNumber}}{Scan number of the spectrum.  Might be null for artificial spectra with no scan number (e.g. Simulated Isotope patterns or merged spectra) integer [optional]}

\item{\code{cosineQuery}}{True if spectrum is in cosine query normalized format.  Such spectrum is compatible with SpectralLibraryMatch peak assignments to reference spectra. character}

\item{\code{precursorPeak}}{\link{SimplePeak} [optional]}

\item{\code{peaks}}{The peaks of this spectrum which might contain additional annotations such as molecular formulas. list(\link{SimplePeak})}

\item{\code{absIntensityFactor}}{Factor to convert relative intensities to absolute intensities.  Might be null or 1 for spectra where absolute intensities are not available (e.g. artificial or merged spectra)  <p>  DEPRECATED: Spectra are always returned with raw intensities.  Use provided normalization factors to normalize on the fly. numeric [optional]}

\item{\code{maxNormFactor}}{Factor to convert absolute intensities to MAX norm. numeric [optional]}

\item{\code{sumNormFactor}}{Factor to convert absolute intensities to SUM norm. numeric [optional]}

\item{\code{l2NormFactor}}{Factor to convert absolute intensities to L2 (Euclidean) norm. numeric [optional]}

\item{\code{firstPeakNormFactor}}{Factor to convert absolute intensities to normalize intensities by first peak intensity. numeric [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BasicSpectrum-new}{\code{BasicSpectrum$new()}}
\item \href{#method-BasicSpectrum-toJSON}{\code{BasicSpectrum$toJSON()}}
\item \href{#method-BasicSpectrum-toList}{\code{BasicSpectrum$toList()}}
\item \href{#method-BasicSpectrum-toSimpleType}{\code{BasicSpectrum$toSimpleType()}}
\item \href{#method-BasicSpectrum-fromJSON}{\code{BasicSpectrum$fromJSON()}}
\item \href{#method-BasicSpectrum-toJSONString}{\code{BasicSpectrum$toJSONString()}}
\item \href{#method-BasicSpectrum-fromJSONString}{\code{BasicSpectrum$fromJSONString()}}
\item \href{#method-BasicSpectrum-validateJSON}{\code{BasicSpectrum$validateJSON()}}
\item \href{#method-BasicSpectrum-toString}{\code{BasicSpectrum$toString()}}
\item \href{#method-BasicSpectrum-isValid}{\code{BasicSpectrum$isValid()}}
\item \href{#method-BasicSpectrum-getInvalidFields}{\code{BasicSpectrum$getInvalidFields()}}
\item \href{#method-BasicSpectrum-print}{\code{BasicSpectrum$print()}}
\item \href{#method-BasicSpectrum-clone}{\code{BasicSpectrum$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicSpectrum-new"></a>}}
\if{latex}{\out{\hypertarget{method-BasicSpectrum-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BasicSpectrum class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpectrum$new(
  cosineQuery,
  peaks,
  name = NULL,
  msLevel = NULL,
  collisionEnergy = NULL,
  instrument = NULL,
  precursorMz = NULL,
  scanNumber = NULL,
  precursorPeak = NULL,
  absIntensityFactor = NULL,
  maxNormFactor = NULL,
  sumNormFactor = NULL,
  l2NormFactor = NULL,
  firstPeakNormFactor = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cosineQuery}}{True if spectrum is in cosine query normalized format.  Such spectrum is compatible with SpectralLibraryMatch peak assignments to reference spectra.}

\item{\code{peaks}}{The peaks of this spectrum which might contain additional annotations such as molecular formulas.}

\item{\code{name}}{Optional display name of this spectrum.}

\item{\code{msLevel}}{MS level of the measured spectrum.  Artificial spectra with no msLevel (e.g. Simulated Isotope patterns) use null or zero}

\item{\code{collisionEnergy}}{Collision energy used for MS/MS spectra  Null for spectra where collision energy is not applicable}

\item{\code{instrument}}{Instrument information.}

\item{\code{precursorMz}}{Precursor m/z of the MS/MS spectrum  Null for spectra where precursor m/z is not applicable}

\item{\code{scanNumber}}{Scan number of the spectrum.  Might be null for artificial spectra with no scan number (e.g. Simulated Isotope patterns or merged spectra)}

\item{\code{precursorPeak}}{precursorPeak}

\item{\code{absIntensityFactor}}{Factor to convert relative intensities to absolute intensities.  Might be null or 1 for spectra where absolute intensities are not available (e.g. artificial or merged spectra)  <p>  DEPRECATED: Spectra are always returned with raw intensities.  Use provided normalization factors to normalize on the fly.}

\item{\code{maxNormFactor}}{Factor to convert absolute intensities to MAX norm.}

\item{\code{sumNormFactor}}{Factor to convert absolute intensities to SUM norm.}

\item{\code{l2NormFactor}}{Factor to convert absolute intensities to L2 (Euclidean) norm.}

\item{\code{firstPeakNormFactor}}{Factor to convert absolute intensities to normalize intensities by first peak intensity.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicSpectrum-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BasicSpectrum-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpectrum$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicSpectrum-toList"></a>}}
\if{latex}{\out{\hypertarget{method-BasicSpectrum-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpectrum$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BasicSpectrum as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of BasicSpectrum (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicSpectrum-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-BasicSpectrum-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert BasicSpectrum to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpectrum$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicSpectrum-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BasicSpectrum-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of BasicSpectrum
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpectrum$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BasicSpectrum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicSpectrum-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BasicSpectrum-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpectrum$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
BasicSpectrum in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicSpectrum-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BasicSpectrum-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of BasicSpectrum
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpectrum$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BasicSpectrum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicSpectrum-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BasicSpectrum-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to BasicSpectrum and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpectrum$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicSpectrum-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BasicSpectrum-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpectrum$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of BasicSpectrum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicSpectrum-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-BasicSpectrum-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpectrum$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicSpectrum-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-BasicSpectrum-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpectrum$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicSpectrum-print"></a>}}
\if{latex}{\out{\hypertarget{method-BasicSpectrum-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpectrum$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicSpectrum-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BasicSpectrum-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpectrum$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
