% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_fingerprint.R
\docType{class}
\name{BinaryFingerprint}
\alias{BinaryFingerprint}
\title{BinaryFingerprint}
\format{
An \code{R6Class} generator object
}
\description{
BinaryFingerprint Class

BinaryFingerprint Class
}
\details{
Create a new BinaryFingerprint
}
\examples{

## ------------------------------------------------
## Method `BinaryFingerprint$toList`
## ------------------------------------------------

# convert array of BinaryFingerprint (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bitsSet}}{Array that contains all RELATIVE indices (masked FP) of bits that are set (are 1) list(integer) [optional]}

\item{\code{length}}{Size of the fingerprint (masked fp), e.g. to reconstruct the binary array from the array of set bits integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BinaryFingerprint-new}{\code{BinaryFingerprint$new()}}
\item \href{#method-BinaryFingerprint-toJSON}{\code{BinaryFingerprint$toJSON()}}
\item \href{#method-BinaryFingerprint-toList}{\code{BinaryFingerprint$toList()}}
\item \href{#method-BinaryFingerprint-toSimpleType}{\code{BinaryFingerprint$toSimpleType()}}
\item \href{#method-BinaryFingerprint-fromJSON}{\code{BinaryFingerprint$fromJSON()}}
\item \href{#method-BinaryFingerprint-toJSONString}{\code{BinaryFingerprint$toJSONString()}}
\item \href{#method-BinaryFingerprint-fromJSONString}{\code{BinaryFingerprint$fromJSONString()}}
\item \href{#method-BinaryFingerprint-validateJSON}{\code{BinaryFingerprint$validateJSON()}}
\item \href{#method-BinaryFingerprint-toString}{\code{BinaryFingerprint$toString()}}
\item \href{#method-BinaryFingerprint-isValid}{\code{BinaryFingerprint$isValid()}}
\item \href{#method-BinaryFingerprint-getInvalidFields}{\code{BinaryFingerprint$getInvalidFields()}}
\item \href{#method-BinaryFingerprint-print}{\code{BinaryFingerprint$print()}}
\item \href{#method-BinaryFingerprint-clone}{\code{BinaryFingerprint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFingerprint-new"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFingerprint-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BinaryFingerprint class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFingerprint$new(bitsSet = NULL, length = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bitsSet}}{Array that contains all RELATIVE indices (masked FP) of bits that are set (are 1)}

\item{\code{length}}{Size of the fingerprint (masked fp), e.g. to reconstruct the binary array from the array of set bits}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFingerprint-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFingerprint-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFingerprint$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFingerprint-toList"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFingerprint-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFingerprint$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BinaryFingerprint as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of BinaryFingerprint (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFingerprint-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFingerprint-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert BinaryFingerprint to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFingerprint$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFingerprint-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFingerprint-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of BinaryFingerprint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFingerprint$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BinaryFingerprint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFingerprint-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFingerprint-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFingerprint$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
BinaryFingerprint in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFingerprint-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFingerprint-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of BinaryFingerprint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFingerprint$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BinaryFingerprint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFingerprint-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFingerprint-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to BinaryFingerprint and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFingerprint$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFingerprint-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFingerprint-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFingerprint$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of BinaryFingerprint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFingerprint-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFingerprint-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFingerprint$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFingerprint-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFingerprint-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFingerprint$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFingerprint-print"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFingerprint-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFingerprint$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFingerprint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFingerprint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFingerprint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
