% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_transformer_parameters.R
\docType{class}
\name{BioTransformerParameters}
\alias{BioTransformerParameters}
\title{BioTransformerParameters}
\format{
An \code{R6Class} generator object
}
\description{
BioTransformerParameters Class

BioTransformerParameters Class
}
\details{
Create a new BioTransformerParameters
}
\examples{

## ------------------------------------------------
## Method `BioTransformerParameters$toList`
## ------------------------------------------------

# convert array of BioTransformerParameters (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cyp450Mode}}{Specify the Phase I/Cyp450 mode for all provided BioTransformerSequenceSteps. Will only be applied to Steps that  require the Cyp450 mode as parameter. Can be null in cases where only BioTransformerSequenceSteps are specified  that do not need the Cyp450 mode. character}

\item{\code{p2Mode}}{Specify the Phase II mode for all provided BioTransformerSequenceSteps. Will only be applied to Steps that  require the Phase II mode as parameter. Can be null in cases where only BioTransformerSequenceSteps are specified  that do not need the Phase II mode. character}

\item{\code{useDB}}{Specify whether structures should additionally be retrieved from the HMDB database. character [optional]}

\item{\code{bioTransformerSequenceSteps}}{Specify BioTransformerSequenceSteps to be applied to input structures. MultiStep MetabolicTransformations can  only be used as singletons (list size of one). list(\link{BioTransformerSequenceStep})}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BioTransformerParameters-new}{\code{BioTransformerParameters$new()}}
\item \href{#method-BioTransformerParameters-toJSON}{\code{BioTransformerParameters$toJSON()}}
\item \href{#method-BioTransformerParameters-toList}{\code{BioTransformerParameters$toList()}}
\item \href{#method-BioTransformerParameters-toSimpleType}{\code{BioTransformerParameters$toSimpleType()}}
\item \href{#method-BioTransformerParameters-fromJSON}{\code{BioTransformerParameters$fromJSON()}}
\item \href{#method-BioTransformerParameters-toJSONString}{\code{BioTransformerParameters$toJSONString()}}
\item \href{#method-BioTransformerParameters-fromJSONString}{\code{BioTransformerParameters$fromJSONString()}}
\item \href{#method-BioTransformerParameters-validateJSON}{\code{BioTransformerParameters$validateJSON()}}
\item \href{#method-BioTransformerParameters-toString}{\code{BioTransformerParameters$toString()}}
\item \href{#method-BioTransformerParameters-isValid}{\code{BioTransformerParameters$isValid()}}
\item \href{#method-BioTransformerParameters-getInvalidFields}{\code{BioTransformerParameters$getInvalidFields()}}
\item \href{#method-BioTransformerParameters-print}{\code{BioTransformerParameters$print()}}
\item \href{#method-BioTransformerParameters-clone}{\code{BioTransformerParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BioTransformerParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerParameters$new(
  cyp450Mode,
  p2Mode,
  bioTransformerSequenceSteps,
  useDB = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cyp450Mode}}{Specify the Phase I/Cyp450 mode for all provided BioTransformerSequenceSteps. Will only be applied to Steps that  require the Cyp450 mode as parameter. Can be null in cases where only BioTransformerSequenceSteps are specified  that do not need the Cyp450 mode.}

\item{\code{p2Mode}}{Specify the Phase II mode for all provided BioTransformerSequenceSteps. Will only be applied to Steps that  require the Phase II mode as parameter. Can be null in cases where only BioTransformerSequenceSteps are specified  that do not need the Phase II mode.}

\item{\code{bioTransformerSequenceSteps}}{Specify BioTransformerSequenceSteps to be applied to input structures. MultiStep MetabolicTransformations can  only be used as singletons (list size of one).}

\item{\code{useDB}}{Specify whether structures should additionally be retrieved from the HMDB database.. Default to TRUE.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerParameters-toList"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerParameters-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerParameters$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BioTransformerParameters as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of BioTransformerParameters (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerParameters-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerParameters-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert BioTransformerParameters to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerParameters$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of BioTransformerParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BioTransformerParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerParameters$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
BioTransformerParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of BioTransformerParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BioTransformerParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to BioTransformerParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of BioTransformerParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
