% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_transformer_sequence_step.R
\docType{class}
\name{BioTransformerSequenceStep}
\alias{BioTransformerSequenceStep}
\title{BioTransformerSequenceStep}
\format{
An \code{R6Class} generator object
}
\description{
BioTransformerSequenceStep Class

BioTransformerSequenceStep Class
}
\details{
Create a new BioTransformerSequenceStep
}
\examples{

## ------------------------------------------------
## Method `BioTransformerSequenceStep$toList`
## ------------------------------------------------

# convert array of BioTransformerSequenceStep (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metabolicTransformation}}{character [optional]}

\item{\code{iterations}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BioTransformerSequenceStep-new}{\code{BioTransformerSequenceStep$new()}}
\item \href{#method-BioTransformerSequenceStep-toJSON}{\code{BioTransformerSequenceStep$toJSON()}}
\item \href{#method-BioTransformerSequenceStep-toList}{\code{BioTransformerSequenceStep$toList()}}
\item \href{#method-BioTransformerSequenceStep-toSimpleType}{\code{BioTransformerSequenceStep$toSimpleType()}}
\item \href{#method-BioTransformerSequenceStep-fromJSON}{\code{BioTransformerSequenceStep$fromJSON()}}
\item \href{#method-BioTransformerSequenceStep-toJSONString}{\code{BioTransformerSequenceStep$toJSONString()}}
\item \href{#method-BioTransformerSequenceStep-fromJSONString}{\code{BioTransformerSequenceStep$fromJSONString()}}
\item \href{#method-BioTransformerSequenceStep-validateJSON}{\code{BioTransformerSequenceStep$validateJSON()}}
\item \href{#method-BioTransformerSequenceStep-toString}{\code{BioTransformerSequenceStep$toString()}}
\item \href{#method-BioTransformerSequenceStep-isValid}{\code{BioTransformerSequenceStep$isValid()}}
\item \href{#method-BioTransformerSequenceStep-getInvalidFields}{\code{BioTransformerSequenceStep$getInvalidFields()}}
\item \href{#method-BioTransformerSequenceStep-print}{\code{BioTransformerSequenceStep$print()}}
\item \href{#method-BioTransformerSequenceStep-clone}{\code{BioTransformerSequenceStep$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerSequenceStep-new"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerSequenceStep-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BioTransformerSequenceStep class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerSequenceStep$new(
  metabolicTransformation = NULL,
  iterations = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metabolicTransformation}}{metabolicTransformation}

\item{\code{iterations}}{iterations}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerSequenceStep-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerSequenceStep-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerSequenceStep$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerSequenceStep-toList"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerSequenceStep-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerSequenceStep$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BioTransformerSequenceStep as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of BioTransformerSequenceStep (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerSequenceStep-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerSequenceStep-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert BioTransformerSequenceStep to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerSequenceStep$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerSequenceStep-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerSequenceStep-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of BioTransformerSequenceStep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerSequenceStep$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BioTransformerSequenceStep
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerSequenceStep-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerSequenceStep-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerSequenceStep$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
BioTransformerSequenceStep in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerSequenceStep-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerSequenceStep-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of BioTransformerSequenceStep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerSequenceStep$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BioTransformerSequenceStep
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerSequenceStep-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerSequenceStep-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to BioTransformerSequenceStep and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerSequenceStep$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerSequenceStep-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerSequenceStep-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerSequenceStep$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of BioTransformerSequenceStep
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerSequenceStep-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerSequenceStep-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerSequenceStep$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerSequenceStep-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerSequenceStep-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerSequenceStep$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerSequenceStep-print"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerSequenceStep-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerSequenceStep$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioTransformerSequenceStep-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BioTransformerSequenceStep-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioTransformerSequenceStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
