% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopus_predictions.R
\docType{class}
\name{CanopusPredictions}
\alias{CanopusPredictions}
\title{CanopusPredictions}
\format{
An \code{R6Class} generator object
}
\description{
Container class that holds the CANOPUS compound class predictions for alle predictable compound classes.  This is the full CANOPUS result.

CanopusPredictions Class
}
\details{
Create a new CanopusPredictions
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classyFireClasses}}{All predicted ClassyFire classes list(\link{CompoundClass}) [optional]}

\item{\code{npcClasses}}{All predicted NPC classes list(\link{CompoundClass}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CanopusPredictions-new}{\code{CanopusPredictions$new()}}
\item \href{#method-CanopusPredictions-toJSON}{\code{CanopusPredictions$toJSON()}}
\item \href{#method-CanopusPredictions-fromJSON}{\code{CanopusPredictions$fromJSON()}}
\item \href{#method-CanopusPredictions-toJSONString}{\code{CanopusPredictions$toJSONString()}}
\item \href{#method-CanopusPredictions-fromJSONString}{\code{CanopusPredictions$fromJSONString()}}
\item \href{#method-CanopusPredictions-validateJSON}{\code{CanopusPredictions$validateJSON()}}
\item \href{#method-CanopusPredictions-toString}{\code{CanopusPredictions$toString()}}
\item \href{#method-CanopusPredictions-isValid}{\code{CanopusPredictions$isValid()}}
\item \href{#method-CanopusPredictions-getInvalidFields}{\code{CanopusPredictions$getInvalidFields()}}
\item \href{#method-CanopusPredictions-print}{\code{CanopusPredictions$print()}}
\item \href{#method-CanopusPredictions-clone}{\code{CanopusPredictions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CanopusPredictions-new"></a>}}
\if{latex}{\out{\hypertarget{method-CanopusPredictions-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CanopusPredictions class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CanopusPredictions$new(classyFireClasses = NULL, npcClasses = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{classyFireClasses}}{All predicted ClassyFire classes}

\item{\code{npcClasses}}{All predicted NPC classes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CanopusPredictions-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CanopusPredictions-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CanopusPredictions$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CanopusPredictions in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CanopusPredictions-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CanopusPredictions-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of CanopusPredictions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CanopusPredictions$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CanopusPredictions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CanopusPredictions-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CanopusPredictions-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CanopusPredictions$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CanopusPredictions in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CanopusPredictions-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CanopusPredictions-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of CanopusPredictions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CanopusPredictions$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CanopusPredictions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CanopusPredictions-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CanopusPredictions-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to CanopusPredictions and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CanopusPredictions$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CanopusPredictions-toString"></a>}}
\if{latex}{\out{\hypertarget{method-CanopusPredictions-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CanopusPredictions$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of CanopusPredictions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CanopusPredictions-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-CanopusPredictions-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CanopusPredictions$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CanopusPredictions-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-CanopusPredictions-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CanopusPredictions$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CanopusPredictions-print"></a>}}
\if{latex}{\out{\hypertarget{method-CanopusPredictions-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CanopusPredictions$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CanopusPredictions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CanopusPredictions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CanopusPredictions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
