% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\docType{class}
\name{Category}
\alias{Category}
\title{Category}
\format{
An \code{R6Class} generator object
}
\description{
Category Class

Category Class
}
\details{
Create a new Category
}
\examples{

## ------------------------------------------------
## Method `Category$toList`
## ------------------------------------------------

# convert array of Category (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{categoryId}}{character [optional]}

\item{\code{categoryName}}{character [optional]}

\item{\code{overallQuality}}{character [optional]}

\item{\code{items}}{list(\link{QualityItem}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Category-new}{\code{Category$new()}}
\item \href{#method-Category-toJSON}{\code{Category$toJSON()}}
\item \href{#method-Category-toList}{\code{Category$toList()}}
\item \href{#method-Category-toSimpleType}{\code{Category$toSimpleType()}}
\item \href{#method-Category-fromJSON}{\code{Category$fromJSON()}}
\item \href{#method-Category-toJSONString}{\code{Category$toJSONString()}}
\item \href{#method-Category-fromJSONString}{\code{Category$fromJSONString()}}
\item \href{#method-Category-validateJSON}{\code{Category$validateJSON()}}
\item \href{#method-Category-toString}{\code{Category$toString()}}
\item \href{#method-Category-isValid}{\code{Category$isValid()}}
\item \href{#method-Category-getInvalidFields}{\code{Category$getInvalidFields()}}
\item \href{#method-Category-print}{\code{Category$print()}}
\item \href{#method-Category-clone}{\code{Category$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-new"></a>}}
\if{latex}{\out{\hypertarget{method-Category-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Category class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$new(
  categoryId = NULL,
  categoryName = NULL,
  overallQuality = NULL,
  items = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{categoryId}}{categoryId}

\item{\code{categoryName}}{categoryName}

\item{\code{overallQuality}}{overallQuality}

\item{\code{items}}{items}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Category-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Category-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Category as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of Category (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-Category-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert Category to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Category-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Category
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Category
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Category-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Category in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Category-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Category
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Category
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Category-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Category and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Category-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Category
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Category-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Category-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-print"></a>}}
\if{latex}{\out{\hypertarget{method-Category-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Category-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Category-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Category$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
