% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classy_fire_class_statistics_api.R
\docType{class}
\name{ClassyFireClassStatisticsApi}
\alias{ClassyFireClassStatisticsApi}
\title{ClassyFireClassStatistics operations}
\format{
An \code{R6Class} generator object
}
\description{
ClassyFireClassStatisticsApi
}
\details{
SIRIUS Nightsky API

REST API that provides the full functionality of SIRIUS and its web services as background service. It is intended as entry-point for scripting languages and software integration SDKs.This API is exposed by SIRIUS 6

The version of the OpenAPI document: 3.1
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  ComputeClassyfireClassFoldChangesExperimental  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project-space to compute the fold change in.
var_fold_change_job_submission <- FoldChangeJobSubmission$new("leftRunGroup_example", "rightRunGroup_example", c("AVG"), c("APEX_INTENSITY")) # FoldChangeJobSubmission | Parameters of fold change job
var_opt_fields <- c("none") # array[character] | job opt fields. (Optional)

#[EXPERIMENTAL] Compute the fold change between two groups of runs
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$ComputeClassyfireClassFoldChangesExperimental(var_project_id, var_fold_change_job_submission, opt_fields = var_opt_fieldsdata_file = "result.txt")
result <- api_instance$classy_fire_class_statistics_api$ComputeClassyfireClassFoldChangesExperimental(var_project_id, var_fold_change_job_submission, opt_fields = var_opt_fields)
dput(result)


####################  DeleteClassyfireClassFoldChangesExperimental  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project-space to delete from.
var_left_group_name <- "left_group_name_example" # character | name of the left group.
var_right_group_name <- "right_group_name_example" # character | name of the right group.
var_aggregation <- "AVG" # character |  (Optional)
var_quantification <- "APEX_INTENSITY" # character |  (Optional)

#[EXPERIMENTAL] Delete fold changes
api_instance <- rsirius_api$new()

api_instance$classy_fire_class_statistics_api$DeleteClassyfireClassFoldChangesExperimental(var_project_id, var_left_group_name, var_right_group_name, aggregation = var_aggregation, quantification = var_quantification)


####################  GetClassyfireClassFoldChangeTableExperimental  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project-space to read from.
var_aggregation <- "AVG" # character | aggregation type. (Optional)
var_quantification <- "APEX_INTENSITY" # character | quantification type. (Optional)

#[EXPERIMENTAL] Get table of all fold changes in the project space
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetClassyfireClassFoldChangeTableExperimental(var_project_id, aggregation = var_aggregation, quantification = var_quantificationdata_file = "result.txt")
result <- api_instance$classy_fire_class_statistics_api$GetClassyfireClassFoldChangeTableExperimental(var_project_id, aggregation = var_aggregation, quantification = var_quantification)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClassyFireClassStatisticsApi-new}{\code{ClassyFireClassStatisticsApi$new()}}
\item \href{#method-ClassyFireClassStatisticsApi-ComputeClassyfireClassFoldChangesExperimental}{\code{ClassyFireClassStatisticsApi$ComputeClassyfireClassFoldChangesExperimental()}}
\item \href{#method-ClassyFireClassStatisticsApi-ComputeClassyfireClassFoldChangesExperimentalWithHttpInfo}{\code{ClassyFireClassStatisticsApi$ComputeClassyfireClassFoldChangesExperimentalWithHttpInfo()}}
\item \href{#method-ClassyFireClassStatisticsApi-DeleteClassyfireClassFoldChangesExperimental}{\code{ClassyFireClassStatisticsApi$DeleteClassyfireClassFoldChangesExperimental()}}
\item \href{#method-ClassyFireClassStatisticsApi-DeleteClassyfireClassFoldChangesExperimentalWithHttpInfo}{\code{ClassyFireClassStatisticsApi$DeleteClassyfireClassFoldChangesExperimentalWithHttpInfo()}}
\item \href{#method-ClassyFireClassStatisticsApi-GetClassyfireClassFoldChangeTableExperimental}{\code{ClassyFireClassStatisticsApi$GetClassyfireClassFoldChangeTableExperimental()}}
\item \href{#method-ClassyFireClassStatisticsApi-GetClassyfireClassFoldChangeTableExperimentalWithHttpInfo}{\code{ClassyFireClassStatisticsApi$GetClassyfireClassFoldChangeTableExperimentalWithHttpInfo()}}
\item \href{#method-ClassyFireClassStatisticsApi-clone}{\code{ClassyFireClassStatisticsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassyFireClassStatisticsApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClassyFireClassStatisticsApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ClassyFireClassStatisticsApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassyFireClassStatisticsApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassyFireClassStatisticsApi-ComputeClassyfireClassFoldChangesExperimental"></a>}}
\if{latex}{\out{\hypertarget{method-ClassyFireClassStatisticsApi-ComputeClassyfireClassFoldChangesExperimental}{}}}
\subsection{Method \code{ComputeClassyfireClassFoldChangesExperimental()}}{
[EXPERIMENTAL] Compute the fold change between two groups of runs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassyFireClassStatisticsApi$ComputeClassyfireClassFoldChangesExperimental(
  project_id,
  fold_change_job_submission,
  opt_fields = list("progress"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to compute the fold change in.}

\item{\code{fold_change_job_submission}}{Parameters of fold change job}

\item{\code{opt_fields}}{(optional) job opt fields. (default value: ["progress"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Job
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassyFireClassStatisticsApi-ComputeClassyfireClassFoldChangesExperimentalWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ClassyFireClassStatisticsApi-ComputeClassyfireClassFoldChangesExperimentalWithHttpInfo}{}}}
\subsection{Method \code{ComputeClassyfireClassFoldChangesExperimentalWithHttpInfo()}}{
[EXPERIMENTAL] Compute the fold change between two groups of runs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassyFireClassStatisticsApi$ComputeClassyfireClassFoldChangesExperimentalWithHttpInfo(
  project_id,
  fold_change_job_submission,
  opt_fields = list("progress"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to compute the fold change in.}

\item{\code{fold_change_job_submission}}{Parameters of fold change job}

\item{\code{opt_fields}}{(optional) job opt fields. (default value: ["progress"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (Job) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassyFireClassStatisticsApi-DeleteClassyfireClassFoldChangesExperimental"></a>}}
\if{latex}{\out{\hypertarget{method-ClassyFireClassStatisticsApi-DeleteClassyfireClassFoldChangesExperimental}{}}}
\subsection{Method \code{DeleteClassyfireClassFoldChangesExperimental()}}{
[EXPERIMENTAL] Delete fold changes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassyFireClassStatisticsApi$DeleteClassyfireClassFoldChangesExperimental(
  project_id,
  left_group_name,
  right_group_name,
  aggregation = "AVG",
  quantification = "APEX_INTENSITY",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to delete from.}

\item{\code{left_group_name}}{name of the left group.}

\item{\code{right_group_name}}{name of the right group.}

\item{\code{aggregation}}{(optional) No description (default value: "AVG")}

\item{\code{quantification}}{(optional) No description (default value: "APEX_INTENSITY")}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassyFireClassStatisticsApi-DeleteClassyfireClassFoldChangesExperimentalWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ClassyFireClassStatisticsApi-DeleteClassyfireClassFoldChangesExperimentalWithHttpInfo}{}}}
\subsection{Method \code{DeleteClassyfireClassFoldChangesExperimentalWithHttpInfo()}}{
[EXPERIMENTAL] Delete fold changes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassyFireClassStatisticsApi$DeleteClassyfireClassFoldChangesExperimentalWithHttpInfo(
  project_id,
  left_group_name,
  right_group_name,
  aggregation = "AVG",
  quantification = "APEX_INTENSITY",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to delete from.}

\item{\code{left_group_name}}{name of the left group.}

\item{\code{right_group_name}}{name of the right group.}

\item{\code{aggregation}}{(optional) No description (default value: "AVG")}

\item{\code{quantification}}{(optional) No description (default value: "APEX_INTENSITY")}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassyFireClassStatisticsApi-GetClassyfireClassFoldChangeTableExperimental"></a>}}
\if{latex}{\out{\hypertarget{method-ClassyFireClassStatisticsApi-GetClassyfireClassFoldChangeTableExperimental}{}}}
\subsection{Method \code{GetClassyfireClassFoldChangeTableExperimental()}}{
[EXPERIMENTAL] Get table of all fold changes in the project space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassyFireClassStatisticsApi$GetClassyfireClassFoldChangeTableExperimental(
  project_id,
  aggregation = "AVG",
  quantification = "APEX_INTENSITY",
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{aggregation}}{(optional) aggregation type. (default value: "AVG")}

\item{\code{quantification}}{(optional) quantification type. (default value: "APEX_INTENSITY")}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
StatisticsTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassyFireClassStatisticsApi-GetClassyfireClassFoldChangeTableExperimentalWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ClassyFireClassStatisticsApi-GetClassyfireClassFoldChangeTableExperimentalWithHttpInfo}{}}}
\subsection{Method \code{GetClassyfireClassFoldChangeTableExperimentalWithHttpInfo()}}{
[EXPERIMENTAL] Get table of all fold changes in the project space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassyFireClassStatisticsApi$GetClassyfireClassFoldChangeTableExperimentalWithHttpInfo(
  project_id,
  aggregation = "AVG",
  quantification = "APEX_INTENSITY",
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{aggregation}}{(optional) aggregation type. (default value: "AVG")}

\item{\code{quantification}}{(optional) quantification type. (default value: "APEX_INTENSITY")}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (StatisticsTable) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassyFireClassStatisticsApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClassyFireClassStatisticsApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassyFireClassStatisticsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
