% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collision_energy.R
\docType{class}
\name{CollisionEnergy}
\alias{CollisionEnergy}
\title{CollisionEnergy}
\format{
An \code{R6Class} generator object
}
\description{
CollisionEnergy Class
}
\details{
Create a new CollisionEnergy
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{minEnergy}}{numeric [optional]}

\item{\code{maxEnergy}}{numeric [optional]}

\item{\code{corrected}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CollisionEnergy-new}{\code{CollisionEnergy$new()}}
\item \href{#method-CollisionEnergy-toJSON}{\code{CollisionEnergy$toJSON()}}
\item \href{#method-CollisionEnergy-fromJSON}{\code{CollisionEnergy$fromJSON()}}
\item \href{#method-CollisionEnergy-toJSONString}{\code{CollisionEnergy$toJSONString()}}
\item \href{#method-CollisionEnergy-fromJSONString}{\code{CollisionEnergy$fromJSONString()}}
\item \href{#method-CollisionEnergy-validateJSON}{\code{CollisionEnergy$validateJSON()}}
\item \href{#method-CollisionEnergy-toString}{\code{CollisionEnergy$toString()}}
\item \href{#method-CollisionEnergy-isValid}{\code{CollisionEnergy$isValid()}}
\item \href{#method-CollisionEnergy-getInvalidFields}{\code{CollisionEnergy$getInvalidFields()}}
\item \href{#method-CollisionEnergy-print}{\code{CollisionEnergy$print()}}
\item \href{#method-CollisionEnergy-clone}{\code{CollisionEnergy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollisionEnergy-new"></a>}}
\if{latex}{\out{\hypertarget{method-CollisionEnergy-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CollisionEnergy class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollisionEnergy$new(minEnergy = NULL, maxEnergy = NULL, corrected = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{minEnergy}}{minEnergy}

\item{\code{maxEnergy}}{maxEnergy}

\item{\code{corrected}}{corrected}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollisionEnergy-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CollisionEnergy-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollisionEnergy$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CollisionEnergy in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollisionEnergy-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CollisionEnergy-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of CollisionEnergy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollisionEnergy$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CollisionEnergy
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollisionEnergy-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CollisionEnergy-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollisionEnergy$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CollisionEnergy in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollisionEnergy-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CollisionEnergy-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of CollisionEnergy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollisionEnergy$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CollisionEnergy
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollisionEnergy-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CollisionEnergy-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to CollisionEnergy and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollisionEnergy$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollisionEnergy-toString"></a>}}
\if{latex}{\out{\hypertarget{method-CollisionEnergy-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollisionEnergy$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of CollisionEnergy
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollisionEnergy-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-CollisionEnergy-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollisionEnergy$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollisionEnergy-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-CollisionEnergy-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollisionEnergy$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollisionEnergy-print"></a>}}
\if{latex}{\out{\hypertarget{method-CollisionEnergy-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollisionEnergy$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollisionEnergy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CollisionEnergy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollisionEnergy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
