% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_opt_field.R
\docType{class}
\name{CompoundOptField}
\alias{CompoundOptField}
\title{CompoundOptField}
\format{
An \code{R6Class} generator object
}
\description{
CompoundOptField Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CompoundOptField-new}{\code{CompoundOptField$new()}}
\item \href{#method-CompoundOptField-toJSON}{\code{CompoundOptField$toJSON()}}
\item \href{#method-CompoundOptField-fromJSON}{\code{CompoundOptField$fromJSON()}}
\item \href{#method-CompoundOptField-toJSONString}{\code{CompoundOptField$toJSONString()}}
\item \href{#method-CompoundOptField-fromJSONString}{\code{CompoundOptField$fromJSONString()}}
\item \href{#method-CompoundOptField-clone}{\code{CompoundOptField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundOptField-new"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundOptField-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CompoundOptField class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundOptField$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundOptField-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundOptField-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundOptField$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CompoundOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundOptField-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundOptField-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of CompoundOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundOptField$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CompoundOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundOptField-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundOptField-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundOptField$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CompoundOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundOptField-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundOptField-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of CompoundOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundOptField$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CompoundOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundOptField-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundOptField-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundOptField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
