% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds_api.R
\docType{class}
\name{CompoundsApi}
\alias{CompoundsApi}
\title{Compounds operations}
\format{
An \code{R6Class} generator object
}
\description{
CompoundsApi
}
\details{
SIRIUS Nightsky API

REST API that provides the full functionality of SIRIUS and its web services as background service. It is intended as entry-point for scripting languages and software integration SDKs.This API is exposed by SIRIUS 6

The version of the OpenAPI document: 3.2
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  AddCompounds  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project-space to import into.
var_compound_import <- c(CompoundImport$new(c(FeatureImport$new(123, 123, "name_example", "externalFeatureId_example", c("detectedAdducts_example"), 123, 123, 123, "NOT_APPLICABLE", BasicSpectrum$new("cosineQuery_example", c(SimplePeak$new(123, 123)), "name_example", 123, "collisionEnergy_example", "instrument_example", 123, 123, SimplePeak$new(123, 123), 123, 123, 123, 123, 123), c(BasicSpectrum$new("cosineQuery_example", c(SimplePeak$new(..., ...)), "name_example", 123, "collisionEnergy_example", "instrument_example", 123, 123, SimplePeak$new(123, 123), 123, 123, 123, 123, 123)), c(BasicSpectrum$new("cosineQuery_example", c(SimplePeak$new(..., ...)), "name_example", 123, "collisionEnergy_example", "instrument_example", 123, 123, SimplePeak$new(123, 123), 123, 123, 123, 123, 123)))), "name_example")) # array[CompoundImport] | the compound data to be imported
var_profile <- "profile_example" # character | profile describing the instrument used to measure the data. Used to merge spectra. (Optional)
var_opt_fields <- c("none") # array[character] | set of optional fields to be included. Use 'none' to override defaults. (Optional)
var_opt_fields_features <- c("none") # array[character] | set of optional fields of the nested features to be included. Use 'none' to override defaults. (Optional)

#Import compounds and their contained features.
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$AddCompounds(var_project_id, var_compound_import, profile = var_profile, opt_fields = var_opt_fields, opt_fields_features = var_opt_fields_featuresdata_file = "result.txt")
result <- api_instance$compounds_api$AddCompounds(var_project_id, var_compound_import, profile = var_profile, opt_fields = var_opt_fields, opt_fields_features = var_opt_fields_features)
dput(result)


####################  DeleteCompound  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project-space to delete from.
var_compound_id <- "compound_id_example" # character | identifier of the compound to delete.

#Delete compound (group of ion identities) with the given identifier (and the included features) from the  specified project-space.
api_instance <- rsirius_api$new()

api_instance$compounds_api$DeleteCompound(var_project_id, var_compound_id)


####################  GetCompound  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project-space to read from.
var_compound_id <- "compound_id_example" # character | identifier of the compound (group of ion identities) to access.
var_ms_data_search_prepared <- FALSE # character | Returns all fragment spectra in a preprocessed form as used for fast                             Cosine/Modified Cosine computation. Gives you spectra compatible with SpectralLibraryMatch                             peak assignments and reference spectra. (Optional)
var_opt_fields <- c("none") # array[character] | set of optional fields to be included. Use 'none' only to override defaults. (Optional)
var_opt_fields_features <- c("none") # array[character] |  (Optional)

#Get compound (group of ion identities) with the given identifier from the specified project-space.
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetCompound(var_project_id, var_compound_id, ms_data_search_prepared = var_ms_data_search_prepared, opt_fields = var_opt_fields, opt_fields_features = var_opt_fields_featuresdata_file = "result.txt")
result <- api_instance$compounds_api$GetCompound(var_project_id, var_compound_id, ms_data_search_prepared = var_ms_data_search_prepared, opt_fields = var_opt_fields, opt_fields_features = var_opt_fields_features)
dput(result)


####################  GetCompoundQuantTable  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project-space to read from.
var_search_query <- "search_query_example" # character | Optional query in lucene syntax selecting compounds by id. Omit this parameter to quantify all compounds. (Optional)
var_type <- "APEX_INTENSITY" # character | quantification type. (Optional)
var_opt_fields <- c(QuantTableOptField$new()) # array[QuantTableOptField] |  (Optional)

#Returns the full quantification table of compounds
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetCompoundQuantTable(var_project_id, search_query = var_search_query, type = var_type, opt_fields = var_opt_fieldsdata_file = "result.txt")
result <- api_instance$compounds_api$GetCompoundQuantTable(var_project_id, search_query = var_search_query, type = var_type, opt_fields = var_opt_fields)
dput(result)


####################  GetCompounds  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project-space to read from.
var_ms_data_search_prepared <- FALSE # character | Returns all fragment spectra in a preprocessed form as used for fast                             Cosine/Modified Cosine computation. Gives you spectra compatible with SpectralLibraryMatch                             peak assignments and reference spectra. (Optional)
var_opt_fields <- c("none") # array[character] | set of optional fields to be included. Use 'none' only to override defaults. (Optional)
var_opt_fields_features <- c("none") # array[character] |  (Optional)

#[DEPRECATED] List of all available compounds (group of ion identities) in the given project-space
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetCompounds(var_project_id, ms_data_search_prepared = var_ms_data_search_prepared, opt_fields = var_opt_fields, opt_fields_features = var_opt_fields_featuresdata_file = "result.txt")
result <- api_instance$compounds_api$GetCompounds(var_project_id, ms_data_search_prepared = var_ms_data_search_prepared, opt_fields = var_opt_fields, opt_fields_features = var_opt_fields_features)
dput(result)


####################  GetCompoundsPage  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project-space to read from.
var_search_query <- "search_query_example" # character | Optional search query in lucene syntax. Not yet supported for compounds; a non-empty                      query responds with 405 METHOD_NOT_ALLOWED. Omit this parameter to page over all compounds. (Optional)
var_page <- 0 # integer | Zero-based page index (0..N) (Optional)
var_size <- 20 # integer | The size of the page to be returned (Optional)
var_sort <- c("inner_example") # array[character] | Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported. (Optional)
var_ms_data_search_prepared <- FALSE # character | Returns all fragment spectra in a preprocessed form as used for fast                             Cosine/Modified Cosine computation. Gives you spectra compatible with SpectralLibraryMatch                             peak assignments and reference spectra. (Optional)
var_opt_fields <- c("none") # array[character] | set of optional fields to be included. Use 'none' only to override defaults. (Optional)
var_opt_fields_features <- c("none") # array[character] |  (Optional)

#Page of available compounds (group of ion identities) in the given project-space
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetCompoundsPage(var_project_id, search_query = var_search_query, page = var_page, size = var_size, sort = var_sort, ms_data_search_prepared = var_ms_data_search_prepared, opt_fields = var_opt_fields, opt_fields_features = var_opt_fields_featuresdata_file = "result.txt")
result <- api_instance$compounds_api$GetCompoundsPage(var_project_id, search_query = var_search_query, page = var_page, size = var_size, sort = var_sort, ms_data_search_prepared = var_ms_data_search_prepared, opt_fields = var_opt_fields, opt_fields_features = var_opt_fields_features)
dput(result)


####################  GetCompoundsSearchableFields  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project space to read the searchable compound fields from.

#Get all fields that can be used in the searchQuery parameter of compound endpoints
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetCompoundsSearchableFields(var_project_iddata_file = "result.txt")
result <- api_instance$compounds_api$GetCompoundsSearchableFields(var_project_id)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CompoundsApi-new}{\code{CompoundsApi$new()}}
\item \href{#method-CompoundsApi-AddCompounds}{\code{CompoundsApi$AddCompounds()}}
\item \href{#method-CompoundsApi-AddCompoundsWithHttpInfo}{\code{CompoundsApi$AddCompoundsWithHttpInfo()}}
\item \href{#method-CompoundsApi-DeleteCompound}{\code{CompoundsApi$DeleteCompound()}}
\item \href{#method-CompoundsApi-DeleteCompoundWithHttpInfo}{\code{CompoundsApi$DeleteCompoundWithHttpInfo()}}
\item \href{#method-CompoundsApi-GetCompound}{\code{CompoundsApi$GetCompound()}}
\item \href{#method-CompoundsApi-GetCompoundWithHttpInfo}{\code{CompoundsApi$GetCompoundWithHttpInfo()}}
\item \href{#method-CompoundsApi-GetCompoundQuantTable}{\code{CompoundsApi$GetCompoundQuantTable()}}
\item \href{#method-CompoundsApi-GetCompoundQuantTableWithHttpInfo}{\code{CompoundsApi$GetCompoundQuantTableWithHttpInfo()}}
\item \href{#method-CompoundsApi-GetCompounds}{\code{CompoundsApi$GetCompounds()}}
\item \href{#method-CompoundsApi-GetCompoundsWithHttpInfo}{\code{CompoundsApi$GetCompoundsWithHttpInfo()}}
\item \href{#method-CompoundsApi-GetCompoundsPage}{\code{CompoundsApi$GetCompoundsPage()}}
\item \href{#method-CompoundsApi-GetCompoundsPageWithHttpInfo}{\code{CompoundsApi$GetCompoundsPageWithHttpInfo()}}
\item \href{#method-CompoundsApi-GetCompoundsSearchableFields}{\code{CompoundsApi$GetCompoundsSearchableFields()}}
\item \href{#method-CompoundsApi-GetCompoundsSearchableFieldsWithHttpInfo}{\code{CompoundsApi$GetCompoundsSearchableFieldsWithHttpInfo()}}
\item \href{#method-CompoundsApi-clone}{\code{CompoundsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CompoundsApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-AddCompounds"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-AddCompounds}{}}}
\subsection{Method \code{AddCompounds()}}{
Import compounds and their contained features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$AddCompounds(
  project_id,
  compound_import,
  profile = NULL,
  opt_fields = list("none"),
  opt_fields_features = list("none"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to import into.}

\item{\code{compound_import}}{the compound data to be imported}

\item{\code{profile}}{(optional) profile describing the instrument used to measure the data. Used to merge spectra.}

\item{\code{opt_fields}}{(optional) set of optional fields to be included. Use 'none' to override defaults. (default value: ["none"])}

\item{\code{opt_fields_features}}{(optional) set of optional fields of the nested features to be included. Use 'none' to override defaults. (default value: ["none"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array[Compound]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-AddCompoundsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-AddCompoundsWithHttpInfo}{}}}
\subsection{Method \code{AddCompoundsWithHttpInfo()}}{
Import compounds and their contained features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$AddCompoundsWithHttpInfo(
  project_id,
  compound_import,
  profile = NULL,
  opt_fields = list("none"),
  opt_fields_features = list("none"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to import into.}

\item{\code{compound_import}}{the compound data to be imported}

\item{\code{profile}}{(optional) profile describing the instrument used to measure the data. Used to merge spectra.}

\item{\code{opt_fields}}{(optional) set of optional fields to be included. Use 'none' to override defaults. (default value: ["none"])}

\item{\code{opt_fields_features}}{(optional) set of optional fields of the nested features to be included. Use 'none' to override defaults. (default value: ["none"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (array[Compound]) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-DeleteCompound"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-DeleteCompound}{}}}
\subsection{Method \code{DeleteCompound()}}{
Delete compound (group of ion identities) with the given identifier (and the included features) from the  specified project-space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$DeleteCompound(project_id, compound_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to delete from.}

\item{\code{compound_id}}{identifier of the compound to delete.}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-DeleteCompoundWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-DeleteCompoundWithHttpInfo}{}}}
\subsection{Method \code{DeleteCompoundWithHttpInfo()}}{
Delete compound (group of ion identities) with the given identifier (and the included features) from the  specified project-space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$DeleteCompoundWithHttpInfo(project_id, compound_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to delete from.}

\item{\code{compound_id}}{identifier of the compound to delete.}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-GetCompound"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-GetCompound}{}}}
\subsection{Method \code{GetCompound()}}{
Get compound (group of ion identities) with the given identifier from the specified project-space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$GetCompound(
  project_id,
  compound_id,
  ms_data_search_prepared = FALSE,
  opt_fields = list("none"),
  opt_fields_features = list("none"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{compound_id}}{identifier of the compound (group of ion identities) to access.}

\item{\code{ms_data_search_prepared}}{(optional) Returns all fragment spectra in a preprocessed form as used for fast                             Cosine/Modified Cosine computation. Gives you spectra compatible with SpectralLibraryMatch                             peak assignments and reference spectra. (default value: FALSE)}

\item{\code{opt_fields}}{(optional) set of optional fields to be included. Use 'none' only to override defaults. (default value: ["none"])}

\item{\code{opt_fields_features}}{(optional) No description (default value: ["none"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Compound
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-GetCompoundWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-GetCompoundWithHttpInfo}{}}}
\subsection{Method \code{GetCompoundWithHttpInfo()}}{
Get compound (group of ion identities) with the given identifier from the specified project-space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$GetCompoundWithHttpInfo(
  project_id,
  compound_id,
  ms_data_search_prepared = FALSE,
  opt_fields = list("none"),
  opt_fields_features = list("none"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{compound_id}}{identifier of the compound (group of ion identities) to access.}

\item{\code{ms_data_search_prepared}}{(optional) Returns all fragment spectra in a preprocessed form as used for fast                             Cosine/Modified Cosine computation. Gives you spectra compatible with SpectralLibraryMatch                             peak assignments and reference spectra. (default value: FALSE)}

\item{\code{opt_fields}}{(optional) set of optional fields to be included. Use 'none' only to override defaults. (default value: ["none"])}

\item{\code{opt_fields_features}}{(optional) No description (default value: ["none"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (Compound) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-GetCompoundQuantTable"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-GetCompoundQuantTable}{}}}
\subsection{Method \code{GetCompoundQuantTable()}}{
Returns the full quantification table of compounds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$GetCompoundQuantTable(
  project_id,
  search_query = NULL,
  type = "APEX_INTENSITY",
  opt_fields = list("none"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{search_query}}{(optional) Optional query in lucene syntax selecting compounds by id. Omit this parameter to quantify all compounds.}

\item{\code{type}}{(optional) quantification type. (default value: "APEX_INTENSITY")}

\item{\code{opt_fields}}{(optional) No description (default value: ["none"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
QuantTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-GetCompoundQuantTableWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-GetCompoundQuantTableWithHttpInfo}{}}}
\subsection{Method \code{GetCompoundQuantTableWithHttpInfo()}}{
Returns the full quantification table of compounds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$GetCompoundQuantTableWithHttpInfo(
  project_id,
  search_query = NULL,
  type = "APEX_INTENSITY",
  opt_fields = list("none"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{search_query}}{(optional) Optional query in lucene syntax selecting compounds by id. Omit this parameter to quantify all compounds.}

\item{\code{type}}{(optional) quantification type. (default value: "APEX_INTENSITY")}

\item{\code{opt_fields}}{(optional) No description (default value: ["none"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (QuantTable) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-GetCompounds"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-GetCompounds}{}}}
\subsection{Method \code{GetCompounds()}}{
[DEPRECATED] List of all available compounds (group of ion identities) in the given project-space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$GetCompounds(
  project_id,
  ms_data_search_prepared = FALSE,
  opt_fields = list("none"),
  opt_fields_features = list("none"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{ms_data_search_prepared}}{(optional) Returns all fragment spectra in a preprocessed form as used for fast                             Cosine/Modified Cosine computation. Gives you spectra compatible with SpectralLibraryMatch                             peak assignments and reference spectra. (default value: FALSE)}

\item{\code{opt_fields}}{(optional) set of optional fields to be included. Use 'none' only to override defaults. (default value: ["none"])}

\item{\code{opt_fields_features}}{(optional) No description (default value: ["none"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array[Compound]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-GetCompoundsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-GetCompoundsWithHttpInfo}{}}}
\subsection{Method \code{GetCompoundsWithHttpInfo()}}{
[DEPRECATED] List of all available compounds (group of ion identities) in the given project-space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$GetCompoundsWithHttpInfo(
  project_id,
  ms_data_search_prepared = FALSE,
  opt_fields = list("none"),
  opt_fields_features = list("none"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{ms_data_search_prepared}}{(optional) Returns all fragment spectra in a preprocessed form as used for fast                             Cosine/Modified Cosine computation. Gives you spectra compatible with SpectralLibraryMatch                             peak assignments and reference spectra. (default value: FALSE)}

\item{\code{opt_fields}}{(optional) set of optional fields to be included. Use 'none' only to override defaults. (default value: ["none"])}

\item{\code{opt_fields_features}}{(optional) No description (default value: ["none"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (array[Compound]) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-GetCompoundsPage"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-GetCompoundsPage}{}}}
\subsection{Method \code{GetCompoundsPage()}}{
Page of available compounds (group of ion identities) in the given project-space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$GetCompoundsPage(
  project_id,
  search_query = NULL,
  page = 0,
  size = 20,
  sort = NULL,
  ms_data_search_prepared = FALSE,
  opt_fields = list("none"),
  opt_fields_features = list("none"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{search_query}}{(optional) Optional search query in lucene syntax. Not yet supported for compounds; a non-empty                      query responds with 405 METHOD_NOT_ALLOWED. Omit this parameter to page over all compounds.}

\item{\code{page}}{(optional) Zero-based page index (0..N) (default value: 0)}

\item{\code{size}}{(optional) The size of the page to be returned (default value: 20)}

\item{\code{sort}}{(optional) Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.}

\item{\code{ms_data_search_prepared}}{(optional) Returns all fragment spectra in a preprocessed form as used for fast                             Cosine/Modified Cosine computation. Gives you spectra compatible with SpectralLibraryMatch                             peak assignments and reference spectra. (default value: FALSE)}

\item{\code{opt_fields}}{(optional) set of optional fields to be included. Use 'none' only to override defaults. (default value: ["none"])}

\item{\code{opt_fields_features}}{(optional) No description (default value: ["none"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PagedModelCompound
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-GetCompoundsPageWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-GetCompoundsPageWithHttpInfo}{}}}
\subsection{Method \code{GetCompoundsPageWithHttpInfo()}}{
Page of available compounds (group of ion identities) in the given project-space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$GetCompoundsPageWithHttpInfo(
  project_id,
  search_query = NULL,
  page = 0,
  size = 20,
  sort = NULL,
  ms_data_search_prepared = FALSE,
  opt_fields = list("none"),
  opt_fields_features = list("none"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{search_query}}{(optional) Optional search query in lucene syntax. Not yet supported for compounds; a non-empty                      query responds with 405 METHOD_NOT_ALLOWED. Omit this parameter to page over all compounds.}

\item{\code{page}}{(optional) Zero-based page index (0..N) (default value: 0)}

\item{\code{size}}{(optional) The size of the page to be returned (default value: 20)}

\item{\code{sort}}{(optional) Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.}

\item{\code{ms_data_search_prepared}}{(optional) Returns all fragment spectra in a preprocessed form as used for fast                             Cosine/Modified Cosine computation. Gives you spectra compatible with SpectralLibraryMatch                             peak assignments and reference spectra. (default value: FALSE)}

\item{\code{opt_fields}}{(optional) set of optional fields to be included. Use 'none' only to override defaults. (default value: ["none"])}

\item{\code{opt_fields_features}}{(optional) No description (default value: ["none"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (PagedModelCompound) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-GetCompoundsSearchableFields"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-GetCompoundsSearchableFields}{}}}
\subsection{Method \code{GetCompoundsSearchableFields()}}{
Get all fields that can be used in the searchQuery parameter of compound endpoints
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$GetCompoundsSearchableFields(project_id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project space to read the searchable compound fields from.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array[SearchableField]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-GetCompoundsSearchableFieldsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-GetCompoundsSearchableFieldsWithHttpInfo}{}}}
\subsection{Method \code{GetCompoundsSearchableFieldsWithHttpInfo()}}{
Get all fields that can be used in the searchQuery parameter of compound endpoints
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$GetCompoundsSearchableFieldsWithHttpInfo(
  project_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project space to read the searchable compound fields from.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (array[SearchableField]) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompoundsApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CompoundsApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompoundsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
