% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_mode.R
\docType{class}
\name{ConfidenceMode}
\alias{ConfidenceMode}
\title{ConfidenceMode}
\format{
An \code{R6Class} generator object
}
\description{
ConfidenceMode Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConfidenceMode-new}{\code{ConfidenceMode$new()}}
\item \href{#method-ConfidenceMode-toJSON}{\code{ConfidenceMode$toJSON()}}
\item \href{#method-ConfidenceMode-fromJSON}{\code{ConfidenceMode$fromJSON()}}
\item \href{#method-ConfidenceMode-toJSONString}{\code{ConfidenceMode$toJSONString()}}
\item \href{#method-ConfidenceMode-fromJSONString}{\code{ConfidenceMode$fromJSONString()}}
\item \href{#method-ConfidenceMode-clone}{\code{ConfidenceMode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfidenceMode-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConfidenceMode-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ConfidenceMode class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfidenceMode$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfidenceMode-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConfidenceMode-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfidenceMode$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ConfidenceMode in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfidenceMode-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConfidenceMode-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ConfidenceMode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfidenceMode$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConfidenceMode
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfidenceMode-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConfidenceMode-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfidenceMode$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ConfidenceMode in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfidenceMode-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConfidenceMode-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ConfidenceMode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfidenceMode$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConfidenceMode
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfidenceMode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConfidenceMode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfidenceMode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
