% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_error.R
\docType{class}
\name{ConnectionError}
\alias{ConnectionError}
\title{ConnectionError}
\format{
An \code{R6Class} generator object
}
\description{
ConnectionError Class

ConnectionError Class
}
\details{
Create a new ConnectionError
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{errorType}}{character}

\item{\code{errorKlass}}{character}

\item{\code{siriusErrorCode}}{integer}

\item{\code{siriusMessage}}{character}

\item{\code{serverResponseErrorCode}}{integer [optional]}

\item{\code{serverResponseErrorMessage}}{character [optional]}

\item{\code{error}}{character [optional]}

\item{\code{warning}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConnectionError-new}{\code{ConnectionError$new()}}
\item \href{#method-ConnectionError-toJSON}{\code{ConnectionError$toJSON()}}
\item \href{#method-ConnectionError-fromJSON}{\code{ConnectionError$fromJSON()}}
\item \href{#method-ConnectionError-toJSONString}{\code{ConnectionError$toJSONString()}}
\item \href{#method-ConnectionError-fromJSONString}{\code{ConnectionError$fromJSONString()}}
\item \href{#method-ConnectionError-validateJSON}{\code{ConnectionError$validateJSON()}}
\item \href{#method-ConnectionError-toString}{\code{ConnectionError$toString()}}
\item \href{#method-ConnectionError-isValid}{\code{ConnectionError$isValid()}}
\item \href{#method-ConnectionError-getInvalidFields}{\code{ConnectionError$getInvalidFields()}}
\item \href{#method-ConnectionError-print}{\code{ConnectionError$print()}}
\item \href{#method-ConnectionError-clone}{\code{ConnectionError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionError-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionError-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ConnectionError class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionError$new(
  errorType,
  errorKlass,
  siriusErrorCode,
  siriusMessage,
  serverResponseErrorCode = NULL,
  serverResponseErrorMessage = NULL,
  error = NULL,
  warning = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{errorType}}{errorType}

\item{\code{errorKlass}}{errorKlass}

\item{\code{siriusErrorCode}}{siriusErrorCode}

\item{\code{siriusMessage}}{siriusMessage}

\item{\code{serverResponseErrorCode}}{serverResponseErrorCode}

\item{\code{serverResponseErrorMessage}}{serverResponseErrorMessage}

\item{\code{error}}{error}

\item{\code{warning}}{warning}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionError-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionError-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionError$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ConnectionError in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionError-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionError-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ConnectionError
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionError$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConnectionError
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionError-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionError-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionError$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ConnectionError in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionError-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionError-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ConnectionError
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionError$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConnectionError
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionError-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionError-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ConnectionError and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionError$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionError-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionError-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionError$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ConnectionError
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionError-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionError-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionError$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionError-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionError-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionError$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionError-print"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionError-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionError$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
