% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_annotations_csi.R
\docType{class}
\name{ConsensusAnnotationsCSI}
\alias{ConsensusAnnotationsCSI}
\title{ConsensusAnnotationsCSI}
\format{
An \code{R6Class} generator object
}
\description{
ConsensusAnnotationsCSI Class

ConsensusAnnotationsCSI Class
}
\details{
Create a new ConsensusAnnotationsCSI
}
\examples{

## ------------------------------------------------
## Method `ConsensusAnnotationsCSI$toList`
## ------------------------------------------------

# convert array of ConsensusAnnotationsCSI (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{selectionCriterion}}{Null if this is a custom selection character [optional]}

\item{\code{csiFingerIdStructure}}{\link{StructureCandidate} [optional]}

\item{\code{confidenceExactMatch}}{Confidence value that represents the certainty that reported consensus structure is exactly the measured one  If multiple features support this consensus structure the maximum confidence is reported numeric [optional]}

\item{\code{confidenceApproxMatch}}{Confidence value that represents the certainty that the exact consensus structure or a very similar  structure (e.g. measured by Maximum Common Edge Subgraph Distance) is the measured one.  If multiple features support this consensus structure the maximum confidence is reported numeric [optional]}

\item{\code{molecularFormula}}{Molecular formula of the consensus annotation  Might be null if no consensus formula is available. character [optional]}

\item{\code{compoundClasses}}{\link{CompoundClasses} [optional]}

\item{\code{supportingFeatureIds}}{FeatureIds where the topAnnotation supports this annotation. list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConsensusAnnotationsCSI-new}{\code{ConsensusAnnotationsCSI$new()}}
\item \href{#method-ConsensusAnnotationsCSI-toJSON}{\code{ConsensusAnnotationsCSI$toJSON()}}
\item \href{#method-ConsensusAnnotationsCSI-toList}{\code{ConsensusAnnotationsCSI$toList()}}
\item \href{#method-ConsensusAnnotationsCSI-toSimpleType}{\code{ConsensusAnnotationsCSI$toSimpleType()}}
\item \href{#method-ConsensusAnnotationsCSI-fromJSON}{\code{ConsensusAnnotationsCSI$fromJSON()}}
\item \href{#method-ConsensusAnnotationsCSI-toJSONString}{\code{ConsensusAnnotationsCSI$toJSONString()}}
\item \href{#method-ConsensusAnnotationsCSI-fromJSONString}{\code{ConsensusAnnotationsCSI$fromJSONString()}}
\item \href{#method-ConsensusAnnotationsCSI-validateJSON}{\code{ConsensusAnnotationsCSI$validateJSON()}}
\item \href{#method-ConsensusAnnotationsCSI-toString}{\code{ConsensusAnnotationsCSI$toString()}}
\item \href{#method-ConsensusAnnotationsCSI-isValid}{\code{ConsensusAnnotationsCSI$isValid()}}
\item \href{#method-ConsensusAnnotationsCSI-getInvalidFields}{\code{ConsensusAnnotationsCSI$getInvalidFields()}}
\item \href{#method-ConsensusAnnotationsCSI-print}{\code{ConsensusAnnotationsCSI$print()}}
\item \href{#method-ConsensusAnnotationsCSI-clone}{\code{ConsensusAnnotationsCSI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusAnnotationsCSI-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusAnnotationsCSI-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ConsensusAnnotationsCSI class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusAnnotationsCSI$new(
  selectionCriterion = NULL,
  csiFingerIdStructure = NULL,
  confidenceExactMatch = NULL,
  confidenceApproxMatch = NULL,
  molecularFormula = NULL,
  compoundClasses = NULL,
  supportingFeatureIds = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selectionCriterion}}{Null if this is a custom selection}

\item{\code{csiFingerIdStructure}}{csiFingerIdStructure}

\item{\code{confidenceExactMatch}}{Confidence value that represents the certainty that reported consensus structure is exactly the measured one  If multiple features support this consensus structure the maximum confidence is reported}

\item{\code{confidenceApproxMatch}}{Confidence value that represents the certainty that the exact consensus structure or a very similar  structure (e.g. measured by Maximum Common Edge Subgraph Distance) is the measured one.  If multiple features support this consensus structure the maximum confidence is reported}

\item{\code{molecularFormula}}{Molecular formula of the consensus annotation  Might be null if no consensus formula is available.}

\item{\code{compoundClasses}}{compoundClasses}

\item{\code{supportingFeatureIds}}{FeatureIds where the topAnnotation supports this annotation.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusAnnotationsCSI-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusAnnotationsCSI-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusAnnotationsCSI$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusAnnotationsCSI-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusAnnotationsCSI-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusAnnotationsCSI$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ConsensusAnnotationsCSI as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of ConsensusAnnotationsCSI (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusAnnotationsCSI-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusAnnotationsCSI-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert ConsensusAnnotationsCSI to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusAnnotationsCSI$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusAnnotationsCSI-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusAnnotationsCSI-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ConsensusAnnotationsCSI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusAnnotationsCSI$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConsensusAnnotationsCSI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusAnnotationsCSI-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusAnnotationsCSI-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusAnnotationsCSI$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ConsensusAnnotationsCSI in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusAnnotationsCSI-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusAnnotationsCSI-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ConsensusAnnotationsCSI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusAnnotationsCSI$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConsensusAnnotationsCSI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusAnnotationsCSI-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusAnnotationsCSI-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ConsensusAnnotationsCSI and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusAnnotationsCSI$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusAnnotationsCSI-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusAnnotationsCSI-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusAnnotationsCSI$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ConsensusAnnotationsCSI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusAnnotationsCSI-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusAnnotationsCSI-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusAnnotationsCSI$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusAnnotationsCSI-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusAnnotationsCSI-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusAnnotationsCSI$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusAnnotationsCSI-print"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusAnnotationsCSI-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusAnnotationsCSI$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusAnnotationsCSI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusAnnotationsCSI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusAnnotationsCSI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
