% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_criterion_csi.R
\docType{class}
\name{ConsensusCriterionCSI}
\alias{ConsensusCriterionCSI}
\title{ConsensusCriterionCSI}
\format{
An \code{R6Class} generator object
}
\description{
ConsensusCriterionCSI Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConsensusCriterionCSI-new}{\code{ConsensusCriterionCSI$new()}}
\item \href{#method-ConsensusCriterionCSI-toJSON}{\code{ConsensusCriterionCSI$toJSON()}}
\item \href{#method-ConsensusCriterionCSI-fromJSON}{\code{ConsensusCriterionCSI$fromJSON()}}
\item \href{#method-ConsensusCriterionCSI-toJSONString}{\code{ConsensusCriterionCSI$toJSONString()}}
\item \href{#method-ConsensusCriterionCSI-fromJSONString}{\code{ConsensusCriterionCSI$fromJSONString()}}
\item \href{#method-ConsensusCriterionCSI-clone}{\code{ConsensusCriterionCSI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusCriterionCSI-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusCriterionCSI-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ConsensusCriterionCSI class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusCriterionCSI$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusCriterionCSI-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusCriterionCSI-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusCriterionCSI$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ConsensusCriterionCSI in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusCriterionCSI-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusCriterionCSI-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ConsensusCriterionCSI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusCriterionCSI$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConsensusCriterionCSI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusCriterionCSI-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusCriterionCSI-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusCriterionCSI$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ConsensusCriterionCSI in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusCriterionCSI-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusCriterionCSI-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ConsensusCriterionCSI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusCriterionCSI$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConsensusCriterionCSI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusCriterionCSI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusCriterionCSI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusCriterionCSI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
