% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_criterion_de_novo.R
\docType{class}
\name{ConsensusCriterionDeNovo}
\alias{ConsensusCriterionDeNovo}
\title{ConsensusCriterionDeNovo}
\format{
An \code{R6Class} generator object
}
\description{
ConsensusCriterionDeNovo Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConsensusCriterionDeNovo-new}{\code{ConsensusCriterionDeNovo$new()}}
\item \href{#method-ConsensusCriterionDeNovo-toJSON}{\code{ConsensusCriterionDeNovo$toJSON()}}
\item \href{#method-ConsensusCriterionDeNovo-fromJSON}{\code{ConsensusCriterionDeNovo$fromJSON()}}
\item \href{#method-ConsensusCriterionDeNovo-toJSONString}{\code{ConsensusCriterionDeNovo$toJSONString()}}
\item \href{#method-ConsensusCriterionDeNovo-fromJSONString}{\code{ConsensusCriterionDeNovo$fromJSONString()}}
\item \href{#method-ConsensusCriterionDeNovo-clone}{\code{ConsensusCriterionDeNovo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusCriterionDeNovo-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusCriterionDeNovo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ConsensusCriterionDeNovo class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusCriterionDeNovo$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusCriterionDeNovo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusCriterionDeNovo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusCriterionDeNovo$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ConsensusCriterionDeNovo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusCriterionDeNovo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusCriterionDeNovo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ConsensusCriterionDeNovo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusCriterionDeNovo$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConsensusCriterionDeNovo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusCriterionDeNovo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusCriterionDeNovo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusCriterionDeNovo$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ConsensusCriterionDeNovo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusCriterionDeNovo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusCriterionDeNovo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ConsensusCriterionDeNovo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusCriterionDeNovo$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConsensusCriterionDeNovo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConsensusCriterionDeNovo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConsensusCriterionDeNovo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsensusCriterionDeNovo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
