% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_event.R
\docType{class}
\name{DataImportEvent}
\alias{DataImportEvent}
\title{DataImportEvent}
\format{
An \code{R6Class} generator object
}
\description{
DataImportEvent Class

DataImportEvent Class
}
\details{
Create a new DataImportEvent
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{importJobId}}{character [optional]}

\item{\code{importedCompoundIds}}{list(character)}

\item{\code{importedFeatureIds}}{list(character)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataImportEvent-new}{\code{DataImportEvent$new()}}
\item \href{#method-DataImportEvent-toJSON}{\code{DataImportEvent$toJSON()}}
\item \href{#method-DataImportEvent-fromJSON}{\code{DataImportEvent$fromJSON()}}
\item \href{#method-DataImportEvent-toJSONString}{\code{DataImportEvent$toJSONString()}}
\item \href{#method-DataImportEvent-fromJSONString}{\code{DataImportEvent$fromJSONString()}}
\item \href{#method-DataImportEvent-validateJSON}{\code{DataImportEvent$validateJSON()}}
\item \href{#method-DataImportEvent-toString}{\code{DataImportEvent$toString()}}
\item \href{#method-DataImportEvent-isValid}{\code{DataImportEvent$isValid()}}
\item \href{#method-DataImportEvent-getInvalidFields}{\code{DataImportEvent$getInvalidFields()}}
\item \href{#method-DataImportEvent-print}{\code{DataImportEvent$print()}}
\item \href{#method-DataImportEvent-clone}{\code{DataImportEvent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImportEvent-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataImportEvent-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DataImportEvent class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImportEvent$new(
  importedCompoundIds,
  importedFeatureIds,
  importJobId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{importedCompoundIds}}{importedCompoundIds}

\item{\code{importedFeatureIds}}{importedFeatureIds}

\item{\code{importJobId}}{importJobId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImportEvent-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataImportEvent-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImportEvent$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DataImportEvent in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImportEvent-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataImportEvent-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DataImportEvent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImportEvent$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DataImportEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImportEvent-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DataImportEvent-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImportEvent$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DataImportEvent in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImportEvent-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DataImportEvent-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DataImportEvent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImportEvent$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DataImportEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImportEvent-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataImportEvent-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to DataImportEvent and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImportEvent$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImportEvent-toString"></a>}}
\if{latex}{\out{\hypertarget{method-DataImportEvent-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImportEvent$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of DataImportEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImportEvent-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-DataImportEvent-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImportEvent$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImportEvent-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-DataImportEvent-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImportEvent$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImportEvent-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataImportEvent-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImportEvent$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImportEvent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataImportEvent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImportEvent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
