% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\docType{class}
\name{DataQuality}
\alias{DataQuality}
\title{DataQuality}
\format{
An \code{R6Class} generator object
}
\description{
DataQuality Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataQuality-new}{\code{DataQuality$new()}}
\item \href{#method-DataQuality-toJSON}{\code{DataQuality$toJSON()}}
\item \href{#method-DataQuality-fromJSON}{\code{DataQuality$fromJSON()}}
\item \href{#method-DataQuality-toJSONString}{\code{DataQuality$toJSONString()}}
\item \href{#method-DataQuality-fromJSONString}{\code{DataQuality$fromJSONString()}}
\item \href{#method-DataQuality-clone}{\code{DataQuality$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataQuality-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataQuality-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DataQuality class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataQuality$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataQuality-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataQuality-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataQuality$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DataQuality in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataQuality-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataQuality-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DataQuality
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataQuality$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DataQuality
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataQuality-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DataQuality-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataQuality$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DataQuality in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataQuality-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DataQuality-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DataQuality
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataQuality$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DataQuality
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataQuality-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataQuality-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataQuality$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
