% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_link.R
\docType{class}
\name{DBLink}
\alias{DBLink}
\title{DBLink}
\format{
An \code{R6Class} generator object
}
\description{
DBLink Class
}
\details{
Create a new DBLink
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{id}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DBLink-new}{\code{DBLink$new()}}
\item \href{#method-DBLink-toJSON}{\code{DBLink$toJSON()}}
\item \href{#method-DBLink-fromJSON}{\code{DBLink$fromJSON()}}
\item \href{#method-DBLink-toJSONString}{\code{DBLink$toJSONString()}}
\item \href{#method-DBLink-fromJSONString}{\code{DBLink$fromJSONString()}}
\item \href{#method-DBLink-validateJSON}{\code{DBLink$validateJSON()}}
\item \href{#method-DBLink-toString}{\code{DBLink$toString()}}
\item \href{#method-DBLink-isValid}{\code{DBLink$isValid()}}
\item \href{#method-DBLink-getInvalidFields}{\code{DBLink$getInvalidFields()}}
\item \href{#method-DBLink-print}{\code{DBLink$print()}}
\item \href{#method-DBLink-clone}{\code{DBLink$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBLink-new"></a>}}
\if{latex}{\out{\hypertarget{method-DBLink-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DBLink class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBLink$new(name = NULL, id = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{id}}{id}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBLink-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DBLink-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBLink$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DBLink in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBLink-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DBLink-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DBLink
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBLink$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DBLink
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBLink-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DBLink-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBLink$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DBLink in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBLink-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DBLink-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DBLink
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBLink$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DBLink
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBLink-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DBLink-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to DBLink and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBLink$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBLink-toString"></a>}}
\if{latex}{\out{\hypertarget{method-DBLink-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBLink$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of DBLink
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBLink-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-DBLink-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBLink$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBLink-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-DBLink-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBLink$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBLink-print"></a>}}
\if{latex}{\out{\hypertarget{method-DBLink-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBLink$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBLink-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DBLink-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBLink$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
