% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\docType{class}
\name{Feature}
\alias{Feature}
\title{Feature}
\format{
An \code{R6Class} generator object
}
\description{
A feature as it has been detected in one single run, in contrast to an AlignedFeature which combines the  features of the same compound over all runs it was detected in.  <p>  It provides the properties that are specific to the run it was detected in, such as its position on the  retention time axis of that run and its quantity within that run.

Feature Class
}
\details{
Create a new Feature
}
\examples{

## ------------------------------------------------
## Method `Feature$toList`
## ------------------------------------------------

# convert array of Feature (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{featureId}}{Identifier character [optional]}

\item{\code{alignedFeatureId}}{ID of the AlignedFeature this feature belongs to character [optional]}

\item{\code{runId}}{ID of the run this feature belongs to character [optional]}

\item{\code{averageMz}}{Average m/z over the whole feature numeric [optional]}

\item{\code{apexMz}}{m/z at the apex of the feature, the m/z this feature was measured at in its run numeric [optional]}

\item{\code{rtStartSeconds}}{Start of the feature on the retention time axis in seconds numeric [optional]}

\item{\code{rtEndSeconds}}{End of the feature on the retention time axis in seconds numeric [optional]}

\item{\code{rtApexSeconds}}{Apex of the feature on the retention time axis in seconds numeric [optional]}

\item{\code{rtFwhmSeconds}}{Full width at half maximum of the feature on the retention time axis in seconds numeric [optional]}

\item{\code{apexIntensity}}{Feature quantity measured as the intensity of the apex of the feature numeric [optional]}

\item{\code{areaUnderCurve}}{Feature quantity measured as the area under the curve of the whole feature numeric [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Feature-new}{\code{Feature$new()}}
\item \href{#method-Feature-toJSON}{\code{Feature$toJSON()}}
\item \href{#method-Feature-toList}{\code{Feature$toList()}}
\item \href{#method-Feature-toSimpleType}{\code{Feature$toSimpleType()}}
\item \href{#method-Feature-fromJSON}{\code{Feature$fromJSON()}}
\item \href{#method-Feature-toJSONString}{\code{Feature$toJSONString()}}
\item \href{#method-Feature-fromJSONString}{\code{Feature$fromJSONString()}}
\item \href{#method-Feature-validateJSON}{\code{Feature$validateJSON()}}
\item \href{#method-Feature-toString}{\code{Feature$toString()}}
\item \href{#method-Feature-isValid}{\code{Feature$isValid()}}
\item \href{#method-Feature-getInvalidFields}{\code{Feature$getInvalidFields()}}
\item \href{#method-Feature-print}{\code{Feature$print()}}
\item \href{#method-Feature-clone}{\code{Feature$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-new"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Feature class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$new(
  featureId = NULL,
  alignedFeatureId = NULL,
  runId = NULL,
  averageMz = NULL,
  apexMz = NULL,
  rtStartSeconds = NULL,
  rtEndSeconds = NULL,
  rtApexSeconds = NULL,
  rtFwhmSeconds = NULL,
  apexIntensity = NULL,
  areaUnderCurve = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{featureId}}{Identifier}

\item{\code{alignedFeatureId}}{ID of the AlignedFeature this feature belongs to}

\item{\code{runId}}{ID of the run this feature belongs to}

\item{\code{averageMz}}{Average m/z over the whole feature}

\item{\code{apexMz}}{m/z at the apex of the feature, the m/z this feature was measured at in its run}

\item{\code{rtStartSeconds}}{Start of the feature on the retention time axis in seconds}

\item{\code{rtEndSeconds}}{End of the feature on the retention time axis in seconds}

\item{\code{rtApexSeconds}}{Apex of the feature on the retention time axis in seconds}

\item{\code{rtFwhmSeconds}}{Full width at half maximum of the feature on the retention time axis in seconds}

\item{\code{apexIntensity}}{Feature quantity measured as the intensity of the apex of the feature}

\item{\code{areaUnderCurve}}{Feature quantity measured as the area under the curve of the whole feature}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Feature as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of Feature (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert Feature to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Feature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Feature
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Feature in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Feature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Feature
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Feature and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Feature
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-print"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
