% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_annotations.R
\docType{class}
\name{FeatureAnnotations}
\alias{FeatureAnnotations}
\title{FeatureAnnotations}
\format{
An \code{R6Class} generator object
}
\description{
Summary of the results of a feature (aligned over runs). Can be added to a AlignedFeature.  The different annotation fields within this summary object are null if the corresponding  feature does not contain the represented results. If fields are non-null  the corresponding result has been computed but might still be empty.

FeatureAnnotations Class
}
\details{
Create a new FeatureAnnotations
}
\examples{

## ------------------------------------------------
## Method `FeatureAnnotations$toList`
## ------------------------------------------------

# convert array of FeatureAnnotations (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{formulaAnnotation}}{\link{FormulaCandidate} [optional]}

\item{\code{structureAnnotation}}{\link{StructureCandidateScored} [optional]}

\item{\code{compoundClassAnnotation}}{\link{CompoundClasses} [optional]}

\item{\code{confidenceExactMatch}}{Confidence Score that represents the confidence whether the top hit is correct. numeric [optional]}

\item{\code{confidenceApproxMatch}}{Confidence Score that represents the confidence whether the top hit or a very similar hit (estimated by MCES distance) is correct. numeric [optional]}

\item{\code{expansiveSearchState}}{Result that shows if structure annotation was expanded by using PubChem as fallback and if so, which confidence mode was used (as per input parameter) character [optional]}

\item{\code{specifiedDatabases}}{List of databases that have been specified for structure db search. Null if no structure db search has been performed. list(character) [optional]}

\item{\code{expandedDatabases}}{List of databases that have been used to expand search space during expansive search. Null if no structure db search has been performed. list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureAnnotations-new}{\code{FeatureAnnotations$new()}}
\item \href{#method-FeatureAnnotations-toJSON}{\code{FeatureAnnotations$toJSON()}}
\item \href{#method-FeatureAnnotations-toList}{\code{FeatureAnnotations$toList()}}
\item \href{#method-FeatureAnnotations-toSimpleType}{\code{FeatureAnnotations$toSimpleType()}}
\item \href{#method-FeatureAnnotations-fromJSON}{\code{FeatureAnnotations$fromJSON()}}
\item \href{#method-FeatureAnnotations-toJSONString}{\code{FeatureAnnotations$toJSONString()}}
\item \href{#method-FeatureAnnotations-fromJSONString}{\code{FeatureAnnotations$fromJSONString()}}
\item \href{#method-FeatureAnnotations-validateJSON}{\code{FeatureAnnotations$validateJSON()}}
\item \href{#method-FeatureAnnotations-toString}{\code{FeatureAnnotations$toString()}}
\item \href{#method-FeatureAnnotations-isValid}{\code{FeatureAnnotations$isValid()}}
\item \href{#method-FeatureAnnotations-getInvalidFields}{\code{FeatureAnnotations$getInvalidFields()}}
\item \href{#method-FeatureAnnotations-print}{\code{FeatureAnnotations$print()}}
\item \href{#method-FeatureAnnotations-clone}{\code{FeatureAnnotations$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnnotations-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnnotations-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FeatureAnnotations class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnnotations$new(
  formulaAnnotation = NULL,
  structureAnnotation = NULL,
  compoundClassAnnotation = NULL,
  confidenceExactMatch = NULL,
  confidenceApproxMatch = NULL,
  expansiveSearchState = NULL,
  specifiedDatabases = NULL,
  expandedDatabases = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formulaAnnotation}}{formulaAnnotation}

\item{\code{structureAnnotation}}{structureAnnotation}

\item{\code{compoundClassAnnotation}}{compoundClassAnnotation}

\item{\code{confidenceExactMatch}}{Confidence Score that represents the confidence whether the top hit is correct.}

\item{\code{confidenceApproxMatch}}{Confidence Score that represents the confidence whether the top hit or a very similar hit (estimated by MCES distance) is correct.}

\item{\code{expansiveSearchState}}{Result that shows if structure annotation was expanded by using PubChem as fallback and if so, which confidence mode was used (as per input parameter)}

\item{\code{specifiedDatabases}}{List of databases that have been specified for structure db search. Null if no structure db search has been performed.}

\item{\code{expandedDatabases}}{List of databases that have been used to expand search space during expansive search. Null if no structure db search has been performed.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnnotations-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnnotations-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnnotations$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnnotations-toList"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnnotations-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnnotations$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FeatureAnnotations as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of FeatureAnnotations (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnnotations-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnnotations-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert FeatureAnnotations to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnnotations$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnnotations-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnnotations-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FeatureAnnotations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnnotations$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FeatureAnnotations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnnotations-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnnotations-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnnotations$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
FeatureAnnotations in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnnotations-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnnotations-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FeatureAnnotations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnnotations$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FeatureAnnotations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnnotations-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnnotations-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FeatureAnnotations and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnnotations$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnnotations-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnnotations-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnnotations$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FeatureAnnotations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnnotations-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnnotations-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnnotations$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnnotations-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnnotations-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnnotations$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnnotations-print"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnnotations-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnnotations$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnnotations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnnotations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnnotations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
