% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_import.R
\docType{class}
\name{FeatureImport}
\alias{FeatureImport}
\title{FeatureImport}
\format{
An \code{R6Class} generator object
}
\description{
Represents an (aligned) feature to be imported into a SIRIUS project.  At least one of the Mass Spec data sources (e.g. mergedMs1, ms1Spectra, ms2Spectra) needs to be given.  Otherwise, the import will fail.

FeatureImport Class
}
\details{
Create a new FeatureImport
}
\examples{

## ------------------------------------------------
## Method `FeatureImport$toList`
## ------------------------------------------------

# convert array of FeatureImport (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{externalFeatureId}}{Externally provided FeatureId (by some preprocessing tool). This FeatureId is NOT used by SIRIUS but is stored to ease mapping information back to the source. character [optional]}

\item{\code{ionMass}}{numeric}

\item{\code{charge}}{integer}

\item{\code{detectedAdducts}}{Detected adducts of this feature. Can be NULL or empty if no adducts are known. list(character) [optional]}

\item{\code{rtStartSeconds}}{numeric [optional]}

\item{\code{rtEndSeconds}}{numeric [optional]}

\item{\code{rtApexSeconds}}{numeric [optional]}

\item{\code{dataQuality}}{A optional feature quality flag that can be used to filter features to be shown in the gui or to be considered for further analysis. character [optional]}

\item{\code{mergedMs1}}{\link{BasicSpectrum} [optional]}

\item{\code{ms1Spectra}}{List of MS1Spectra belonging to this feature. These spectra will be merged and only a representative  mergedMs1 spectrum will be stored in SIRIUS. At least one of these spectra should contain the  isotope pattern of the precursor ion.  Note: Will be ignored if 'mergedMs1' is given. list(\link{BasicSpectrum}) [optional]}

\item{\code{ms2Spectra}}{List of MS/MS spectra that belong to this feature. list(\link{BasicSpectrum}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureImport-new}{\code{FeatureImport$new()}}
\item \href{#method-FeatureImport-toJSON}{\code{FeatureImport$toJSON()}}
\item \href{#method-FeatureImport-toList}{\code{FeatureImport$toList()}}
\item \href{#method-FeatureImport-toSimpleType}{\code{FeatureImport$toSimpleType()}}
\item \href{#method-FeatureImport-fromJSON}{\code{FeatureImport$fromJSON()}}
\item \href{#method-FeatureImport-toJSONString}{\code{FeatureImport$toJSONString()}}
\item \href{#method-FeatureImport-fromJSONString}{\code{FeatureImport$fromJSONString()}}
\item \href{#method-FeatureImport-validateJSON}{\code{FeatureImport$validateJSON()}}
\item \href{#method-FeatureImport-toString}{\code{FeatureImport$toString()}}
\item \href{#method-FeatureImport-isValid}{\code{FeatureImport$isValid()}}
\item \href{#method-FeatureImport-getInvalidFields}{\code{FeatureImport$getInvalidFields()}}
\item \href{#method-FeatureImport-print}{\code{FeatureImport$print()}}
\item \href{#method-FeatureImport-clone}{\code{FeatureImport$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImport-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImport-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FeatureImport class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImport$new(
  ionMass,
  charge,
  name = NULL,
  externalFeatureId = NULL,
  detectedAdducts = NULL,
  rtStartSeconds = NULL,
  rtEndSeconds = NULL,
  rtApexSeconds = NULL,
  dataQuality = NULL,
  mergedMs1 = NULL,
  ms1Spectra = NULL,
  ms2Spectra = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ionMass}}{ionMass}

\item{\code{charge}}{charge}

\item{\code{name}}{name}

\item{\code{externalFeatureId}}{Externally provided FeatureId (by some preprocessing tool). This FeatureId is NOT used by SIRIUS but is stored to ease mapping information back to the source.}

\item{\code{detectedAdducts}}{Detected adducts of this feature. Can be NULL or empty if no adducts are known.}

\item{\code{rtStartSeconds}}{rtStartSeconds}

\item{\code{rtEndSeconds}}{rtEndSeconds}

\item{\code{rtApexSeconds}}{rtApexSeconds}

\item{\code{dataQuality}}{A optional feature quality flag that can be used to filter features to be shown in the gui or to be considered for further analysis.}

\item{\code{mergedMs1}}{mergedMs1}

\item{\code{ms1Spectra}}{List of MS1Spectra belonging to this feature. These spectra will be merged and only a representative  mergedMs1 spectrum will be stored in SIRIUS. At least one of these spectra should contain the  isotope pattern of the precursor ion.  Note: Will be ignored if 'mergedMs1' is given.}

\item{\code{ms2Spectra}}{List of MS/MS spectra that belong to this feature.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImport-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImport-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImport$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImport-toList"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImport-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImport$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FeatureImport as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of FeatureImport (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImport-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImport-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert FeatureImport to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImport$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImport-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImport-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FeatureImport
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImport$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FeatureImport
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImport-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImport-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImport$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
FeatureImport in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImport-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImport-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FeatureImport
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImport$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FeatureImport
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImport-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImport-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FeatureImport and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImport$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImport-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImport-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImport$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FeatureImport
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImport-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImport-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImport$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImport-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImport-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImport$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImport-print"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImport-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImport$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImport-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImport-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImport$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
