% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint_prediction.R
\docType{class}
\name{FingerprintPrediction}
\alias{FingerprintPrediction}
\title{FingerprintPrediction}
\format{
An \code{R6Class} generator object
}
\description{
User/developer friendly parameter subset for the CSI:FingerID Fingerprint tool  Needs results from the Formula identification tool (Sirius).

FingerprintPrediction Class
}
\details{
Create a new FingerprintPrediction
}
\examples{

## ------------------------------------------------
## Method `FingerprintPrediction$toList`
## ------------------------------------------------

# convert array of FingerprintPrediction (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{enabled}}{Indicates whether the tool is enabled. character [optional]}

\item{\code{useScoreThreshold}}{If true, an adaptive soft threshold will be applied to only compute Fingerprints for promising formula candidates  Enabling is highly recommended. character [optional]}

\item{\code{alwaysPredictHighRefMatches}}{If true, fingerprints, compound classes and structures will be predicted for formula candidates whose  reference spectrum similarity is above Sirius.minReferenceMatchScoreToInject, no matter which  score threshold rules would otherwise apply.  If NULL default value will be used. character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FingerprintPrediction-new}{\code{FingerprintPrediction$new()}}
\item \href{#method-FingerprintPrediction-toJSON}{\code{FingerprintPrediction$toJSON()}}
\item \href{#method-FingerprintPrediction-toList}{\code{FingerprintPrediction$toList()}}
\item \href{#method-FingerprintPrediction-toSimpleType}{\code{FingerprintPrediction$toSimpleType()}}
\item \href{#method-FingerprintPrediction-fromJSON}{\code{FingerprintPrediction$fromJSON()}}
\item \href{#method-FingerprintPrediction-toJSONString}{\code{FingerprintPrediction$toJSONString()}}
\item \href{#method-FingerprintPrediction-fromJSONString}{\code{FingerprintPrediction$fromJSONString()}}
\item \href{#method-FingerprintPrediction-validateJSON}{\code{FingerprintPrediction$validateJSON()}}
\item \href{#method-FingerprintPrediction-toString}{\code{FingerprintPrediction$toString()}}
\item \href{#method-FingerprintPrediction-isValid}{\code{FingerprintPrediction$isValid()}}
\item \href{#method-FingerprintPrediction-getInvalidFields}{\code{FingerprintPrediction$getInvalidFields()}}
\item \href{#method-FingerprintPrediction-print}{\code{FingerprintPrediction$print()}}
\item \href{#method-FingerprintPrediction-clone}{\code{FingerprintPrediction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FingerprintPrediction-new"></a>}}
\if{latex}{\out{\hypertarget{method-FingerprintPrediction-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FingerprintPrediction class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FingerprintPrediction$new(
  enabled = NULL,
  useScoreThreshold = NULL,
  alwaysPredictHighRefMatches = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enabled}}{Indicates whether the tool is enabled.}

\item{\code{useScoreThreshold}}{If true, an adaptive soft threshold will be applied to only compute Fingerprints for promising formula candidates  Enabling is highly recommended.}

\item{\code{alwaysPredictHighRefMatches}}{If true, fingerprints, compound classes and structures will be predicted for formula candidates whose  reference spectrum similarity is above Sirius.minReferenceMatchScoreToInject, no matter which  score threshold rules would otherwise apply.  If NULL default value will be used.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FingerprintPrediction-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FingerprintPrediction-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FingerprintPrediction$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FingerprintPrediction-toList"></a>}}
\if{latex}{\out{\hypertarget{method-FingerprintPrediction-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FingerprintPrediction$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FingerprintPrediction as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of FingerprintPrediction (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FingerprintPrediction-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-FingerprintPrediction-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert FingerprintPrediction to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FingerprintPrediction$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FingerprintPrediction-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FingerprintPrediction-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FingerprintPrediction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FingerprintPrediction$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FingerprintPrediction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FingerprintPrediction-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FingerprintPrediction-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FingerprintPrediction$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
FingerprintPrediction in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FingerprintPrediction-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FingerprintPrediction-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FingerprintPrediction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FingerprintPrediction$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FingerprintPrediction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FingerprintPrediction-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FingerprintPrediction-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FingerprintPrediction and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FingerprintPrediction$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FingerprintPrediction-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FingerprintPrediction-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FingerprintPrediction$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FingerprintPrediction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FingerprintPrediction-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FingerprintPrediction-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FingerprintPrediction$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FingerprintPrediction-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FingerprintPrediction-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FingerprintPrediction$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FingerprintPrediction-print"></a>}}
\if{latex}{\out{\hypertarget{method-FingerprintPrediction-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FingerprintPrediction$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FingerprintPrediction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FingerprintPrediction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FingerprintPrediction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
