% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change_job_submission.R
\docType{class}
\name{FoldChangeJobSubmission}
\alias{FoldChangeJobSubmission}
\title{FoldChangeJobSubmission}
\format{
An \code{R6Class} generator object
}
\description{
FoldChangeJobSubmission Class

FoldChangeJobSubmission Class
}
\details{
Create a new FoldChangeJobSubmission
}
\examples{

## ------------------------------------------------
## Method `FoldChangeJobSubmission$toList`
## ------------------------------------------------

# convert array of FoldChangeJobSubmission (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{leftRunGroup}}{character [optional]}

\item{\code{rightRunGroup}}{character [optional]}

\item{\code{aggregationTypes}}{list(character) [optional]}

\item{\code{quantificationMeasures}}{list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FoldChangeJobSubmission-new}{\code{FoldChangeJobSubmission$new()}}
\item \href{#method-FoldChangeJobSubmission-toJSON}{\code{FoldChangeJobSubmission$toJSON()}}
\item \href{#method-FoldChangeJobSubmission-toList}{\code{FoldChangeJobSubmission$toList()}}
\item \href{#method-FoldChangeJobSubmission-toSimpleType}{\code{FoldChangeJobSubmission$toSimpleType()}}
\item \href{#method-FoldChangeJobSubmission-fromJSON}{\code{FoldChangeJobSubmission$fromJSON()}}
\item \href{#method-FoldChangeJobSubmission-toJSONString}{\code{FoldChangeJobSubmission$toJSONString()}}
\item \href{#method-FoldChangeJobSubmission-fromJSONString}{\code{FoldChangeJobSubmission$fromJSONString()}}
\item \href{#method-FoldChangeJobSubmission-validateJSON}{\code{FoldChangeJobSubmission$validateJSON()}}
\item \href{#method-FoldChangeJobSubmission-toString}{\code{FoldChangeJobSubmission$toString()}}
\item \href{#method-FoldChangeJobSubmission-isValid}{\code{FoldChangeJobSubmission$isValid()}}
\item \href{#method-FoldChangeJobSubmission-getInvalidFields}{\code{FoldChangeJobSubmission$getInvalidFields()}}
\item \href{#method-FoldChangeJobSubmission-print}{\code{FoldChangeJobSubmission$print()}}
\item \href{#method-FoldChangeJobSubmission-clone}{\code{FoldChangeJobSubmission$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoldChangeJobSubmission-new"></a>}}
\if{latex}{\out{\hypertarget{method-FoldChangeJobSubmission-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FoldChangeJobSubmission class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoldChangeJobSubmission$new(
  leftRunGroup = NULL,
  rightRunGroup = NULL,
  aggregationTypes = NULL,
  quantificationMeasures = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{leftRunGroup}}{leftRunGroup}

\item{\code{rightRunGroup}}{rightRunGroup}

\item{\code{aggregationTypes}}{aggregationTypes}

\item{\code{quantificationMeasures}}{quantificationMeasures}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoldChangeJobSubmission-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FoldChangeJobSubmission-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoldChangeJobSubmission$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoldChangeJobSubmission-toList"></a>}}
\if{latex}{\out{\hypertarget{method-FoldChangeJobSubmission-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoldChangeJobSubmission$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FoldChangeJobSubmission as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of FoldChangeJobSubmission (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoldChangeJobSubmission-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-FoldChangeJobSubmission-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert FoldChangeJobSubmission to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoldChangeJobSubmission$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoldChangeJobSubmission-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FoldChangeJobSubmission-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FoldChangeJobSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoldChangeJobSubmission$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FoldChangeJobSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoldChangeJobSubmission-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FoldChangeJobSubmission-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoldChangeJobSubmission$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
FoldChangeJobSubmission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoldChangeJobSubmission-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FoldChangeJobSubmission-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FoldChangeJobSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoldChangeJobSubmission$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FoldChangeJobSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoldChangeJobSubmission-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FoldChangeJobSubmission-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FoldChangeJobSubmission and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoldChangeJobSubmission$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoldChangeJobSubmission-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FoldChangeJobSubmission-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoldChangeJobSubmission$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FoldChangeJobSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoldChangeJobSubmission-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FoldChangeJobSubmission-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoldChangeJobSubmission$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoldChangeJobSubmission-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FoldChangeJobSubmission-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoldChangeJobSubmission$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoldChangeJobSubmission-print"></a>}}
\if{latex}{\out{\hypertarget{method-FoldChangeJobSubmission-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoldChangeJobSubmission$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoldChangeJobSubmission-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FoldChangeJobSubmission-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoldChangeJobSubmission$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
