% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_candidate_opt_field.R
\docType{class}
\name{FormulaCandidateOptField}
\alias{FormulaCandidateOptField}
\title{FormulaCandidateOptField}
\format{
An \code{R6Class} generator object
}
\description{
FormulaCandidateOptField Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FormulaCandidateOptField-new}{\code{FormulaCandidateOptField$new()}}
\item \href{#method-FormulaCandidateOptField-toJSON}{\code{FormulaCandidateOptField$toJSON()}}
\item \href{#method-FormulaCandidateOptField-fromJSON}{\code{FormulaCandidateOptField$fromJSON()}}
\item \href{#method-FormulaCandidateOptField-toJSONString}{\code{FormulaCandidateOptField$toJSONString()}}
\item \href{#method-FormulaCandidateOptField-fromJSONString}{\code{FormulaCandidateOptField$fromJSONString()}}
\item \href{#method-FormulaCandidateOptField-clone}{\code{FormulaCandidateOptField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormulaCandidateOptField-new"></a>}}
\if{latex}{\out{\hypertarget{method-FormulaCandidateOptField-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FormulaCandidateOptField class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormulaCandidateOptField$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormulaCandidateOptField-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FormulaCandidateOptField-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormulaCandidateOptField$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FormulaCandidateOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormulaCandidateOptField-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FormulaCandidateOptField-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FormulaCandidateOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormulaCandidateOptField$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FormulaCandidateOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormulaCandidateOptField-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FormulaCandidateOptField-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormulaCandidateOptField$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FormulaCandidateOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormulaCandidateOptField-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FormulaCandidateOptField-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FormulaCandidateOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormulaCandidateOptField$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FormulaCandidateOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormulaCandidateOptField-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FormulaCandidateOptField-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormulaCandidateOptField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
