% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_result.R
\docType{class}
\name{ImportResult}
\alias{ImportResult}
\title{ImportResult}
\format{
An \code{R6Class} generator object
}
\description{
ImportResult Class

ImportResult Class
}
\details{
Create a new ImportResult
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{affectedCompoundIds}}{List of compoundIds that have been imported. list(character)}

\item{\code{affectedAlignedFeatureIds}}{List of alignedFeatureIds that have been imported.. list(character)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImportResult-new}{\code{ImportResult$new()}}
\item \href{#method-ImportResult-toJSON}{\code{ImportResult$toJSON()}}
\item \href{#method-ImportResult-fromJSON}{\code{ImportResult$fromJSON()}}
\item \href{#method-ImportResult-toJSONString}{\code{ImportResult$toJSONString()}}
\item \href{#method-ImportResult-fromJSONString}{\code{ImportResult$fromJSONString()}}
\item \href{#method-ImportResult-validateJSON}{\code{ImportResult$validateJSON()}}
\item \href{#method-ImportResult-toString}{\code{ImportResult$toString()}}
\item \href{#method-ImportResult-isValid}{\code{ImportResult$isValid()}}
\item \href{#method-ImportResult-getInvalidFields}{\code{ImportResult$getInvalidFields()}}
\item \href{#method-ImportResult-print}{\code{ImportResult$print()}}
\item \href{#method-ImportResult-clone}{\code{ImportResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImportResult-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ImportResult class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportResult$new(affectedCompoundIds, affectedAlignedFeatureIds, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{affectedCompoundIds}}{List of compoundIds that have been imported.}

\item{\code{affectedAlignedFeatureIds}}{List of alignedFeatureIds that have been imported..}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportResult-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImportResult-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportResult$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ImportResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportResult-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImportResult-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ImportResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportResult$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ImportResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportResult-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ImportResult-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportResult$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ImportResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportResult-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ImportResult-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ImportResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportResult$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ImportResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportResult-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImportResult-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ImportResult and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportResult$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportResult-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ImportResult-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportResult$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ImportResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportResult-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ImportResult-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportResult$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportResult-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ImportResult-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportResult$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportResult-print"></a>}}
\if{latex}{\out{\hypertarget{method-ImportResult-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportResult$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImportResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
