% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\docType{class}
\name{Info}
\alias{Info}
\title{Info}
\format{
An \code{R6Class} generator object
}
\description{
Info Class

Info Class
}
\details{
Create a new Info
}
\examples{

## ------------------------------------------------
## Method `Info$toList`
## ------------------------------------------------

# convert array of Info (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nightSkyApiVersion}}{API version of the SIRIUS Nightsky API character [optional]}

\item{\code{siriusVersion}}{Version of the SIRIUS application character [optional]}

\item{\code{latestSiriusVersion}}{Latest available Version of the SIRIUS application character [optional]}

\item{\code{latestSiriusLink}}{Link to the latest available Version of the SIRIUS application character [optional]}

\item{\code{updateAvailable}}{True if a newer SIRIUS version is available. character [optional]}

\item{\code{siriusLibVersion}}{Version of the SIRIUS libraries character [optional]}

\item{\code{fingerIdLibVersion}}{Version of the CSI:FingerID libraries character [optional]}

\item{\code{chemDbVersion}}{Version of the Chemical Database available via SIRIUS web services character [optional]}

\item{\code{fingerIdModelVersion}}{Version of the Machine learning models used for Fingerprint, Compound Class and Structure Prediction  Not available if web service is not reachable. character [optional]}

\item{\code{fingerprintId}}{Version of the Molecular Fingerprint used by SIRIUS character [optional]}

\item{\code{availableILPSolvers}}{Set of solvers that are configured correctly and can be loaded list(character)}

\item{\code{supportedILPSolvers}}{Set of ILP solvers that are supported, and their version information named list(character)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Info-new}{\code{Info$new()}}
\item \href{#method-Info-toJSON}{\code{Info$toJSON()}}
\item \href{#method-Info-toList}{\code{Info$toList()}}
\item \href{#method-Info-toSimpleType}{\code{Info$toSimpleType()}}
\item \href{#method-Info-fromJSON}{\code{Info$fromJSON()}}
\item \href{#method-Info-toJSONString}{\code{Info$toJSONString()}}
\item \href{#method-Info-fromJSONString}{\code{Info$fromJSONString()}}
\item \href{#method-Info-validateJSON}{\code{Info$validateJSON()}}
\item \href{#method-Info-toString}{\code{Info$toString()}}
\item \href{#method-Info-isValid}{\code{Info$isValid()}}
\item \href{#method-Info-getInvalidFields}{\code{Info$getInvalidFields()}}
\item \href{#method-Info-print}{\code{Info$print()}}
\item \href{#method-Info-clone}{\code{Info$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-new"></a>}}
\if{latex}{\out{\hypertarget{method-Info-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Info class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$new(
  availableILPSolvers,
  supportedILPSolvers,
  nightSkyApiVersion = NULL,
  siriusVersion = NULL,
  latestSiriusVersion = NULL,
  latestSiriusLink = NULL,
  updateAvailable = NULL,
  siriusLibVersion = NULL,
  fingerIdLibVersion = NULL,
  chemDbVersion = NULL,
  fingerIdModelVersion = NULL,
  fingerprintId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{availableILPSolvers}}{Set of solvers that are configured correctly and can be loaded}

\item{\code{supportedILPSolvers}}{Set of ILP solvers that are supported, and their version information}

\item{\code{nightSkyApiVersion}}{API version of the SIRIUS Nightsky API}

\item{\code{siriusVersion}}{Version of the SIRIUS application}

\item{\code{latestSiriusVersion}}{Latest available Version of the SIRIUS application}

\item{\code{latestSiriusLink}}{Link to the latest available Version of the SIRIUS application}

\item{\code{updateAvailable}}{True if a newer SIRIUS version is available.}

\item{\code{siriusLibVersion}}{Version of the SIRIUS libraries}

\item{\code{fingerIdLibVersion}}{Version of the CSI:FingerID libraries}

\item{\code{chemDbVersion}}{Version of the Chemical Database available via SIRIUS web services}

\item{\code{fingerIdModelVersion}}{Version of the Machine learning models used for Fingerprint, Compound Class and Structure Prediction  Not available if web service is not reachable.}

\item{\code{fingerprintId}}{Version of the Molecular Fingerprint used by SIRIUS}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Info-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Info-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Info as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of Info (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-Info-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert Info to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Info-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Info
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Info-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Info in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Info-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Info
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Info-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Info and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Info-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Info
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Info-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Info-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-print"></a>}}
\if{latex}{\out{\hypertarget{method-Info-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Info-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
