% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_api.R
\docType{class}
\name{InfoApi}
\alias{InfoApi}
\title{Info operations}
\format{
An \code{R6Class} generator object
}
\description{
InfoApi
}
\details{
SIRIUS Nightsky API

REST API that provides the full functionality of SIRIUS and its web services as background service. It is intended as entry-point for scripting languages and software integration SDKs.This API is exposed by SIRIUS 6

The version of the OpenAPI document: 3.2
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  GetConnectionCheck  ####################

library(RSirius)

api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetConnectionCheck(data_file = "result.txt")
result <- api_instance$info_api$GetConnectionCheck()
dput(result)


####################  GetInfo  ####################

library(RSirius)
var_server_info <- TRUE # character |  (Optional)
var_update_info <- TRUE # character |  (Optional)

api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetInfo(server_info = var_server_info, update_info = var_update_infodata_file = "result.txt")
result <- api_instance$info_api$GetInfo(server_info = var_server_info, update_info = var_update_info)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InfoApi-new}{\code{InfoApi$new()}}
\item \href{#method-InfoApi-GetConnectionCheck}{\code{InfoApi$GetConnectionCheck()}}
\item \href{#method-InfoApi-GetConnectionCheckWithHttpInfo}{\code{InfoApi$GetConnectionCheckWithHttpInfo()}}
\item \href{#method-InfoApi-GetInfo}{\code{InfoApi$GetInfo()}}
\item \href{#method-InfoApi-GetInfoWithHttpInfo}{\code{InfoApi$GetInfoWithHttpInfo()}}
\item \href{#method-InfoApi-clone}{\code{InfoApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-InfoApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InfoApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoApi-GetConnectionCheck"></a>}}
\if{latex}{\out{\hypertarget{method-InfoApi-GetConnectionCheck}{}}}
\subsection{Method \code{GetConnectionCheck()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoApi$GetConnectionCheck(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ConnectionCheck
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoApi-GetConnectionCheckWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-InfoApi-GetConnectionCheckWithHttpInfo}{}}}
\subsection{Method \code{GetConnectionCheckWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoApi$GetConnectionCheckWithHttpInfo(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (ConnectionCheck) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoApi-GetInfo"></a>}}
\if{latex}{\out{\hypertarget{method-InfoApi-GetInfo}{}}}
\subsection{Method \code{GetInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoApi$GetInfo(server_info = TRUE, update_info = TRUE, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{server_info}}{(optional) No description (default value: TRUE)}

\item{\code{update_info}}{(optional) No description (default value: TRUE)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Info
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoApi-GetInfoWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-InfoApi-GetInfoWithHttpInfo}{}}}
\subsection{Method \code{GetInfoWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoApi$GetInfoWithHttpInfo(
  server_info = TRUE,
  update_info = TRUE,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{server_info}}{(optional) No description (default value: TRUE)}

\item{\code{update_info}}{(optional) No description (default value: TRUE)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (Info) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InfoApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
