% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument.R
\docType{class}
\name{Instrument}
\alias{Instrument}
\title{Instrument}
\format{
An \code{R6Class} generator object
}
\description{
Instrument Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Instrument-new}{\code{Instrument$new()}}
\item \href{#method-Instrument-toJSON}{\code{Instrument$toJSON()}}
\item \href{#method-Instrument-fromJSON}{\code{Instrument$fromJSON()}}
\item \href{#method-Instrument-toJSONString}{\code{Instrument$toJSONString()}}
\item \href{#method-Instrument-fromJSONString}{\code{Instrument$fromJSONString()}}
\item \href{#method-Instrument-clone}{\code{Instrument$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-new"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Instrument class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Instrument in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Instrument
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Instrument
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Instrument in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Instrument
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Instrument
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
