% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_profile.R
\docType{class}
\name{InstrumentProfile}
\alias{InstrumentProfile}
\title{InstrumentProfile}
\format{
An \code{R6Class} generator object
}
\description{
InstrumentProfile Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InstrumentProfile-new}{\code{InstrumentProfile$new()}}
\item \href{#method-InstrumentProfile-toJSON}{\code{InstrumentProfile$toJSON()}}
\item \href{#method-InstrumentProfile-fromJSON}{\code{InstrumentProfile$fromJSON()}}
\item \href{#method-InstrumentProfile-toJSONString}{\code{InstrumentProfile$toJSONString()}}
\item \href{#method-InstrumentProfile-fromJSONString}{\code{InstrumentProfile$fromJSONString()}}
\item \href{#method-InstrumentProfile-clone}{\code{InstrumentProfile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InstrumentProfile-new"></a>}}
\if{latex}{\out{\hypertarget{method-InstrumentProfile-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InstrumentProfile class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstrumentProfile$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InstrumentProfile-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InstrumentProfile-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstrumentProfile$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InstrumentProfile in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InstrumentProfile-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InstrumentProfile-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of InstrumentProfile
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstrumentProfile$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InstrumentProfile
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InstrumentProfile-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InstrumentProfile-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstrumentProfile$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InstrumentProfile in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InstrumentProfile-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InstrumentProfile-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of InstrumentProfile
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstrumentProfile$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InstrumentProfile
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InstrumentProfile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InstrumentProfile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstrumentProfile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
