% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_pattern_annotation.R
\docType{class}
\name{IsotopePatternAnnotation}
\alias{IsotopePatternAnnotation}
\title{IsotopePatternAnnotation}
\format{
An \code{R6Class} generator object
}
\description{
IsotopePatternAnnotation Class

IsotopePatternAnnotation Class
}
\details{
Create a new IsotopePatternAnnotation
}
\examples{

## ------------------------------------------------
## Method `IsotopePatternAnnotation$toList`
## ------------------------------------------------

# convert array of IsotopePatternAnnotation (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isotopePattern}}{\link{BasicSpectrum} [optional]}

\item{\code{simulatedPattern}}{\link{BasicSpectrum} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IsotopePatternAnnotation-new}{\code{IsotopePatternAnnotation$new()}}
\item \href{#method-IsotopePatternAnnotation-toJSON}{\code{IsotopePatternAnnotation$toJSON()}}
\item \href{#method-IsotopePatternAnnotation-toList}{\code{IsotopePatternAnnotation$toList()}}
\item \href{#method-IsotopePatternAnnotation-toSimpleType}{\code{IsotopePatternAnnotation$toSimpleType()}}
\item \href{#method-IsotopePatternAnnotation-fromJSON}{\code{IsotopePatternAnnotation$fromJSON()}}
\item \href{#method-IsotopePatternAnnotation-toJSONString}{\code{IsotopePatternAnnotation$toJSONString()}}
\item \href{#method-IsotopePatternAnnotation-fromJSONString}{\code{IsotopePatternAnnotation$fromJSONString()}}
\item \href{#method-IsotopePatternAnnotation-validateJSON}{\code{IsotopePatternAnnotation$validateJSON()}}
\item \href{#method-IsotopePatternAnnotation-toString}{\code{IsotopePatternAnnotation$toString()}}
\item \href{#method-IsotopePatternAnnotation-isValid}{\code{IsotopePatternAnnotation$isValid()}}
\item \href{#method-IsotopePatternAnnotation-getInvalidFields}{\code{IsotopePatternAnnotation$getInvalidFields()}}
\item \href{#method-IsotopePatternAnnotation-print}{\code{IsotopePatternAnnotation$print()}}
\item \href{#method-IsotopePatternAnnotation-clone}{\code{IsotopePatternAnnotation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsotopePatternAnnotation-new"></a>}}
\if{latex}{\out{\hypertarget{method-IsotopePatternAnnotation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new IsotopePatternAnnotation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsotopePatternAnnotation$new(
  isotopePattern = NULL,
  simulatedPattern = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isotopePattern}}{isotopePattern}

\item{\code{simulatedPattern}}{simulatedPattern}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsotopePatternAnnotation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-IsotopePatternAnnotation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsotopePatternAnnotation$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsotopePatternAnnotation-toList"></a>}}
\if{latex}{\out{\hypertarget{method-IsotopePatternAnnotation-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsotopePatternAnnotation$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
IsotopePatternAnnotation as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of IsotopePatternAnnotation (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsotopePatternAnnotation-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-IsotopePatternAnnotation-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert IsotopePatternAnnotation to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsotopePatternAnnotation$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsotopePatternAnnotation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-IsotopePatternAnnotation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of IsotopePatternAnnotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsotopePatternAnnotation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of IsotopePatternAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsotopePatternAnnotation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-IsotopePatternAnnotation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsotopePatternAnnotation$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
IsotopePatternAnnotation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsotopePatternAnnotation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-IsotopePatternAnnotation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of IsotopePatternAnnotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsotopePatternAnnotation$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of IsotopePatternAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsotopePatternAnnotation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-IsotopePatternAnnotation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to IsotopePatternAnnotation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsotopePatternAnnotation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsotopePatternAnnotation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-IsotopePatternAnnotation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsotopePatternAnnotation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of IsotopePatternAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsotopePatternAnnotation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-IsotopePatternAnnotation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsotopePatternAnnotation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsotopePatternAnnotation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-IsotopePatternAnnotation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsotopePatternAnnotation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsotopePatternAnnotation-print"></a>}}
\if{latex}{\out{\hypertarget{method-IsotopePatternAnnotation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsotopePatternAnnotation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsotopePatternAnnotation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IsotopePatternAnnotation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsotopePatternAnnotation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
