% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_opt_field.R
\docType{class}
\name{JobOptField}
\alias{JobOptField}
\title{JobOptField}
\format{
An \code{R6Class} generator object
}
\description{
JobOptField Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JobOptField-new}{\code{JobOptField$new()}}
\item \href{#method-JobOptField-toJSON}{\code{JobOptField$toJSON()}}
\item \href{#method-JobOptField-fromJSON}{\code{JobOptField$fromJSON()}}
\item \href{#method-JobOptField-toJSONString}{\code{JobOptField$toJSONString()}}
\item \href{#method-JobOptField-fromJSONString}{\code{JobOptField$fromJSONString()}}
\item \href{#method-JobOptField-clone}{\code{JobOptField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobOptField-new"></a>}}
\if{latex}{\out{\hypertarget{method-JobOptField-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new JobOptField class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobOptField$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobOptField-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobOptField-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobOptField$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JobOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobOptField-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobOptField-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of JobOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobOptField$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobOptField-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobOptField-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobOptField$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JobOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobOptField-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobOptField-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of JobOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobOptField$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobOptField-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JobOptField-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobOptField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
