% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_feature_summary_quality_data.R
\docType{class}
\name{LCMSFeatureSummaryQualityData}
\alias{LCMSFeatureSummaryQualityData}
\title{LCMSFeatureSummaryQualityData}
\format{
An \code{R6Class} generator object
}
\description{
This quality information is available for LC-MS data and includes quality measures for  - the chromatographic peak  - isotope peaks  - detected adducts  - MS/MS peaks  This information is the same as provided in the LCMS-view in the GUI

LCMSFeatureSummaryQualityData Class
}
\details{
Create a new LCMSFeatureSummaryQualityData
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{peakQualityResult}}{\link{LCMSQualityCheckResult} [optional]}

\item{\code{isotopeQualityResult}}{\link{LCMSQualityCheckResult} [optional]}

\item{\code{adductQualityResult}}{\link{LCMSQualityCheckResult} [optional]}

\item{\code{ms2QualityResult}}{\link{LCMSQualityCheckResult} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LCMSFeatureSummaryQualityData-new}{\code{LCMSFeatureSummaryQualityData$new()}}
\item \href{#method-LCMSFeatureSummaryQualityData-toJSON}{\code{LCMSFeatureSummaryQualityData$toJSON()}}
\item \href{#method-LCMSFeatureSummaryQualityData-fromJSON}{\code{LCMSFeatureSummaryQualityData$fromJSON()}}
\item \href{#method-LCMSFeatureSummaryQualityData-toJSONString}{\code{LCMSFeatureSummaryQualityData$toJSONString()}}
\item \href{#method-LCMSFeatureSummaryQualityData-fromJSONString}{\code{LCMSFeatureSummaryQualityData$fromJSONString()}}
\item \href{#method-LCMSFeatureSummaryQualityData-validateJSON}{\code{LCMSFeatureSummaryQualityData$validateJSON()}}
\item \href{#method-LCMSFeatureSummaryQualityData-toString}{\code{LCMSFeatureSummaryQualityData$toString()}}
\item \href{#method-LCMSFeatureSummaryQualityData-isValid}{\code{LCMSFeatureSummaryQualityData$isValid()}}
\item \href{#method-LCMSFeatureSummaryQualityData-getInvalidFields}{\code{LCMSFeatureSummaryQualityData$getInvalidFields()}}
\item \href{#method-LCMSFeatureSummaryQualityData-print}{\code{LCMSFeatureSummaryQualityData$print()}}
\item \href{#method-LCMSFeatureSummaryQualityData-clone}{\code{LCMSFeatureSummaryQualityData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSFeatureSummaryQualityData-new"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSFeatureSummaryQualityData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new LCMSFeatureSummaryQualityData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSFeatureSummaryQualityData$new(
  peakQualityResult = NULL,
  isotopeQualityResult = NULL,
  adductQualityResult = NULL,
  ms2QualityResult = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peakQualityResult}}{peakQualityResult}

\item{\code{isotopeQualityResult}}{isotopeQualityResult}

\item{\code{adductQualityResult}}{adductQualityResult}

\item{\code{ms2QualityResult}}{ms2QualityResult}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSFeatureSummaryQualityData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSFeatureSummaryQualityData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSFeatureSummaryQualityData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
LCMSFeatureSummaryQualityData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSFeatureSummaryQualityData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSFeatureSummaryQualityData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of LCMSFeatureSummaryQualityData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSFeatureSummaryQualityData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of LCMSFeatureSummaryQualityData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSFeatureSummaryQualityData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSFeatureSummaryQualityData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSFeatureSummaryQualityData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
LCMSFeatureSummaryQualityData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSFeatureSummaryQualityData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSFeatureSummaryQualityData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of LCMSFeatureSummaryQualityData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSFeatureSummaryQualityData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of LCMSFeatureSummaryQualityData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSFeatureSummaryQualityData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSFeatureSummaryQualityData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to LCMSFeatureSummaryQualityData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSFeatureSummaryQualityData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSFeatureSummaryQualityData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSFeatureSummaryQualityData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSFeatureSummaryQualityData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of LCMSFeatureSummaryQualityData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSFeatureSummaryQualityData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSFeatureSummaryQualityData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSFeatureSummaryQualityData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSFeatureSummaryQualityData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSFeatureSummaryQualityData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSFeatureSummaryQualityData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSFeatureSummaryQualityData-print"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSFeatureSummaryQualityData-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSFeatureSummaryQualityData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSFeatureSummaryQualityData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSFeatureSummaryQualityData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSFeatureSummaryQualityData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
