% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_quality_check.R
\docType{class}
\name{LCMSQualityCheck}
\alias{LCMSQualityCheck}
\title{LCMSQualityCheck}
\format{
An \code{R6Class} generator object
}
\description{
LCMSQualityCheck Class
}
\details{
Create a new LCMSQualityCheck
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{quality}}{character [optional]}

\item{\code{description}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LCMSQualityCheck-new}{\code{LCMSQualityCheck$new()}}
\item \href{#method-LCMSQualityCheck-toJSON}{\code{LCMSQualityCheck$toJSON()}}
\item \href{#method-LCMSQualityCheck-fromJSON}{\code{LCMSQualityCheck$fromJSON()}}
\item \href{#method-LCMSQualityCheck-toJSONString}{\code{LCMSQualityCheck$toJSONString()}}
\item \href{#method-LCMSQualityCheck-fromJSONString}{\code{LCMSQualityCheck$fromJSONString()}}
\item \href{#method-LCMSQualityCheck-validateJSON}{\code{LCMSQualityCheck$validateJSON()}}
\item \href{#method-LCMSQualityCheck-toString}{\code{LCMSQualityCheck$toString()}}
\item \href{#method-LCMSQualityCheck-isValid}{\code{LCMSQualityCheck$isValid()}}
\item \href{#method-LCMSQualityCheck-getInvalidFields}{\code{LCMSQualityCheck$getInvalidFields()}}
\item \href{#method-LCMSQualityCheck-print}{\code{LCMSQualityCheck$print()}}
\item \href{#method-LCMSQualityCheck-clone}{\code{LCMSQualityCheck$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheck-new"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheck-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new LCMSQualityCheck class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheck$new(quality = NULL, description = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quality}}{quality}

\item{\code{description}}{description}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheck-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheck-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheck$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
LCMSQualityCheck in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheck-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheck-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of LCMSQualityCheck
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheck$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of LCMSQualityCheck
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheck-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheck-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheck$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
LCMSQualityCheck in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheck-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheck-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of LCMSQualityCheck
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheck$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of LCMSQualityCheck
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheck-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheck-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to LCMSQualityCheck and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheck$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheck-toString"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheck-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheck$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of LCMSQualityCheck
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheck-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheck-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheck$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheck-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheck-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheck$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheck-print"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheck-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheck$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheck-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheck-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheck$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
