% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_quality_check_result.R
\docType{class}
\name{LCMSQualityCheckResult}
\alias{LCMSQualityCheckResult}
\title{LCMSQualityCheckResult}
\format{
An \code{R6Class} generator object
}
\description{
LCMSQualityCheckResult Class
}
\details{
Create a new LCMSQualityCheckResult
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{checks}}{list(\link{LCMSQualityCheck}) [optional]}

\item{\code{quality}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LCMSQualityCheckResult-new}{\code{LCMSQualityCheckResult$new()}}
\item \href{#method-LCMSQualityCheckResult-toJSON}{\code{LCMSQualityCheckResult$toJSON()}}
\item \href{#method-LCMSQualityCheckResult-fromJSON}{\code{LCMSQualityCheckResult$fromJSON()}}
\item \href{#method-LCMSQualityCheckResult-toJSONString}{\code{LCMSQualityCheckResult$toJSONString()}}
\item \href{#method-LCMSQualityCheckResult-fromJSONString}{\code{LCMSQualityCheckResult$fromJSONString()}}
\item \href{#method-LCMSQualityCheckResult-validateJSON}{\code{LCMSQualityCheckResult$validateJSON()}}
\item \href{#method-LCMSQualityCheckResult-toString}{\code{LCMSQualityCheckResult$toString()}}
\item \href{#method-LCMSQualityCheckResult-isValid}{\code{LCMSQualityCheckResult$isValid()}}
\item \href{#method-LCMSQualityCheckResult-getInvalidFields}{\code{LCMSQualityCheckResult$getInvalidFields()}}
\item \href{#method-LCMSQualityCheckResult-print}{\code{LCMSQualityCheckResult$print()}}
\item \href{#method-LCMSQualityCheckResult-clone}{\code{LCMSQualityCheckResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheckResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheckResult-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new LCMSQualityCheckResult class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheckResult$new(checks = NULL, quality = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{checks}}{checks}

\item{\code{quality}}{quality}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheckResult-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheckResult-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheckResult$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
LCMSQualityCheckResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheckResult-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheckResult-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of LCMSQualityCheckResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheckResult$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of LCMSQualityCheckResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheckResult-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheckResult-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheckResult$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
LCMSQualityCheckResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheckResult-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheckResult-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of LCMSQualityCheckResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheckResult$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of LCMSQualityCheckResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheckResult-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheckResult-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to LCMSQualityCheckResult and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheckResult$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheckResult-toString"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheckResult-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheckResult$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of LCMSQualityCheckResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheckResult-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheckResult-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheckResult$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheckResult-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheckResult-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheckResult$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheckResult-print"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheckResult-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheckResult$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LCMSQualityCheckResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LCMSQualityCheckResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LCMSQualityCheckResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
