% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_edge.R
\docType{class}
\name{LossEdge}
\alias{LossEdge}
\title{LossEdge}
\format{
An \code{R6Class} generator object
}
\description{
LossEdge Class
}
\details{
Create a new LossEdge
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sourceFragment}}{\link{FragmentNode} [optional]}

\item{\code{targetFragment}}{\link{FragmentNode} [optional]}

\item{\code{molecularFormula}}{character [optional]}

\item{\code{score}}{numeric [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LossEdge-new}{\code{LossEdge$new()}}
\item \href{#method-LossEdge-toJSON}{\code{LossEdge$toJSON()}}
\item \href{#method-LossEdge-fromJSON}{\code{LossEdge$fromJSON()}}
\item \href{#method-LossEdge-toJSONString}{\code{LossEdge$toJSONString()}}
\item \href{#method-LossEdge-fromJSONString}{\code{LossEdge$fromJSONString()}}
\item \href{#method-LossEdge-validateJSON}{\code{LossEdge$validateJSON()}}
\item \href{#method-LossEdge-toString}{\code{LossEdge$toString()}}
\item \href{#method-LossEdge-isValid}{\code{LossEdge$isValid()}}
\item \href{#method-LossEdge-getInvalidFields}{\code{LossEdge$getInvalidFields()}}
\item \href{#method-LossEdge-print}{\code{LossEdge$print()}}
\item \href{#method-LossEdge-clone}{\code{LossEdge$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LossEdge-new"></a>}}
\if{latex}{\out{\hypertarget{method-LossEdge-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new LossEdge class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossEdge$new(
  sourceFragment = NULL,
  targetFragment = NULL,
  molecularFormula = NULL,
  score = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sourceFragment}}{sourceFragment}

\item{\code{targetFragment}}{targetFragment}

\item{\code{molecularFormula}}{molecularFormula}

\item{\code{score}}{score}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LossEdge-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LossEdge-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossEdge$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
LossEdge in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LossEdge-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LossEdge-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of LossEdge
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossEdge$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of LossEdge
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LossEdge-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LossEdge-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossEdge$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
LossEdge in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LossEdge-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LossEdge-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of LossEdge
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossEdge$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of LossEdge
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LossEdge-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LossEdge-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to LossEdge and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossEdge$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LossEdge-toString"></a>}}
\if{latex}{\out{\hypertarget{method-LossEdge-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossEdge$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of LossEdge
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LossEdge-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-LossEdge-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossEdge$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LossEdge-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-LossEdge-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossEdge$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LossEdge-print"></a>}}
\if{latex}{\out{\hypertarget{method-LossEdge-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossEdge$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LossEdge-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LossEdge-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LossEdge$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
