% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npc_class_statistics_api.R
\docType{class}
\name{NPCClassStatisticsApi}
\alias{NPCClassStatisticsApi}
\title{NPCClassStatistics operations}
\format{
An \code{R6Class} generator object
}
\description{
NPCClassStatisticsApi
}
\details{
SIRIUS Nightsky API

REST API that provides the full functionality of SIRIUS and its web services as background service. It is intended as entry-point for scripting languages and software integration SDKs.This API is exposed by SIRIUS 6

The version of the OpenAPI document: 3.1
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  ComputeNpcClassFoldChangesExperimental  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project-space to compute the fold change in.
var_fold_change_job_submission <- FoldChangeJobSubmission$new("leftRunGroup_example", "rightRunGroup_example", c("AVG"), c("APEX_INTENSITY")) # FoldChangeJobSubmission | Parameters of fold change job
var_opt_fields <- c("none") # array[character] | job opt fields. (Optional)

#[EXPERIMENTAL] Compute the fold change between two groups of runs
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$ComputeNpcClassFoldChangesExperimental(var_project_id, var_fold_change_job_submission, opt_fields = var_opt_fieldsdata_file = "result.txt")
result <- api_instance$npc_class_statistics_api$ComputeNpcClassFoldChangesExperimental(var_project_id, var_fold_change_job_submission, opt_fields = var_opt_fields)
dput(result)


####################  DeleteNpcClassFoldChangesExperimental  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project-space to delete from.
var_left_group_name <- "left_group_name_example" # character | name of the left group.
var_right_group_name <- "right_group_name_example" # character | name of the right group.
var_aggregation <- "AVG" # character |  (Optional)
var_quantification <- "APEX_INTENSITY" # character |  (Optional)

#[EXPERIMENTAL] Delete fold changes
api_instance <- rsirius_api$new()

api_instance$npc_class_statistics_api$DeleteNpcClassFoldChangesExperimental(var_project_id, var_left_group_name, var_right_group_name, aggregation = var_aggregation, quantification = var_quantification)


####################  GetNpcClassFoldChangeTableExperimental  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project-space to read from.
var_aggregation <- "AVG" # character | aggregation type. (Optional)
var_quantification <- "APEX_INTENSITY" # character | quantification type. (Optional)

#[EXPERIMENTAL] Get table of all fold changes in the project space
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetNpcClassFoldChangeTableExperimental(var_project_id, aggregation = var_aggregation, quantification = var_quantificationdata_file = "result.txt")
result <- api_instance$npc_class_statistics_api$GetNpcClassFoldChangeTableExperimental(var_project_id, aggregation = var_aggregation, quantification = var_quantification)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NPCClassStatisticsApi-new}{\code{NPCClassStatisticsApi$new()}}
\item \href{#method-NPCClassStatisticsApi-ComputeNpcClassFoldChangesExperimental}{\code{NPCClassStatisticsApi$ComputeNpcClassFoldChangesExperimental()}}
\item \href{#method-NPCClassStatisticsApi-ComputeNpcClassFoldChangesExperimentalWithHttpInfo}{\code{NPCClassStatisticsApi$ComputeNpcClassFoldChangesExperimentalWithHttpInfo()}}
\item \href{#method-NPCClassStatisticsApi-DeleteNpcClassFoldChangesExperimental}{\code{NPCClassStatisticsApi$DeleteNpcClassFoldChangesExperimental()}}
\item \href{#method-NPCClassStatisticsApi-DeleteNpcClassFoldChangesExperimentalWithHttpInfo}{\code{NPCClassStatisticsApi$DeleteNpcClassFoldChangesExperimentalWithHttpInfo()}}
\item \href{#method-NPCClassStatisticsApi-GetNpcClassFoldChangeTableExperimental}{\code{NPCClassStatisticsApi$GetNpcClassFoldChangeTableExperimental()}}
\item \href{#method-NPCClassStatisticsApi-GetNpcClassFoldChangeTableExperimentalWithHttpInfo}{\code{NPCClassStatisticsApi$GetNpcClassFoldChangeTableExperimentalWithHttpInfo()}}
\item \href{#method-NPCClassStatisticsApi-clone}{\code{NPCClassStatisticsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NPCClassStatisticsApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-NPCClassStatisticsApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new NPCClassStatisticsApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NPCClassStatisticsApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NPCClassStatisticsApi-ComputeNpcClassFoldChangesExperimental"></a>}}
\if{latex}{\out{\hypertarget{method-NPCClassStatisticsApi-ComputeNpcClassFoldChangesExperimental}{}}}
\subsection{Method \code{ComputeNpcClassFoldChangesExperimental()}}{
[EXPERIMENTAL] Compute the fold change between two groups of runs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NPCClassStatisticsApi$ComputeNpcClassFoldChangesExperimental(
  project_id,
  fold_change_job_submission,
  opt_fields = list("progress"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to compute the fold change in.}

\item{\code{fold_change_job_submission}}{Parameters of fold change job}

\item{\code{opt_fields}}{(optional) job opt fields. (default value: ["progress"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Job
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NPCClassStatisticsApi-ComputeNpcClassFoldChangesExperimentalWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-NPCClassStatisticsApi-ComputeNpcClassFoldChangesExperimentalWithHttpInfo}{}}}
\subsection{Method \code{ComputeNpcClassFoldChangesExperimentalWithHttpInfo()}}{
[EXPERIMENTAL] Compute the fold change between two groups of runs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NPCClassStatisticsApi$ComputeNpcClassFoldChangesExperimentalWithHttpInfo(
  project_id,
  fold_change_job_submission,
  opt_fields = list("progress"),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to compute the fold change in.}

\item{\code{fold_change_job_submission}}{Parameters of fold change job}

\item{\code{opt_fields}}{(optional) job opt fields. (default value: ["progress"])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (Job) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NPCClassStatisticsApi-DeleteNpcClassFoldChangesExperimental"></a>}}
\if{latex}{\out{\hypertarget{method-NPCClassStatisticsApi-DeleteNpcClassFoldChangesExperimental}{}}}
\subsection{Method \code{DeleteNpcClassFoldChangesExperimental()}}{
[EXPERIMENTAL] Delete fold changes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NPCClassStatisticsApi$DeleteNpcClassFoldChangesExperimental(
  project_id,
  left_group_name,
  right_group_name,
  aggregation = "AVG",
  quantification = "APEX_INTENSITY",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to delete from.}

\item{\code{left_group_name}}{name of the left group.}

\item{\code{right_group_name}}{name of the right group.}

\item{\code{aggregation}}{(optional) No description (default value: "AVG")}

\item{\code{quantification}}{(optional) No description (default value: "APEX_INTENSITY")}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NPCClassStatisticsApi-DeleteNpcClassFoldChangesExperimentalWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-NPCClassStatisticsApi-DeleteNpcClassFoldChangesExperimentalWithHttpInfo}{}}}
\subsection{Method \code{DeleteNpcClassFoldChangesExperimentalWithHttpInfo()}}{
[EXPERIMENTAL] Delete fold changes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NPCClassStatisticsApi$DeleteNpcClassFoldChangesExperimentalWithHttpInfo(
  project_id,
  left_group_name,
  right_group_name,
  aggregation = "AVG",
  quantification = "APEX_INTENSITY",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to delete from.}

\item{\code{left_group_name}}{name of the left group.}

\item{\code{right_group_name}}{name of the right group.}

\item{\code{aggregation}}{(optional) No description (default value: "AVG")}

\item{\code{quantification}}{(optional) No description (default value: "APEX_INTENSITY")}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NPCClassStatisticsApi-GetNpcClassFoldChangeTableExperimental"></a>}}
\if{latex}{\out{\hypertarget{method-NPCClassStatisticsApi-GetNpcClassFoldChangeTableExperimental}{}}}
\subsection{Method \code{GetNpcClassFoldChangeTableExperimental()}}{
[EXPERIMENTAL] Get table of all fold changes in the project space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NPCClassStatisticsApi$GetNpcClassFoldChangeTableExperimental(
  project_id,
  aggregation = "AVG",
  quantification = "APEX_INTENSITY",
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{aggregation}}{(optional) aggregation type. (default value: "AVG")}

\item{\code{quantification}}{(optional) quantification type. (default value: "APEX_INTENSITY")}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
StatisticsTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NPCClassStatisticsApi-GetNpcClassFoldChangeTableExperimentalWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-NPCClassStatisticsApi-GetNpcClassFoldChangeTableExperimentalWithHttpInfo}{}}}
\subsection{Method \code{GetNpcClassFoldChangeTableExperimentalWithHttpInfo()}}{
[EXPERIMENTAL] Get table of all fold changes in the project space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NPCClassStatisticsApi$GetNpcClassFoldChangeTableExperimentalWithHttpInfo(
  project_id,
  aggregation = "AVG",
  quantification = "APEX_INTENSITY",
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{aggregation}}{(optional) aggregation type. (default value: "AVG")}

\item{\code{quantification}}{(optional) quantification type. (default value: "APEX_INTENSITY")}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (StatisticsTable) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NPCClassStatisticsApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NPCClassStatisticsApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NPCClassStatisticsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
