% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_compound.R
\docType{class}
\name{PageCompound}
\alias{PageCompound}
\title{PageCompound}
\format{
An \code{R6Class} generator object
}
\description{
PageCompound Class

PageCompound Class
}
\details{
Create a new PageCompound
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalPages}}{integer [optional]}

\item{\code{totalElements}}{integer [optional]}

\item{\code{first}}{character [optional]}

\item{\code{last}}{character [optional]}

\item{\code{size}}{integer [optional]}

\item{\code{content}}{list(\link{Compound}) [optional]}

\item{\code{number}}{integer [optional]}

\item{\code{sort}}{\link{SortObject} [optional]}

\item{\code{numberOfElements}}{integer [optional]}

\item{\code{pageable}}{\link{PageableObject} [optional]}

\item{\code{empty}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PageCompound-new}{\code{PageCompound$new()}}
\item \href{#method-PageCompound-toJSON}{\code{PageCompound$toJSON()}}
\item \href{#method-PageCompound-fromJSON}{\code{PageCompound$fromJSON()}}
\item \href{#method-PageCompound-toJSONString}{\code{PageCompound$toJSONString()}}
\item \href{#method-PageCompound-fromJSONString}{\code{PageCompound$fromJSONString()}}
\item \href{#method-PageCompound-validateJSON}{\code{PageCompound$validateJSON()}}
\item \href{#method-PageCompound-toString}{\code{PageCompound$toString()}}
\item \href{#method-PageCompound-isValid}{\code{PageCompound$isValid()}}
\item \href{#method-PageCompound-getInvalidFields}{\code{PageCompound$getInvalidFields()}}
\item \href{#method-PageCompound-print}{\code{PageCompound$print()}}
\item \href{#method-PageCompound-clone}{\code{PageCompound$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageCompound-new"></a>}}
\if{latex}{\out{\hypertarget{method-PageCompound-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PageCompound class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageCompound$new(
  totalPages = NULL,
  totalElements = NULL,
  first = NULL,
  last = NULL,
  size = NULL,
  content = NULL,
  number = NULL,
  sort = NULL,
  numberOfElements = NULL,
  pageable = NULL,
  empty = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalPages}}{totalPages}

\item{\code{totalElements}}{totalElements}

\item{\code{first}}{first}

\item{\code{last}}{last}

\item{\code{size}}{size}

\item{\code{content}}{content}

\item{\code{number}}{number}

\item{\code{sort}}{sort}

\item{\code{numberOfElements}}{numberOfElements}

\item{\code{pageable}}{pageable}

\item{\code{empty}}{empty}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageCompound-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageCompound-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageCompound$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PageCompound in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageCompound-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageCompound-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PageCompound
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageCompound$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PageCompound
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageCompound-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PageCompound-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageCompound$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PageCompound in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageCompound-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PageCompound-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PageCompound
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageCompound$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PageCompound
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageCompound-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageCompound-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PageCompound and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageCompound$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageCompound-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PageCompound-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageCompound$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PageCompound
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageCompound-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PageCompound-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageCompound$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageCompound-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PageCompound-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageCompound$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageCompound-print"></a>}}
\if{latex}{\out{\hypertarget{method-PageCompound-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageCompound$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageCompound-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PageCompound-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageCompound$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
