% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_model_aligned_feature.R
\docType{class}
\name{PagedModelAlignedFeature}
\alias{PagedModelAlignedFeature}
\title{PagedModelAlignedFeature}
\format{
An \code{R6Class} generator object
}
\description{
PagedModelAlignedFeature Class

PagedModelAlignedFeature Class
}
\details{
Create a new PagedModelAlignedFeature
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content}}{list(\link{AlignedFeature}) [optional]}

\item{\code{page}}{\link{PageMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PagedModelAlignedFeature-new}{\code{PagedModelAlignedFeature$new()}}
\item \href{#method-PagedModelAlignedFeature-toJSON}{\code{PagedModelAlignedFeature$toJSON()}}
\item \href{#method-PagedModelAlignedFeature-fromJSON}{\code{PagedModelAlignedFeature$fromJSON()}}
\item \href{#method-PagedModelAlignedFeature-toJSONString}{\code{PagedModelAlignedFeature$toJSONString()}}
\item \href{#method-PagedModelAlignedFeature-fromJSONString}{\code{PagedModelAlignedFeature$fromJSONString()}}
\item \href{#method-PagedModelAlignedFeature-validateJSON}{\code{PagedModelAlignedFeature$validateJSON()}}
\item \href{#method-PagedModelAlignedFeature-toString}{\code{PagedModelAlignedFeature$toString()}}
\item \href{#method-PagedModelAlignedFeature-isValid}{\code{PagedModelAlignedFeature$isValid()}}
\item \href{#method-PagedModelAlignedFeature-getInvalidFields}{\code{PagedModelAlignedFeature$getInvalidFields()}}
\item \href{#method-PagedModelAlignedFeature-print}{\code{PagedModelAlignedFeature$print()}}
\item \href{#method-PagedModelAlignedFeature-clone}{\code{PagedModelAlignedFeature$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelAlignedFeature-new"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelAlignedFeature-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PagedModelAlignedFeature class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelAlignedFeature$new(content = NULL, page = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{content}

\item{\code{page}}{page}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelAlignedFeature-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelAlignedFeature-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelAlignedFeature$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PagedModelAlignedFeature in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelAlignedFeature-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelAlignedFeature-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PagedModelAlignedFeature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelAlignedFeature$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PagedModelAlignedFeature
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelAlignedFeature-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelAlignedFeature-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelAlignedFeature$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PagedModelAlignedFeature in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelAlignedFeature-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelAlignedFeature-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PagedModelAlignedFeature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelAlignedFeature$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PagedModelAlignedFeature
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelAlignedFeature-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelAlignedFeature-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PagedModelAlignedFeature and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelAlignedFeature$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelAlignedFeature-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelAlignedFeature-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelAlignedFeature$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PagedModelAlignedFeature
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelAlignedFeature-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelAlignedFeature-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelAlignedFeature$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelAlignedFeature-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelAlignedFeature-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelAlignedFeature$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelAlignedFeature-print"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelAlignedFeature-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelAlignedFeature$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelAlignedFeature-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelAlignedFeature-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelAlignedFeature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
