% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_model_feature.R
\docType{class}
\name{PagedModelFeature}
\alias{PagedModelFeature}
\title{PagedModelFeature}
\format{
An \code{R6Class} generator object
}
\description{
PagedModelFeature Class

PagedModelFeature Class
}
\details{
Create a new PagedModelFeature
}
\examples{

## ------------------------------------------------
## Method `PagedModelFeature$toList`
## ------------------------------------------------

# convert array of PagedModelFeature (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content}}{list(\link{Feature}) [optional]}

\item{\code{page}}{\link{PageMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PagedModelFeature-new}{\code{PagedModelFeature$new()}}
\item \href{#method-PagedModelFeature-toJSON}{\code{PagedModelFeature$toJSON()}}
\item \href{#method-PagedModelFeature-toList}{\code{PagedModelFeature$toList()}}
\item \href{#method-PagedModelFeature-toSimpleType}{\code{PagedModelFeature$toSimpleType()}}
\item \href{#method-PagedModelFeature-fromJSON}{\code{PagedModelFeature$fromJSON()}}
\item \href{#method-PagedModelFeature-toJSONString}{\code{PagedModelFeature$toJSONString()}}
\item \href{#method-PagedModelFeature-fromJSONString}{\code{PagedModelFeature$fromJSONString()}}
\item \href{#method-PagedModelFeature-validateJSON}{\code{PagedModelFeature$validateJSON()}}
\item \href{#method-PagedModelFeature-toString}{\code{PagedModelFeature$toString()}}
\item \href{#method-PagedModelFeature-isValid}{\code{PagedModelFeature$isValid()}}
\item \href{#method-PagedModelFeature-getInvalidFields}{\code{PagedModelFeature$getInvalidFields()}}
\item \href{#method-PagedModelFeature-print}{\code{PagedModelFeature$print()}}
\item \href{#method-PagedModelFeature-clone}{\code{PagedModelFeature$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelFeature-new"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelFeature-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PagedModelFeature class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelFeature$new(content = NULL, page = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{content}

\item{\code{page}}{page}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelFeature-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelFeature-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelFeature$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelFeature-toList"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelFeature-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelFeature$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PagedModelFeature as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of PagedModelFeature (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelFeature-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelFeature-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert PagedModelFeature to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelFeature$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelFeature-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelFeature-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PagedModelFeature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelFeature$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PagedModelFeature
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelFeature-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelFeature-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelFeature$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PagedModelFeature in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelFeature-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelFeature-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PagedModelFeature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelFeature$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PagedModelFeature
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelFeature-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelFeature-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PagedModelFeature and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelFeature$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelFeature-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelFeature-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelFeature$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PagedModelFeature
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelFeature-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelFeature-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelFeature$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelFeature-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelFeature-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelFeature$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelFeature-print"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelFeature-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelFeature$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelFeature-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelFeature-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelFeature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
