% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_model_job.R
\docType{class}
\name{PagedModelJob}
\alias{PagedModelJob}
\title{PagedModelJob}
\format{
An \code{R6Class} generator object
}
\description{
PagedModelJob Class

PagedModelJob Class
}
\details{
Create a new PagedModelJob
}
\examples{

## ------------------------------------------------
## Method `PagedModelJob$toList`
## ------------------------------------------------

# convert array of PagedModelJob (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content}}{list(\link{Job}) [optional]}

\item{\code{page}}{\link{PageMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PagedModelJob-new}{\code{PagedModelJob$new()}}
\item \href{#method-PagedModelJob-toJSON}{\code{PagedModelJob$toJSON()}}
\item \href{#method-PagedModelJob-toList}{\code{PagedModelJob$toList()}}
\item \href{#method-PagedModelJob-toSimpleType}{\code{PagedModelJob$toSimpleType()}}
\item \href{#method-PagedModelJob-fromJSON}{\code{PagedModelJob$fromJSON()}}
\item \href{#method-PagedModelJob-toJSONString}{\code{PagedModelJob$toJSONString()}}
\item \href{#method-PagedModelJob-fromJSONString}{\code{PagedModelJob$fromJSONString()}}
\item \href{#method-PagedModelJob-validateJSON}{\code{PagedModelJob$validateJSON()}}
\item \href{#method-PagedModelJob-toString}{\code{PagedModelJob$toString()}}
\item \href{#method-PagedModelJob-isValid}{\code{PagedModelJob$isValid()}}
\item \href{#method-PagedModelJob-getInvalidFields}{\code{PagedModelJob$getInvalidFields()}}
\item \href{#method-PagedModelJob-print}{\code{PagedModelJob$print()}}
\item \href{#method-PagedModelJob-clone}{\code{PagedModelJob$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelJob-new"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelJob-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PagedModelJob class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelJob$new(content = NULL, page = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{content}

\item{\code{page}}{page}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelJob-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelJob-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelJob$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelJob-toList"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelJob-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelJob$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PagedModelJob as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of PagedModelJob (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelJob-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelJob-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert PagedModelJob to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelJob$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelJob-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelJob-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PagedModelJob
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelJob$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PagedModelJob
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelJob-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelJob-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelJob$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PagedModelJob in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelJob-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelJob-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PagedModelJob
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelJob$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PagedModelJob
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelJob-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelJob-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PagedModelJob and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelJob$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelJob-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelJob-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelJob$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PagedModelJob
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelJob-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelJob-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelJob$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelJob-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelJob-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelJob$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelJob-print"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelJob-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelJob$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelJob-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelJob-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelJob$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
