% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_model_spectral_library_match.R
\docType{class}
\name{PagedModelSpectralLibraryMatch}
\alias{PagedModelSpectralLibraryMatch}
\title{PagedModelSpectralLibraryMatch}
\format{
An \code{R6Class} generator object
}
\description{
PagedModelSpectralLibraryMatch Class

PagedModelSpectralLibraryMatch Class
}
\details{
Create a new PagedModelSpectralLibraryMatch
}
\examples{

## ------------------------------------------------
## Method `PagedModelSpectralLibraryMatch$toList`
## ------------------------------------------------

# convert array of PagedModelSpectralLibraryMatch (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content}}{list(\link{SpectralLibraryMatch}) [optional]}

\item{\code{page}}{\link{PageMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PagedModelSpectralLibraryMatch-new}{\code{PagedModelSpectralLibraryMatch$new()}}
\item \href{#method-PagedModelSpectralLibraryMatch-toJSON}{\code{PagedModelSpectralLibraryMatch$toJSON()}}
\item \href{#method-PagedModelSpectralLibraryMatch-toList}{\code{PagedModelSpectralLibraryMatch$toList()}}
\item \href{#method-PagedModelSpectralLibraryMatch-toSimpleType}{\code{PagedModelSpectralLibraryMatch$toSimpleType()}}
\item \href{#method-PagedModelSpectralLibraryMatch-fromJSON}{\code{PagedModelSpectralLibraryMatch$fromJSON()}}
\item \href{#method-PagedModelSpectralLibraryMatch-toJSONString}{\code{PagedModelSpectralLibraryMatch$toJSONString()}}
\item \href{#method-PagedModelSpectralLibraryMatch-fromJSONString}{\code{PagedModelSpectralLibraryMatch$fromJSONString()}}
\item \href{#method-PagedModelSpectralLibraryMatch-validateJSON}{\code{PagedModelSpectralLibraryMatch$validateJSON()}}
\item \href{#method-PagedModelSpectralLibraryMatch-toString}{\code{PagedModelSpectralLibraryMatch$toString()}}
\item \href{#method-PagedModelSpectralLibraryMatch-isValid}{\code{PagedModelSpectralLibraryMatch$isValid()}}
\item \href{#method-PagedModelSpectralLibraryMatch-getInvalidFields}{\code{PagedModelSpectralLibraryMatch$getInvalidFields()}}
\item \href{#method-PagedModelSpectralLibraryMatch-print}{\code{PagedModelSpectralLibraryMatch$print()}}
\item \href{#method-PagedModelSpectralLibraryMatch-clone}{\code{PagedModelSpectralLibraryMatch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelSpectralLibraryMatch-new"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelSpectralLibraryMatch-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PagedModelSpectralLibraryMatch class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelSpectralLibraryMatch$new(content = NULL, page = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{content}

\item{\code{page}}{page}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelSpectralLibraryMatch-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelSpectralLibraryMatch-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelSpectralLibraryMatch$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelSpectralLibraryMatch-toList"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelSpectralLibraryMatch-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelSpectralLibraryMatch$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PagedModelSpectralLibraryMatch as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of PagedModelSpectralLibraryMatch (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelSpectralLibraryMatch-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelSpectralLibraryMatch-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert PagedModelSpectralLibraryMatch to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelSpectralLibraryMatch$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelSpectralLibraryMatch-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelSpectralLibraryMatch-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PagedModelSpectralLibraryMatch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelSpectralLibraryMatch$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PagedModelSpectralLibraryMatch
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelSpectralLibraryMatch-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelSpectralLibraryMatch-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelSpectralLibraryMatch$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PagedModelSpectralLibraryMatch in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelSpectralLibraryMatch-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelSpectralLibraryMatch-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PagedModelSpectralLibraryMatch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelSpectralLibraryMatch$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PagedModelSpectralLibraryMatch
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelSpectralLibraryMatch-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelSpectralLibraryMatch-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PagedModelSpectralLibraryMatch and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelSpectralLibraryMatch$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelSpectralLibraryMatch-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelSpectralLibraryMatch-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelSpectralLibraryMatch$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PagedModelSpectralLibraryMatch
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelSpectralLibraryMatch-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelSpectralLibraryMatch-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelSpectralLibraryMatch$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelSpectralLibraryMatch-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelSpectralLibraryMatch-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelSpectralLibraryMatch$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelSpectralLibraryMatch-print"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelSpectralLibraryMatch-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelSpectralLibraryMatch$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelSpectralLibraryMatch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelSpectralLibraryMatch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelSpectralLibraryMatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
