% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_model_structure_candidate_scored.R
\docType{class}
\name{PagedModelStructureCandidateScored}
\alias{PagedModelStructureCandidateScored}
\title{PagedModelStructureCandidateScored}
\format{
An \code{R6Class} generator object
}
\description{
PagedModelStructureCandidateScored Class

PagedModelStructureCandidateScored Class
}
\details{
Create a new PagedModelStructureCandidateScored
}
\examples{

## ------------------------------------------------
## Method `PagedModelStructureCandidateScored$toList`
## ------------------------------------------------

# convert array of PagedModelStructureCandidateScored (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content}}{list(\link{StructureCandidateScored}) [optional]}

\item{\code{page}}{\link{PageMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PagedModelStructureCandidateScored-new}{\code{PagedModelStructureCandidateScored$new()}}
\item \href{#method-PagedModelStructureCandidateScored-toJSON}{\code{PagedModelStructureCandidateScored$toJSON()}}
\item \href{#method-PagedModelStructureCandidateScored-toList}{\code{PagedModelStructureCandidateScored$toList()}}
\item \href{#method-PagedModelStructureCandidateScored-toSimpleType}{\code{PagedModelStructureCandidateScored$toSimpleType()}}
\item \href{#method-PagedModelStructureCandidateScored-fromJSON}{\code{PagedModelStructureCandidateScored$fromJSON()}}
\item \href{#method-PagedModelStructureCandidateScored-toJSONString}{\code{PagedModelStructureCandidateScored$toJSONString()}}
\item \href{#method-PagedModelStructureCandidateScored-fromJSONString}{\code{PagedModelStructureCandidateScored$fromJSONString()}}
\item \href{#method-PagedModelStructureCandidateScored-validateJSON}{\code{PagedModelStructureCandidateScored$validateJSON()}}
\item \href{#method-PagedModelStructureCandidateScored-toString}{\code{PagedModelStructureCandidateScored$toString()}}
\item \href{#method-PagedModelStructureCandidateScored-isValid}{\code{PagedModelStructureCandidateScored$isValid()}}
\item \href{#method-PagedModelStructureCandidateScored-getInvalidFields}{\code{PagedModelStructureCandidateScored$getInvalidFields()}}
\item \href{#method-PagedModelStructureCandidateScored-print}{\code{PagedModelStructureCandidateScored$print()}}
\item \href{#method-PagedModelStructureCandidateScored-clone}{\code{PagedModelStructureCandidateScored$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelStructureCandidateScored-new"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelStructureCandidateScored-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PagedModelStructureCandidateScored class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelStructureCandidateScored$new(content = NULL, page = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{content}

\item{\code{page}}{page}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelStructureCandidateScored-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelStructureCandidateScored-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelStructureCandidateScored$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelStructureCandidateScored-toList"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelStructureCandidateScored-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelStructureCandidateScored$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PagedModelStructureCandidateScored as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of PagedModelStructureCandidateScored (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelStructureCandidateScored-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelStructureCandidateScored-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert PagedModelStructureCandidateScored to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelStructureCandidateScored$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelStructureCandidateScored-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelStructureCandidateScored-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PagedModelStructureCandidateScored
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelStructureCandidateScored$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PagedModelStructureCandidateScored
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelStructureCandidateScored-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelStructureCandidateScored-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelStructureCandidateScored$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PagedModelStructureCandidateScored in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelStructureCandidateScored-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelStructureCandidateScored-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PagedModelStructureCandidateScored
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelStructureCandidateScored$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PagedModelStructureCandidateScored
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelStructureCandidateScored-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelStructureCandidateScored-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PagedModelStructureCandidateScored and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelStructureCandidateScored$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelStructureCandidateScored-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelStructureCandidateScored-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelStructureCandidateScored$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PagedModelStructureCandidateScored
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelStructureCandidateScored-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelStructureCandidateScored-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelStructureCandidateScored$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelStructureCandidateScored-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelStructureCandidateScored-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelStructureCandidateScored$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelStructureCandidateScored-print"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelStructureCandidateScored-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelStructureCandidateScored$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PagedModelStructureCandidateScored-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PagedModelStructureCandidateScored-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PagedModelStructureCandidateScored$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
