% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_formula_candidate.R
\docType{class}
\name{PageFormulaCandidate}
\alias{PageFormulaCandidate}
\title{PageFormulaCandidate}
\format{
An \code{R6Class} generator object
}
\description{
PageFormulaCandidate Class

PageFormulaCandidate Class
}
\details{
Create a new PageFormulaCandidate
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalPages}}{integer [optional]}

\item{\code{totalElements}}{integer [optional]}

\item{\code{first}}{character [optional]}

\item{\code{last}}{character [optional]}

\item{\code{size}}{integer [optional]}

\item{\code{content}}{list(\link{FormulaCandidate}) [optional]}

\item{\code{number}}{integer [optional]}

\item{\code{sort}}{\link{SortObject} [optional]}

\item{\code{numberOfElements}}{integer [optional]}

\item{\code{pageable}}{\link{PageableObject} [optional]}

\item{\code{empty}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PageFormulaCandidate-new}{\code{PageFormulaCandidate$new()}}
\item \href{#method-PageFormulaCandidate-toJSON}{\code{PageFormulaCandidate$toJSON()}}
\item \href{#method-PageFormulaCandidate-fromJSON}{\code{PageFormulaCandidate$fromJSON()}}
\item \href{#method-PageFormulaCandidate-toJSONString}{\code{PageFormulaCandidate$toJSONString()}}
\item \href{#method-PageFormulaCandidate-fromJSONString}{\code{PageFormulaCandidate$fromJSONString()}}
\item \href{#method-PageFormulaCandidate-validateJSON}{\code{PageFormulaCandidate$validateJSON()}}
\item \href{#method-PageFormulaCandidate-toString}{\code{PageFormulaCandidate$toString()}}
\item \href{#method-PageFormulaCandidate-isValid}{\code{PageFormulaCandidate$isValid()}}
\item \href{#method-PageFormulaCandidate-getInvalidFields}{\code{PageFormulaCandidate$getInvalidFields()}}
\item \href{#method-PageFormulaCandidate-print}{\code{PageFormulaCandidate$print()}}
\item \href{#method-PageFormulaCandidate-clone}{\code{PageFormulaCandidate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageFormulaCandidate-new"></a>}}
\if{latex}{\out{\hypertarget{method-PageFormulaCandidate-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PageFormulaCandidate class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageFormulaCandidate$new(
  totalPages = NULL,
  totalElements = NULL,
  first = NULL,
  last = NULL,
  size = NULL,
  content = NULL,
  number = NULL,
  sort = NULL,
  numberOfElements = NULL,
  pageable = NULL,
  empty = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalPages}}{totalPages}

\item{\code{totalElements}}{totalElements}

\item{\code{first}}{first}

\item{\code{last}}{last}

\item{\code{size}}{size}

\item{\code{content}}{content}

\item{\code{number}}{number}

\item{\code{sort}}{sort}

\item{\code{numberOfElements}}{numberOfElements}

\item{\code{pageable}}{pageable}

\item{\code{empty}}{empty}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageFormulaCandidate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageFormulaCandidate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageFormulaCandidate$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PageFormulaCandidate in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageFormulaCandidate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageFormulaCandidate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PageFormulaCandidate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageFormulaCandidate$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PageFormulaCandidate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageFormulaCandidate-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PageFormulaCandidate-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageFormulaCandidate$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PageFormulaCandidate in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageFormulaCandidate-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PageFormulaCandidate-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PageFormulaCandidate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageFormulaCandidate$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PageFormulaCandidate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageFormulaCandidate-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageFormulaCandidate-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PageFormulaCandidate and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageFormulaCandidate$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageFormulaCandidate-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PageFormulaCandidate-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageFormulaCandidate$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PageFormulaCandidate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageFormulaCandidate-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PageFormulaCandidate-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageFormulaCandidate$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageFormulaCandidate-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PageFormulaCandidate-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageFormulaCandidate$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageFormulaCandidate-print"></a>}}
\if{latex}{\out{\hypertarget{method-PageFormulaCandidate-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageFormulaCandidate$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageFormulaCandidate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PageFormulaCandidate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageFormulaCandidate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
