% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_job.R
\docType{class}
\name{PageJob}
\alias{PageJob}
\title{PageJob}
\format{
An \code{R6Class} generator object
}
\description{
PageJob Class

PageJob Class
}
\details{
Create a new PageJob
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalPages}}{integer [optional]}

\item{\code{totalElements}}{integer [optional]}

\item{\code{first}}{character [optional]}

\item{\code{last}}{character [optional]}

\item{\code{size}}{integer [optional]}

\item{\code{content}}{list(\link{Job}) [optional]}

\item{\code{number}}{integer [optional]}

\item{\code{sort}}{\link{SortObject} [optional]}

\item{\code{numberOfElements}}{integer [optional]}

\item{\code{pageable}}{\link{PageableObject} [optional]}

\item{\code{empty}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PageJob-new}{\code{PageJob$new()}}
\item \href{#method-PageJob-toJSON}{\code{PageJob$toJSON()}}
\item \href{#method-PageJob-fromJSON}{\code{PageJob$fromJSON()}}
\item \href{#method-PageJob-toJSONString}{\code{PageJob$toJSONString()}}
\item \href{#method-PageJob-fromJSONString}{\code{PageJob$fromJSONString()}}
\item \href{#method-PageJob-validateJSON}{\code{PageJob$validateJSON()}}
\item \href{#method-PageJob-toString}{\code{PageJob$toString()}}
\item \href{#method-PageJob-isValid}{\code{PageJob$isValid()}}
\item \href{#method-PageJob-getInvalidFields}{\code{PageJob$getInvalidFields()}}
\item \href{#method-PageJob-print}{\code{PageJob$print()}}
\item \href{#method-PageJob-clone}{\code{PageJob$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageJob-new"></a>}}
\if{latex}{\out{\hypertarget{method-PageJob-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PageJob class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageJob$new(
  totalPages = NULL,
  totalElements = NULL,
  first = NULL,
  last = NULL,
  size = NULL,
  content = NULL,
  number = NULL,
  sort = NULL,
  numberOfElements = NULL,
  pageable = NULL,
  empty = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalPages}}{totalPages}

\item{\code{totalElements}}{totalElements}

\item{\code{first}}{first}

\item{\code{last}}{last}

\item{\code{size}}{size}

\item{\code{content}}{content}

\item{\code{number}}{number}

\item{\code{sort}}{sort}

\item{\code{numberOfElements}}{numberOfElements}

\item{\code{pageable}}{pageable}

\item{\code{empty}}{empty}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageJob-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageJob-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageJob$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PageJob in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageJob-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageJob-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PageJob
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageJob$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PageJob
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageJob-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PageJob-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageJob$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PageJob in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageJob-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PageJob-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PageJob
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageJob$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PageJob
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageJob-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageJob-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PageJob and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageJob$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageJob-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PageJob-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageJob$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PageJob
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageJob-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PageJob-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageJob$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageJob-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PageJob-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageJob$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageJob-print"></a>}}
\if{latex}{\out{\hypertarget{method-PageJob-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageJob$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageJob-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PageJob-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageJob$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
