% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_metadata.R
\docType{class}
\name{PageMetadata}
\alias{PageMetadata}
\title{PageMetadata}
\format{
An \code{R6Class} generator object
}
\description{
PageMetadata Class

PageMetadata Class
}
\details{
Create a new PageMetadata
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{size}}{integer [optional]}

\item{\code{number}}{integer [optional]}

\item{\code{totalElements}}{integer [optional]}

\item{\code{totalPages}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PageMetadata-new}{\code{PageMetadata$new()}}
\item \href{#method-PageMetadata-toJSON}{\code{PageMetadata$toJSON()}}
\item \href{#method-PageMetadata-fromJSON}{\code{PageMetadata$fromJSON()}}
\item \href{#method-PageMetadata-toJSONString}{\code{PageMetadata$toJSONString()}}
\item \href{#method-PageMetadata-fromJSONString}{\code{PageMetadata$fromJSONString()}}
\item \href{#method-PageMetadata-validateJSON}{\code{PageMetadata$validateJSON()}}
\item \href{#method-PageMetadata-toString}{\code{PageMetadata$toString()}}
\item \href{#method-PageMetadata-isValid}{\code{PageMetadata$isValid()}}
\item \href{#method-PageMetadata-getInvalidFields}{\code{PageMetadata$getInvalidFields()}}
\item \href{#method-PageMetadata-print}{\code{PageMetadata$print()}}
\item \href{#method-PageMetadata-clone}{\code{PageMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageMetadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-PageMetadata-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PageMetadata class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageMetadata$new(
  size = NULL,
  number = NULL,
  totalElements = NULL,
  totalPages = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{size}

\item{\code{number}}{number}

\item{\code{totalElements}}{totalElements}

\item{\code{totalPages}}{totalPages}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageMetadata-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageMetadata-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageMetadata$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PageMetadata in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageMetadata-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageMetadata-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PageMetadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageMetadata$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PageMetadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageMetadata-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PageMetadata-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageMetadata$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PageMetadata in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageMetadata-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PageMetadata-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PageMetadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageMetadata$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PageMetadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageMetadata-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageMetadata-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PageMetadata and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageMetadata$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageMetadata-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PageMetadata-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageMetadata$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PageMetadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageMetadata-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PageMetadata-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageMetadata$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageMetadata-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PageMetadata-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageMetadata$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageMetadata-print"></a>}}
\if{latex}{\out{\hypertarget{method-PageMetadata-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageMetadata$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageMetadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PageMetadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
