% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_structure_candidate_formula.R
\docType{class}
\name{PageStructureCandidateFormula}
\alias{PageStructureCandidateFormula}
\title{PageStructureCandidateFormula}
\format{
An \code{R6Class} generator object
}
\description{
PageStructureCandidateFormula Class

PageStructureCandidateFormula Class
}
\details{
Create a new PageStructureCandidateFormula
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalPages}}{integer [optional]}

\item{\code{totalElements}}{integer [optional]}

\item{\code{first}}{character [optional]}

\item{\code{size}}{integer [optional]}

\item{\code{content}}{list(\link{StructureCandidateFormula}) [optional]}

\item{\code{number}}{integer [optional]}

\item{\code{sort}}{\link{SortObject} [optional]}

\item{\code{last}}{character [optional]}

\item{\code{numberOfElements}}{integer [optional]}

\item{\code{pageable}}{\link{PageableObject} [optional]}

\item{\code{empty}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PageStructureCandidateFormula-new}{\code{PageStructureCandidateFormula$new()}}
\item \href{#method-PageStructureCandidateFormula-toJSON}{\code{PageStructureCandidateFormula$toJSON()}}
\item \href{#method-PageStructureCandidateFormula-fromJSON}{\code{PageStructureCandidateFormula$fromJSON()}}
\item \href{#method-PageStructureCandidateFormula-toJSONString}{\code{PageStructureCandidateFormula$toJSONString()}}
\item \href{#method-PageStructureCandidateFormula-fromJSONString}{\code{PageStructureCandidateFormula$fromJSONString()}}
\item \href{#method-PageStructureCandidateFormula-validateJSON}{\code{PageStructureCandidateFormula$validateJSON()}}
\item \href{#method-PageStructureCandidateFormula-toString}{\code{PageStructureCandidateFormula$toString()}}
\item \href{#method-PageStructureCandidateFormula-isValid}{\code{PageStructureCandidateFormula$isValid()}}
\item \href{#method-PageStructureCandidateFormula-getInvalidFields}{\code{PageStructureCandidateFormula$getInvalidFields()}}
\item \href{#method-PageStructureCandidateFormula-print}{\code{PageStructureCandidateFormula$print()}}
\item \href{#method-PageStructureCandidateFormula-clone}{\code{PageStructureCandidateFormula$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateFormula-new"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateFormula-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PageStructureCandidateFormula class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateFormula$new(
  totalPages = NULL,
  totalElements = NULL,
  first = NULL,
  size = NULL,
  content = NULL,
  number = NULL,
  sort = NULL,
  last = NULL,
  numberOfElements = NULL,
  pageable = NULL,
  empty = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalPages}}{totalPages}

\item{\code{totalElements}}{totalElements}

\item{\code{first}}{first}

\item{\code{size}}{size}

\item{\code{content}}{content}

\item{\code{number}}{number}

\item{\code{sort}}{sort}

\item{\code{last}}{last}

\item{\code{numberOfElements}}{numberOfElements}

\item{\code{pageable}}{pageable}

\item{\code{empty}}{empty}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateFormula-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateFormula-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateFormula$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PageStructureCandidateFormula in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateFormula-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateFormula-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PageStructureCandidateFormula
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateFormula$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PageStructureCandidateFormula
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateFormula-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateFormula-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateFormula$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PageStructureCandidateFormula in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateFormula-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateFormula-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PageStructureCandidateFormula
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateFormula$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PageStructureCandidateFormula
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateFormula-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateFormula-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PageStructureCandidateFormula and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateFormula$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateFormula-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateFormula-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateFormula$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PageStructureCandidateFormula
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateFormula-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateFormula-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateFormula$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateFormula-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateFormula-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateFormula$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateFormula-print"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateFormula-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateFormula$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateFormula-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateFormula-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateFormula$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
