% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_structure_candidate_scored.R
\docType{class}
\name{PageStructureCandidateScored}
\alias{PageStructureCandidateScored}
\title{PageStructureCandidateScored}
\format{
An \code{R6Class} generator object
}
\description{
PageStructureCandidateScored Class

PageStructureCandidateScored Class
}
\details{
Create a new PageStructureCandidateScored
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalPages}}{integer [optional]}

\item{\code{totalElements}}{integer [optional]}

\item{\code{last}}{character [optional]}

\item{\code{first}}{character [optional]}

\item{\code{size}}{integer [optional]}

\item{\code{content}}{list(\link{StructureCandidateScored}) [optional]}

\item{\code{number}}{integer [optional]}

\item{\code{sort}}{\link{SortObject} [optional]}

\item{\code{numberOfElements}}{integer [optional]}

\item{\code{pageable}}{\link{PageableObject} [optional]}

\item{\code{empty}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PageStructureCandidateScored-new}{\code{PageStructureCandidateScored$new()}}
\item \href{#method-PageStructureCandidateScored-toJSON}{\code{PageStructureCandidateScored$toJSON()}}
\item \href{#method-PageStructureCandidateScored-fromJSON}{\code{PageStructureCandidateScored$fromJSON()}}
\item \href{#method-PageStructureCandidateScored-toJSONString}{\code{PageStructureCandidateScored$toJSONString()}}
\item \href{#method-PageStructureCandidateScored-fromJSONString}{\code{PageStructureCandidateScored$fromJSONString()}}
\item \href{#method-PageStructureCandidateScored-validateJSON}{\code{PageStructureCandidateScored$validateJSON()}}
\item \href{#method-PageStructureCandidateScored-toString}{\code{PageStructureCandidateScored$toString()}}
\item \href{#method-PageStructureCandidateScored-isValid}{\code{PageStructureCandidateScored$isValid()}}
\item \href{#method-PageStructureCandidateScored-getInvalidFields}{\code{PageStructureCandidateScored$getInvalidFields()}}
\item \href{#method-PageStructureCandidateScored-print}{\code{PageStructureCandidateScored$print()}}
\item \href{#method-PageStructureCandidateScored-clone}{\code{PageStructureCandidateScored$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateScored-new"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateScored-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PageStructureCandidateScored class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateScored$new(
  totalPages = NULL,
  totalElements = NULL,
  last = NULL,
  first = NULL,
  size = NULL,
  content = NULL,
  number = NULL,
  sort = NULL,
  numberOfElements = NULL,
  pageable = NULL,
  empty = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalPages}}{totalPages}

\item{\code{totalElements}}{totalElements}

\item{\code{last}}{last}

\item{\code{first}}{first}

\item{\code{size}}{size}

\item{\code{content}}{content}

\item{\code{number}}{number}

\item{\code{sort}}{sort}

\item{\code{numberOfElements}}{numberOfElements}

\item{\code{pageable}}{pageable}

\item{\code{empty}}{empty}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateScored-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateScored-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateScored$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PageStructureCandidateScored in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateScored-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateScored-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PageStructureCandidateScored
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateScored$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PageStructureCandidateScored
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateScored-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateScored-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateScored$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PageStructureCandidateScored in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateScored-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateScored-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PageStructureCandidateScored
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateScored$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PageStructureCandidateScored
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateScored-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateScored-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PageStructureCandidateScored and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateScored$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateScored-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateScored-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateScored$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PageStructureCandidateScored
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateScored-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateScored-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateScored$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateScored-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateScored-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateScored$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateScored-print"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateScored-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateScored$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageStructureCandidateScored-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PageStructureCandidateScored-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageStructureCandidateScored$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
