% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_pair.R
\docType{class}
\name{PeakPair}
\alias{PeakPair}
\title{PeakPair}
\format{
An \code{R6Class} generator object
}
\description{
PeakPair Class

PeakPair Class
}
\details{
Create a new PeakPair
}
\examples{

## ------------------------------------------------
## Method `PeakPair$toList`
## ------------------------------------------------

# convert array of PeakPair (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{queryPeak}}{integer}

\item{\code{referencePeak}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PeakPair-new}{\code{PeakPair$new()}}
\item \href{#method-PeakPair-toJSON}{\code{PeakPair$toJSON()}}
\item \href{#method-PeakPair-toList}{\code{PeakPair$toList()}}
\item \href{#method-PeakPair-toSimpleType}{\code{PeakPair$toSimpleType()}}
\item \href{#method-PeakPair-fromJSON}{\code{PeakPair$fromJSON()}}
\item \href{#method-PeakPair-toJSONString}{\code{PeakPair$toJSONString()}}
\item \href{#method-PeakPair-fromJSONString}{\code{PeakPair$fromJSONString()}}
\item \href{#method-PeakPair-validateJSON}{\code{PeakPair$validateJSON()}}
\item \href{#method-PeakPair-toString}{\code{PeakPair$toString()}}
\item \href{#method-PeakPair-isValid}{\code{PeakPair$isValid()}}
\item \href{#method-PeakPair-getInvalidFields}{\code{PeakPair$getInvalidFields()}}
\item \href{#method-PeakPair-print}{\code{PeakPair$print()}}
\item \href{#method-PeakPair-clone}{\code{PeakPair$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakPair-new"></a>}}
\if{latex}{\out{\hypertarget{method-PeakPair-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PeakPair class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakPair$new(queryPeak, referencePeak, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{queryPeak}}{queryPeak}

\item{\code{referencePeak}}{referencePeak}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakPair-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PeakPair-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakPair$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakPair-toList"></a>}}
\if{latex}{\out{\hypertarget{method-PeakPair-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakPair$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PeakPair as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of PeakPair (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakPair-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-PeakPair-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert PeakPair to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakPair$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakPair-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PeakPair-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PeakPair
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakPair$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PeakPair
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakPair-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PeakPair-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakPair$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PeakPair in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakPair-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PeakPair-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PeakPair
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakPair$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PeakPair
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakPair-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PeakPair-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PeakPair and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakPair$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakPair-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PeakPair-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakPair$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PeakPair
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakPair-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PeakPair-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakPair$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakPair-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PeakPair-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakPair$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakPair-print"></a>}}
\if{latex}{\out{\hypertarget{method-PeakPair-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakPair$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakPair-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PeakPair-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakPair$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
