% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_detail.R
\docType{class}
\name{ProblemDetail}
\alias{ProblemDetail}
\title{ProblemDetail}
\format{
An \code{R6Class} generator object
}
\description{
ProblemDetail Class

ProblemDetail Class
}
\details{
Create a new ProblemDetail
}
\examples{

## ------------------------------------------------
## Method `ProblemDetail$toList`
## ------------------------------------------------

# convert array of ProblemDetail (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{character [optional]}

\item{\code{title}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{detail}}{character [optional]}

\item{\code{instance}}{character [optional]}

\item{\code{properties}}{named list(object) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProblemDetail-new}{\code{ProblemDetail$new()}}
\item \href{#method-ProblemDetail-toJSON}{\code{ProblemDetail$toJSON()}}
\item \href{#method-ProblemDetail-toList}{\code{ProblemDetail$toList()}}
\item \href{#method-ProblemDetail-toSimpleType}{\code{ProblemDetail$toSimpleType()}}
\item \href{#method-ProblemDetail-fromJSON}{\code{ProblemDetail$fromJSON()}}
\item \href{#method-ProblemDetail-toJSONString}{\code{ProblemDetail$toJSONString()}}
\item \href{#method-ProblemDetail-fromJSONString}{\code{ProblemDetail$fromJSONString()}}
\item \href{#method-ProblemDetail-validateJSON}{\code{ProblemDetail$validateJSON()}}
\item \href{#method-ProblemDetail-toString}{\code{ProblemDetail$toString()}}
\item \href{#method-ProblemDetail-isValid}{\code{ProblemDetail$isValid()}}
\item \href{#method-ProblemDetail-getInvalidFields}{\code{ProblemDetail$getInvalidFields()}}
\item \href{#method-ProblemDetail-print}{\code{ProblemDetail$print()}}
\item \href{#method-ProblemDetail-clone}{\code{ProblemDetail$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemDetail-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemDetail-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ProblemDetail class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemDetail$new(
  type = NULL,
  title = NULL,
  status = NULL,
  detail = NULL,
  instance = NULL,
  properties = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{type}

\item{\code{title}}{title}

\item{\code{status}}{status}

\item{\code{detail}}{detail}

\item{\code{instance}}{instance}

\item{\code{properties}}{properties}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemDetail-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemDetail-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemDetail$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemDetail-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemDetail-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemDetail$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ProblemDetail as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of ProblemDetail (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemDetail-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemDetail-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert ProblemDetail to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemDetail$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemDetail-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemDetail-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ProblemDetail
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemDetail$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ProblemDetail
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemDetail-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemDetail-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemDetail$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ProblemDetail in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemDetail-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemDetail-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ProblemDetail
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemDetail$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ProblemDetail
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemDetail-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemDetail-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ProblemDetail and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemDetail$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemDetail-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemDetail-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemDetail$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ProblemDetail
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemDetail-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemDetail-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemDetail$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemDetail-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemDetail-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemDetail$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemDetail-print"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemDetail-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemDetail$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemDetail-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemDetail-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemDetail$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
