% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_info_opt_field.R
\docType{class}
\name{ProjectInfoOptField}
\alias{ProjectInfoOptField}
\title{ProjectInfoOptField}
\format{
An \code{R6Class} generator object
}
\description{
ProjectInfoOptField Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectInfoOptField-new}{\code{ProjectInfoOptField$new()}}
\item \href{#method-ProjectInfoOptField-toJSON}{\code{ProjectInfoOptField$toJSON()}}
\item \href{#method-ProjectInfoOptField-fromJSON}{\code{ProjectInfoOptField$fromJSON()}}
\item \href{#method-ProjectInfoOptField-toJSONString}{\code{ProjectInfoOptField$toJSONString()}}
\item \href{#method-ProjectInfoOptField-fromJSONString}{\code{ProjectInfoOptField$fromJSONString()}}
\item \href{#method-ProjectInfoOptField-clone}{\code{ProjectInfoOptField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectInfoOptField-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectInfoOptField-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ProjectInfoOptField class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectInfoOptField$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectInfoOptField-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectInfoOptField-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectInfoOptField$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ProjectInfoOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectInfoOptField-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectInfoOptField-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ProjectInfoOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectInfoOptField$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ProjectInfoOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectInfoOptField-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectInfoOptField-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectInfoOptField$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ProjectInfoOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectInfoOptField-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectInfoOptField-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ProjectInfoOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectInfoOptField$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ProjectInfoOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectInfoOptField-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectInfoOptField-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectInfoOptField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
