% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_space_id.R
\docType{class}
\name{ProjectSpaceId}
\alias{ProjectSpaceId}
\title{ProjectSpaceId}
\format{
An \code{R6Class} generator object
}
\description{
ProjectSpaceId Class
}
\details{
Create a new ProjectSpaceId
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{path}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectSpaceId-new}{\code{ProjectSpaceId$new()}}
\item \href{#method-ProjectSpaceId-toJSON}{\code{ProjectSpaceId$toJSON()}}
\item \href{#method-ProjectSpaceId-fromJSON}{\code{ProjectSpaceId$fromJSON()}}
\item \href{#method-ProjectSpaceId-toJSONString}{\code{ProjectSpaceId$toJSONString()}}
\item \href{#method-ProjectSpaceId-fromJSONString}{\code{ProjectSpaceId$fromJSONString()}}
\item \href{#method-ProjectSpaceId-validateJSON}{\code{ProjectSpaceId$validateJSON()}}
\item \href{#method-ProjectSpaceId-toString}{\code{ProjectSpaceId$toString()}}
\item \href{#method-ProjectSpaceId-isValid}{\code{ProjectSpaceId$isValid()}}
\item \href{#method-ProjectSpaceId-getInvalidFields}{\code{ProjectSpaceId$getInvalidFields()}}
\item \href{#method-ProjectSpaceId-print}{\code{ProjectSpaceId$print()}}
\item \href{#method-ProjectSpaceId-clone}{\code{ProjectSpaceId$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSpaceId-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSpaceId-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ProjectSpaceId class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSpaceId$new(name = NULL, path = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{path}}{path}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSpaceId-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSpaceId-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSpaceId$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ProjectSpaceId in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSpaceId-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSpaceId-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ProjectSpaceId
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSpaceId$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ProjectSpaceId
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSpaceId-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSpaceId-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSpaceId$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ProjectSpaceId in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSpaceId-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSpaceId-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ProjectSpaceId
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSpaceId$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ProjectSpaceId
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSpaceId-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSpaceId-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ProjectSpaceId and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSpaceId$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSpaceId-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSpaceId-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSpaceId$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ProjectSpaceId
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSpaceId-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSpaceId-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSpaceId$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSpaceId-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSpaceId-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSpaceId$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSpaceId-print"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSpaceId-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSpaceId$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSpaceId-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSpaceId-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSpaceId$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
