% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_item.R
\docType{class}
\name{QualityItem}
\alias{QualityItem}
\title{QualityItem}
\format{
An \code{R6Class} generator object
}
\description{
QualityItem Class

QualityItem Class
}
\details{
Create a new QualityItem
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character [optional]}

\item{\code{quality}}{character [optional]}

\item{\code{weight}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QualityItem-new}{\code{QualityItem$new()}}
\item \href{#method-QualityItem-toJSON}{\code{QualityItem$toJSON()}}
\item \href{#method-QualityItem-fromJSON}{\code{QualityItem$fromJSON()}}
\item \href{#method-QualityItem-toJSONString}{\code{QualityItem$toJSONString()}}
\item \href{#method-QualityItem-fromJSONString}{\code{QualityItem$fromJSONString()}}
\item \href{#method-QualityItem-validateJSON}{\code{QualityItem$validateJSON()}}
\item \href{#method-QualityItem-toString}{\code{QualityItem$toString()}}
\item \href{#method-QualityItem-isValid}{\code{QualityItem$isValid()}}
\item \href{#method-QualityItem-getInvalidFields}{\code{QualityItem$getInvalidFields()}}
\item \href{#method-QualityItem-print}{\code{QualityItem$print()}}
\item \href{#method-QualityItem-clone}{\code{QualityItem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualityItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-QualityItem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new QualityItem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualityItem$new(description = NULL, quality = NULL, weight = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{description}

\item{\code{quality}}{quality}

\item{\code{weight}}{weight}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualityItem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QualityItem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualityItem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
QualityItem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualityItem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QualityItem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of QualityItem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualityItem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QualityItem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualityItem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QualityItem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualityItem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
QualityItem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualityItem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QualityItem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of QualityItem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualityItem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QualityItem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualityItem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QualityItem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to QualityItem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualityItem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualityItem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-QualityItem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualityItem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of QualityItem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualityItem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-QualityItem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualityItem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualityItem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-QualityItem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualityItem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualityItem-print"></a>}}
\if{latex}{\out{\hypertarget{method-QualityItem-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualityItem$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualityItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QualityItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualityItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
