% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification_table.R
\docType{class}
\name{QuantificationTable}
\alias{QuantificationTable}
\title{QuantificationTable}
\format{
An \code{R6Class} generator object
}
\description{
QuantificationTable Class

QuantificationTable Class
}
\details{
Create a new QuantificationTable
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{quantificationType}}{character [optional]}

\item{\code{rowType}}{character [optional]}

\item{\code{columnType}}{character [optional]}

\item{\code{rowIds}}{list(integer) [optional]}

\item{\code{columnIds}}{list(integer) [optional]}

\item{\code{rowNames}}{list(character) [optional]}

\item{\code{columnNames}}{list(character) [optional]}

\item{\code{values}}{list(list(numeric)) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QuantificationTable-new}{\code{QuantificationTable$new()}}
\item \href{#method-QuantificationTable-toJSON}{\code{QuantificationTable$toJSON()}}
\item \href{#method-QuantificationTable-fromJSON}{\code{QuantificationTable$fromJSON()}}
\item \href{#method-QuantificationTable-toJSONString}{\code{QuantificationTable$toJSONString()}}
\item \href{#method-QuantificationTable-fromJSONString}{\code{QuantificationTable$fromJSONString()}}
\item \href{#method-QuantificationTable-validateJSON}{\code{QuantificationTable$validateJSON()}}
\item \href{#method-QuantificationTable-toString}{\code{QuantificationTable$toString()}}
\item \href{#method-QuantificationTable-isValid}{\code{QuantificationTable$isValid()}}
\item \href{#method-QuantificationTable-getInvalidFields}{\code{QuantificationTable$getInvalidFields()}}
\item \href{#method-QuantificationTable-print}{\code{QuantificationTable$print()}}
\item \href{#method-QuantificationTable-clone}{\code{QuantificationTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTable-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new QuantificationTable class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTable$new(
  quantificationType = NULL,
  rowType = NULL,
  columnType = NULL,
  rowIds = NULL,
  columnIds = NULL,
  rowNames = NULL,
  columnNames = NULL,
  values = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantificationType}}{quantificationType}

\item{\code{rowType}}{rowType}

\item{\code{columnType}}{columnType}

\item{\code{rowIds}}{rowIds}

\item{\code{columnIds}}{columnIds}

\item{\code{rowNames}}{rowNames}

\item{\code{columnNames}}{columnNames}

\item{\code{values}}{values}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTable-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTable-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTable$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
QuantificationTable in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTable-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTable-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of QuantificationTable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTable$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QuantificationTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTable-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTable-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTable$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
QuantificationTable in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTable-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTable-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of QuantificationTable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTable$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QuantificationTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTable-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTable-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to QuantificationTable and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTable$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTable-toString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTable-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTable$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of QuantificationTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTable-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTable-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTable$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTable-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTable-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTable$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTable-print"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTable-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTable$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
