% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification_table_experimental.R
\docType{class}
\name{QuantificationTableExperimental}
\alias{QuantificationTableExperimental}
\title{QuantificationTableExperimental}
\format{
An \code{R6Class} generator object
}
\description{
EXPERIMENTAL: This schema is experimental and may be changed (or even removed) without notice until it is declared stable.

QuantificationTableExperimental Class
}
\details{
Create a new QuantificationTableExperimental
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{quantificationType}}{character [optional]}

\item{\code{rowType}}{character [optional]}

\item{\code{columnType}}{character [optional]}

\item{\code{rowIds}}{list(integer) [optional]}

\item{\code{columnIds}}{list(integer) [optional]}

\item{\code{rowNames}}{list(character) [optional]}

\item{\code{columnNames}}{list(character) [optional]}

\item{\code{values}}{list(list(numeric)) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QuantificationTableExperimental-new}{\code{QuantificationTableExperimental$new()}}
\item \href{#method-QuantificationTableExperimental-toJSON}{\code{QuantificationTableExperimental$toJSON()}}
\item \href{#method-QuantificationTableExperimental-fromJSON}{\code{QuantificationTableExperimental$fromJSON()}}
\item \href{#method-QuantificationTableExperimental-toJSONString}{\code{QuantificationTableExperimental$toJSONString()}}
\item \href{#method-QuantificationTableExperimental-fromJSONString}{\code{QuantificationTableExperimental$fromJSONString()}}
\item \href{#method-QuantificationTableExperimental-validateJSON}{\code{QuantificationTableExperimental$validateJSON()}}
\item \href{#method-QuantificationTableExperimental-toString}{\code{QuantificationTableExperimental$toString()}}
\item \href{#method-QuantificationTableExperimental-isValid}{\code{QuantificationTableExperimental$isValid()}}
\item \href{#method-QuantificationTableExperimental-getInvalidFields}{\code{QuantificationTableExperimental$getInvalidFields()}}
\item \href{#method-QuantificationTableExperimental-print}{\code{QuantificationTableExperimental$print()}}
\item \href{#method-QuantificationTableExperimental-clone}{\code{QuantificationTableExperimental$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTableExperimental-new"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTableExperimental-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new QuantificationTableExperimental class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTableExperimental$new(
  quantificationType = NULL,
  rowType = NULL,
  columnType = NULL,
  rowIds = NULL,
  columnIds = NULL,
  rowNames = NULL,
  columnNames = NULL,
  values = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantificationType}}{quantificationType}

\item{\code{rowType}}{rowType}

\item{\code{columnType}}{columnType}

\item{\code{rowIds}}{rowIds}

\item{\code{columnIds}}{columnIds}

\item{\code{rowNames}}{rowNames}

\item{\code{columnNames}}{columnNames}

\item{\code{values}}{values}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTableExperimental-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTableExperimental-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTableExperimental$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
QuantificationTableExperimental in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTableExperimental-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTableExperimental-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of QuantificationTableExperimental
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTableExperimental$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QuantificationTableExperimental
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTableExperimental-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTableExperimental-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTableExperimental$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
QuantificationTableExperimental in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTableExperimental-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTableExperimental-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of QuantificationTableExperimental
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTableExperimental$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QuantificationTableExperimental
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTableExperimental-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTableExperimental-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to QuantificationTableExperimental and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTableExperimental$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTableExperimental-toString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTableExperimental-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTableExperimental$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of QuantificationTableExperimental
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTableExperimental-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTableExperimental-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTableExperimental$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTableExperimental-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTableExperimental-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTableExperimental$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTableExperimental-print"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTableExperimental-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTableExperimental$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantificationTableExperimental-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QuantificationTableExperimental-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantificationTableExperimental$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
