% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_table.R
\docType{class}
\name{QuantTable}
\alias{QuantTable}
\title{QuantTable}
\format{
An \code{R6Class} generator object
}
\description{
Quantification of features or compounds within the runs they have been detected in. Rows refer to the quantified objects, columns to the runs. Values that could not be quantified are NaN.

QuantTable Class
}
\details{
Create a new QuantTable
}
\examples{

## ------------------------------------------------
## Method `QuantTable$toList`
## ------------------------------------------------

# convert array of QuantTable (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{quantificationMeasure}}{character [optional]}

\item{\code{rowType}}{character [optional]}

\item{\code{rowIds}}{Ids of the quantified objects, features or compounds depending on the row type. list(character) [optional]}

\item{\code{columnIds}}{Ids of the runs the objects are quantified in. list(character) [optional]}

\item{\code{columnNames}}{Names of the runs the objects are quantified in, in the order of columnIds.  <p>  Optional field, only present if requested, since the table can be read by run id alone.  <p>  The name is the one the run carries in the project. It is either the sample name given when the data was  imported, or, if none was given, derived from the measurement itself: the run id inside the file, and only  failing that the file name. A name is therefore not necessarily the name of the file the run came from, and  it is not guaranteed to be unique. Use columnSources to identify the input file, and  columnIds to identify the run. list(character) [optional]}

\item{\code{columnSources}}{Files the runs were imported from, in the order of columnIds, to relate a column back to the input  data. Same value as the source of the corresponding run.  <p>  Optional field, only present if requested, since it is not needed to read the table. list(character) [optional]}

\item{\code{values}}{list(list(numeric)) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QuantTable-new}{\code{QuantTable$new()}}
\item \href{#method-QuantTable-toJSON}{\code{QuantTable$toJSON()}}
\item \href{#method-QuantTable-toList}{\code{QuantTable$toList()}}
\item \href{#method-QuantTable-toSimpleType}{\code{QuantTable$toSimpleType()}}
\item \href{#method-QuantTable-fromJSON}{\code{QuantTable$fromJSON()}}
\item \href{#method-QuantTable-toJSONString}{\code{QuantTable$toJSONString()}}
\item \href{#method-QuantTable-fromJSONString}{\code{QuantTable$fromJSONString()}}
\item \href{#method-QuantTable-validateJSON}{\code{QuantTable$validateJSON()}}
\item \href{#method-QuantTable-toString}{\code{QuantTable$toString()}}
\item \href{#method-QuantTable-isValid}{\code{QuantTable$isValid()}}
\item \href{#method-QuantTable-getInvalidFields}{\code{QuantTable$getInvalidFields()}}
\item \href{#method-QuantTable-print}{\code{QuantTable$print()}}
\item \href{#method-QuantTable-clone}{\code{QuantTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTable-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new QuantTable class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTable$new(
  quantificationMeasure = NULL,
  rowType = NULL,
  rowIds = NULL,
  columnIds = NULL,
  columnNames = NULL,
  columnSources = NULL,
  values = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantificationMeasure}}{quantificationMeasure}

\item{\code{rowType}}{rowType}

\item{\code{rowIds}}{Ids of the quantified objects, features or compounds depending on the row type.}

\item{\code{columnIds}}{Ids of the runs the objects are quantified in.}

\item{\code{columnNames}}{Names of the runs the objects are quantified in, in the order of columnIds.  <p>  Optional field, only present if requested, since the table can be read by run id alone.  <p>  The name is the one the run carries in the project. It is either the sample name given when the data was  imported, or, if none was given, derived from the measurement itself: the run id inside the file, and only  failing that the file name. A name is therefore not necessarily the name of the file the run came from, and  it is not guaranteed to be unique. Use columnSources to identify the input file, and  columnIds to identify the run.}

\item{\code{columnSources}}{Files the runs were imported from, in the order of columnIds, to relate a column back to the input  data. Same value as the source of the corresponding run.  <p>  Optional field, only present if requested, since it is not needed to read the table.}

\item{\code{values}}{values}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTable-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTable-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTable$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTable-toList"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTable-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTable$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
QuantTable as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of QuantTable (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTable-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTable-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert QuantTable to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTable$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTable-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTable-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of QuantTable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTable$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QuantTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTable-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTable-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTable$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
QuantTable in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTable-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTable-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of QuantTable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTable$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QuantTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTable-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTable-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to QuantTable and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTable$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTable-toString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTable-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTable$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of QuantTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTable-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTable-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTable$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTable-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTable-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTable$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTable-print"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTable-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTable$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
