% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_table_opt_field.R
\docType{class}
\name{QuantTableOptField}
\alias{QuantTableOptField}
\title{QuantTableOptField}
\format{
An \code{R6Class} generator object
}
\description{
QuantTableOptField Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QuantTableOptField-new}{\code{QuantTableOptField$new()}}
\item \href{#method-QuantTableOptField-toJSON}{\code{QuantTableOptField$toJSON()}}
\item \href{#method-QuantTableOptField-fromJSON}{\code{QuantTableOptField$fromJSON()}}
\item \href{#method-QuantTableOptField-toJSONString}{\code{QuantTableOptField$toJSONString()}}
\item \href{#method-QuantTableOptField-fromJSONString}{\code{QuantTableOptField$fromJSONString()}}
\item \href{#method-QuantTableOptField-clone}{\code{QuantTableOptField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTableOptField-new"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTableOptField-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new QuantTableOptField class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTableOptField$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTableOptField-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTableOptField-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTableOptField$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
QuantTableOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTableOptField-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTableOptField-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of QuantTableOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTableOptField$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QuantTableOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTableOptField-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTableOptField-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTableOptField$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
QuantTableOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTableOptField-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTableOptField-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of QuantTableOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTableOptField$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QuantTableOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantTableOptField-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QuantTableOptField-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantTableOptField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
