% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_overview.R
\docType{class}
\name{ResultOverview}
\alias{ResultOverview}
\title{ResultOverview}
\format{
An \code{R6Class} generator object
}
\description{
Results that are available for a {@link FormulaResultContainer FormulaResultContainer} represented as boolean or numeric score (if available).  NULL scores indicate that the corresponding result is not available.

ResultOverview Class
}
\details{
Create a new ResultOverview
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{siriusScore}}{Sirius Score (isotope + tree score) of the formula candidate.  If NULL result is not available numeric [optional]}

\item{\code{zodiacScore}}{Zodiac Score of the formula candidate.  If NULL result is not available numeric [optional]}

\item{\code{topCSIScore}}{CSI:FingerID Score of the top ranking structure candidate of the structure database search  performed for this formula candidate.  If NULL structure database result is not available or the structure candidate list is empty. numeric [optional]}

\item{\code{confidenceScore}}{Confidence Score of the  IF NULL structure database result not available a structure candidate hit of another formula candidate is  the top ranking structure candidate. numeric [optional]}

\item{\code{canopusResult}}{True if Canopus compound class prediction results are available. character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResultOverview-new}{\code{ResultOverview$new()}}
\item \href{#method-ResultOverview-toJSON}{\code{ResultOverview$toJSON()}}
\item \href{#method-ResultOverview-fromJSON}{\code{ResultOverview$fromJSON()}}
\item \href{#method-ResultOverview-toJSONString}{\code{ResultOverview$toJSONString()}}
\item \href{#method-ResultOverview-fromJSONString}{\code{ResultOverview$fromJSONString()}}
\item \href{#method-ResultOverview-validateJSON}{\code{ResultOverview$validateJSON()}}
\item \href{#method-ResultOverview-toString}{\code{ResultOverview$toString()}}
\item \href{#method-ResultOverview-isValid}{\code{ResultOverview$isValid()}}
\item \href{#method-ResultOverview-getInvalidFields}{\code{ResultOverview$getInvalidFields()}}
\item \href{#method-ResultOverview-print}{\code{ResultOverview$print()}}
\item \href{#method-ResultOverview-clone}{\code{ResultOverview$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultOverview-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResultOverview-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ResultOverview class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultOverview$new(
  siriusScore = NULL,
  zodiacScore = NULL,
  topCSIScore = NULL,
  confidenceScore = NULL,
  canopusResult = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{siriusScore}}{Sirius Score (isotope + tree score) of the formula candidate.  If NULL result is not available}

\item{\code{zodiacScore}}{Zodiac Score of the formula candidate.  If NULL result is not available}

\item{\code{topCSIScore}}{CSI:FingerID Score of the top ranking structure candidate of the structure database search  performed for this formula candidate.  If NULL structure database result is not available or the structure candidate list is empty.}

\item{\code{confidenceScore}}{Confidence Score of the  IF NULL structure database result not available a structure candidate hit of another formula candidate is  the top ranking structure candidate.}

\item{\code{canopusResult}}{True if Canopus compound class prediction results are available.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultOverview-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResultOverview-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultOverview$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResultOverview in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultOverview-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResultOverview-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ResultOverview
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultOverview$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResultOverview
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultOverview-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResultOverview-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultOverview$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResultOverview in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultOverview-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResultOverview-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ResultOverview
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultOverview$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResultOverview
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultOverview-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResultOverview-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ResultOverview and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultOverview$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultOverview-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ResultOverview-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultOverview$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ResultOverview
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultOverview-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ResultOverview-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultOverview$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultOverview-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ResultOverview-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultOverview$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultOverview-print"></a>}}
\if{latex}{\out{\hypertarget{method-ResultOverview-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultOverview$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultOverview-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResultOverview-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultOverview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
