% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsirius_api.R
\docType{class}
\name{rsirius_api}
\alias{rsirius_api}
\title{ApiClient}
\format{
An \code{R6Class} generator object
}
\description{
ApiClient Class
}
\details{
SIRIUS Nightsky API

REST API that provides the full functionality of SIRIUS and its web services as background service. It is intended as entry-point for scripting languages and software integration SDKs.This API is exposed by SIRIUS 6

The version of the OpenAPI document: 3.2
Generated by: https://openapi-generator.tech

RSirius API Class

A single point of access to the RSirius API.

NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
Ref: https://openapi-generator.tech
Do not edit the class manually.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{API client}

\item{\code{actuator_api}}{an instance of ActuatorApi}

\item{\code{compounds_api}}{an instance of CompoundsApi}

\item{\code{features_api}}{an instance of FeaturesApi}

\item{\code{gui_api}}{an instance of GuiApi}

\item{\code{info_api}}{an instance of InfoApi}

\item{\code{jobs_api}}{an instance of JobsApi}

\item{\code{login_and_account_api}}{an instance of LoginAndAccountApi}

\item{\code{projects_api}}{an instance of ProjectsApi}

\item{\code{runs_api}}{an instance of RunsApi}

\item{\code{searchable_databases_api}}{an instance of SearchableDatabasesApi}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rsirius_api-new}{\code{rsirius_api$new()}}
\item \href{#method-rsirius_api-wait_for_job_completion}{\code{rsirius_api$wait_for_job_completion()}}
\item \href{#method-rsirius_api-clone}{\code{rsirius_api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rsirius_api-new"></a>}}
\if{latex}{\out{\hypertarget{method-rsirius_api-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RSirius API Class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rsirius_api$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client (optional).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rsirius_api-wait_for_job_completion"></a>}}
\if{latex}{\out{\hypertarget{method-rsirius_api-wait_for_job_completion}{}}}
\subsection{Method \code{wait_for_job_completion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rsirius_api$wait_for_job_completion(project_id, job_id, timeout_in_sec = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rsirius_api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rsirius_api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rsirius_api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
