% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_opt_field.R
\docType{class}
\name{RunOptField}
\alias{RunOptField}
\title{RunOptField}
\format{
An \code{R6Class} generator object
}
\description{
RunOptField Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RunOptField-new}{\code{RunOptField$new()}}
\item \href{#method-RunOptField-toJSON}{\code{RunOptField$toJSON()}}
\item \href{#method-RunOptField-toSimpleType}{\code{RunOptField$toSimpleType()}}
\item \href{#method-RunOptField-fromJSON}{\code{RunOptField$fromJSON()}}
\item \href{#method-RunOptField-toJSONString}{\code{RunOptField$toJSONString()}}
\item \href{#method-RunOptField-fromJSONString}{\code{RunOptField$fromJSONString()}}
\item \href{#method-RunOptField-clone}{\code{RunOptField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunOptField-new"></a>}}
\if{latex}{\out{\hypertarget{method-RunOptField-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RunOptField class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunOptField$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunOptField-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RunOptField-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunOptField$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunOptField-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-RunOptField-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert RunOptField to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunOptField$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunOptField-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RunOptField-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of RunOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunOptField$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RunOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunOptField-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RunOptField-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunOptField$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
RunOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunOptField-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RunOptField-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of RunOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunOptField$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RunOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunOptField-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RunOptField-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunOptField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
