% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs_api.R
\docType{class}
\name{RunsApi}
\alias{RunsApi}
\title{Runs operations}
\format{
An \code{R6Class} generator object
}
\description{
RunsApi
}
\details{
SIRIUS Nightsky API

REST API that provides the full functionality of SIRIUS and its web services as background service. It is intended as entry-point for scripting languages and software integration SDKs.This API is exposed by SIRIUS 6

The version of the OpenAPI document: 3.2
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  GetRun  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project-space to read from.
var_run_id <- "run_id_example" # character | identifier of run to access.
var_opt_fields <- c(RunOptField$new()) # array[RunOptField] | set of optional fields to be included. Use 'none' only to override defaults. (Optional)

#Get run with the given identifier from the specified project-space.
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetRun(var_project_id, var_run_id, opt_fields = var_opt_fieldsdata_file = "result.txt")
result <- api_instance$runs_api$GetRun(var_project_id, var_run_id, opt_fields = var_opt_fields)
dput(result)


####################  GetRunsPage  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project space to get runs from.
var_search_query <- "search_query_example" # character | Optional search query in lucene syntax. Omit this parameter to page over all runs. (Optional)
var_page <- 0 # integer | Zero-based page index (0..N) (Optional)
var_size <- 20 # integer | The size of the page to be returned (Optional)
var_sort <- c("inner_example") # array[character] | Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported. (Optional)
var_opt_fields <- c(RunOptField$new()) # array[RunOptField] | set of optional fields to be included. Use 'none' only to override defaults. (Optional)

#Get runs in the given project-space
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetRunsPage(var_project_id, search_query = var_search_query, page = var_page, size = var_size, sort = var_sort, opt_fields = var_opt_fieldsdata_file = "result.txt")
result <- api_instance$runs_api$GetRunsPage(var_project_id, search_query = var_search_query, page = var_page, size = var_size, sort = var_sort, opt_fields = var_opt_fields)
dput(result)


####################  GetRunsSearchableFields  ####################

library(RSirius)
var_project_id <- "project_id_example" # character | project space to read the searchable run fields from.

#Get all fields that can be used in the searchQuery parameter of run endpoints
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetRunsSearchableFields(var_project_iddata_file = "result.txt")
result <- api_instance$runs_api$GetRunsSearchableFields(var_project_id)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RunsApi-new}{\code{RunsApi$new()}}
\item \href{#method-RunsApi-GetRun}{\code{RunsApi$GetRun()}}
\item \href{#method-RunsApi-GetRunWithHttpInfo}{\code{RunsApi$GetRunWithHttpInfo()}}
\item \href{#method-RunsApi-GetRunsPage}{\code{RunsApi$GetRunsPage()}}
\item \href{#method-RunsApi-GetRunsPageWithHttpInfo}{\code{RunsApi$GetRunsPageWithHttpInfo()}}
\item \href{#method-RunsApi-GetRunsSearchableFields}{\code{RunsApi$GetRunsSearchableFields()}}
\item \href{#method-RunsApi-GetRunsSearchableFieldsWithHttpInfo}{\code{RunsApi$GetRunsSearchableFieldsWithHttpInfo()}}
\item \href{#method-RunsApi-clone}{\code{RunsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunsApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-RunsApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RunsApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunsApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunsApi-GetRun"></a>}}
\if{latex}{\out{\hypertarget{method-RunsApi-GetRun}{}}}
\subsection{Method \code{GetRun()}}{
Get run with the given identifier from the specified project-space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunsApi$GetRun(project_id, run_id, opt_fields = list(), data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{run_id}}{identifier of run to access.}

\item{\code{opt_fields}}{(optional) set of optional fields to be included. Use 'none' only to override defaults. (default value: [])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Run
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunsApi-GetRunWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-RunsApi-GetRunWithHttpInfo}{}}}
\subsection{Method \code{GetRunWithHttpInfo()}}{
Get run with the given identifier from the specified project-space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunsApi$GetRunWithHttpInfo(
  project_id,
  run_id,
  opt_fields = list(),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project-space to read from.}

\item{\code{run_id}}{identifier of run to access.}

\item{\code{opt_fields}}{(optional) set of optional fields to be included. Use 'none' only to override defaults. (default value: [])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (Run) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunsApi-GetRunsPage"></a>}}
\if{latex}{\out{\hypertarget{method-RunsApi-GetRunsPage}{}}}
\subsection{Method \code{GetRunsPage()}}{
Get runs in the given project-space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunsApi$GetRunsPage(
  project_id,
  search_query = NULL,
  page = 0,
  size = 20,
  sort = NULL,
  opt_fields = list(),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project space to get runs from.}

\item{\code{search_query}}{(optional) Optional search query in lucene syntax. Omit this parameter to page over all runs.}

\item{\code{page}}{(optional) Zero-based page index (0..N) (default value: 0)}

\item{\code{size}}{(optional) The size of the page to be returned (default value: 20)}

\item{\code{sort}}{(optional) Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.}

\item{\code{opt_fields}}{(optional) set of optional fields to be included. Use 'none' only to override defaults. (default value: [])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PagedModelRun
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunsApi-GetRunsPageWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-RunsApi-GetRunsPageWithHttpInfo}{}}}
\subsection{Method \code{GetRunsPageWithHttpInfo()}}{
Get runs in the given project-space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunsApi$GetRunsPageWithHttpInfo(
  project_id,
  search_query = NULL,
  page = 0,
  size = 20,
  sort = NULL,
  opt_fields = list(),
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project space to get runs from.}

\item{\code{search_query}}{(optional) Optional search query in lucene syntax. Omit this parameter to page over all runs.}

\item{\code{page}}{(optional) Zero-based page index (0..N) (default value: 0)}

\item{\code{size}}{(optional) The size of the page to be returned (default value: 20)}

\item{\code{sort}}{(optional) Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.}

\item{\code{opt_fields}}{(optional) set of optional fields to be included. Use 'none' only to override defaults. (default value: [])}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (PagedModelRun) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunsApi-GetRunsSearchableFields"></a>}}
\if{latex}{\out{\hypertarget{method-RunsApi-GetRunsSearchableFields}{}}}
\subsection{Method \code{GetRunsSearchableFields()}}{
Get all fields that can be used in the searchQuery parameter of run endpoints
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunsApi$GetRunsSearchableFields(project_id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project space to read the searchable run fields from.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array[SearchableField]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunsApi-GetRunsSearchableFieldsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-RunsApi-GetRunsSearchableFieldsWithHttpInfo}{}}}
\subsection{Method \code{GetRunsSearchableFieldsWithHttpInfo()}}{
Get all fields that can be used in the searchQuery parameter of run endpoints
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunsApi$GetRunsSearchableFieldsWithHttpInfo(project_id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{project space to read the searchable run fields from.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (array[SearchableField]) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunsApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RunsApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
