% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchable_database_parameters.R
\docType{class}
\name{SearchableDatabaseParameters}
\alias{SearchableDatabaseParameters}
\title{SearchableDatabaseParameters}
\format{
An \code{R6Class} generator object
}
\description{
SearchableDatabaseParameters Class

SearchableDatabaseParameters Class
}
\details{
Create a new SearchableDatabaseParameters
}
\examples{

## ------------------------------------------------
## Method `SearchableDatabaseParameters$toList`
## ------------------------------------------------

# convert array of SearchableDatabaseParameters (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{displayName}}{display name of the database  Should be short character [optional]}

\item{\code{location}}{Storage location of user database  Might be NULL for non-user databases or if default location is used. character [optional]}

\item{\code{matchRtOfReferenceSpectra}}{Indicates whether this database shall be used to use retention time information for library matching.  Typically used for in-house spectral libraries that were measured on the same  chromatographic setup as the samples being analyzed, so that retention times are comparable. character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SearchableDatabaseParameters-new}{\code{SearchableDatabaseParameters$new()}}
\item \href{#method-SearchableDatabaseParameters-toJSON}{\code{SearchableDatabaseParameters$toJSON()}}
\item \href{#method-SearchableDatabaseParameters-toList}{\code{SearchableDatabaseParameters$toList()}}
\item \href{#method-SearchableDatabaseParameters-toSimpleType}{\code{SearchableDatabaseParameters$toSimpleType()}}
\item \href{#method-SearchableDatabaseParameters-fromJSON}{\code{SearchableDatabaseParameters$fromJSON()}}
\item \href{#method-SearchableDatabaseParameters-toJSONString}{\code{SearchableDatabaseParameters$toJSONString()}}
\item \href{#method-SearchableDatabaseParameters-fromJSONString}{\code{SearchableDatabaseParameters$fromJSONString()}}
\item \href{#method-SearchableDatabaseParameters-validateJSON}{\code{SearchableDatabaseParameters$validateJSON()}}
\item \href{#method-SearchableDatabaseParameters-toString}{\code{SearchableDatabaseParameters$toString()}}
\item \href{#method-SearchableDatabaseParameters-isValid}{\code{SearchableDatabaseParameters$isValid()}}
\item \href{#method-SearchableDatabaseParameters-getInvalidFields}{\code{SearchableDatabaseParameters$getInvalidFields()}}
\item \href{#method-SearchableDatabaseParameters-print}{\code{SearchableDatabaseParameters$print()}}
\item \href{#method-SearchableDatabaseParameters-clone}{\code{SearchableDatabaseParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabaseParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabaseParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SearchableDatabaseParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabaseParameters$new(
  displayName = NULL,
  location = NULL,
  matchRtOfReferenceSpectra = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{displayName}}{display name of the database  Should be short}

\item{\code{location}}{Storage location of user database  Might be NULL for non-user databases or if default location is used.}

\item{\code{matchRtOfReferenceSpectra}}{Indicates whether this database shall be used to use retention time information for library matching.  Typically used for in-house spectral libraries that were measured on the same  chromatographic setup as the samples being analyzed, so that retention times are comparable.. Default to FALSE.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabaseParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabaseParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabaseParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabaseParameters-toList"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabaseParameters-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabaseParameters$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SearchableDatabaseParameters as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of SearchableDatabaseParameters (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabaseParameters-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabaseParameters-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert SearchableDatabaseParameters to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabaseParameters$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabaseParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabaseParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of SearchableDatabaseParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabaseParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SearchableDatabaseParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabaseParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabaseParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabaseParameters$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SearchableDatabaseParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabaseParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabaseParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of SearchableDatabaseParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabaseParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SearchableDatabaseParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabaseParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabaseParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to SearchableDatabaseParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabaseParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabaseParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabaseParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabaseParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of SearchableDatabaseParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabaseParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabaseParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabaseParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabaseParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabaseParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabaseParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabaseParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabaseParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabaseParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabaseParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabaseParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabaseParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
