% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchable_field_type.R
\docType{class}
\name{SearchableFieldType}
\alias{SearchableFieldType}
\title{SearchableFieldType}
\format{
An \code{R6Class} generator object
}
\description{
SearchableFieldType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SearchableFieldType-new}{\code{SearchableFieldType$new()}}
\item \href{#method-SearchableFieldType-toJSON}{\code{SearchableFieldType$toJSON()}}
\item \href{#method-SearchableFieldType-fromJSON}{\code{SearchableFieldType$fromJSON()}}
\item \href{#method-SearchableFieldType-toJSONString}{\code{SearchableFieldType$toJSONString()}}
\item \href{#method-SearchableFieldType-fromJSONString}{\code{SearchableFieldType$fromJSONString()}}
\item \href{#method-SearchableFieldType-clone}{\code{SearchableFieldType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableFieldType-new"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableFieldType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SearchableFieldType class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableFieldType$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableFieldType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableFieldType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableFieldType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SearchableFieldType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableFieldType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableFieldType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of SearchableFieldType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableFieldType$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SearchableFieldType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableFieldType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableFieldType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableFieldType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SearchableFieldType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableFieldType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableFieldType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of SearchableFieldType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableFieldType$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SearchableFieldType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableFieldType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableFieldType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableFieldType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
