% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_context.R
\docType{class}
\name{SecurityContext}
\alias{SecurityContext}
\title{SecurityContext}
\format{
An \code{R6Class} generator object
}
\description{
SecurityContext Class

SecurityContext Class
}
\details{
Create a new SecurityContext
}
\examples{

## ------------------------------------------------
## Method `SecurityContext$toList`
## ------------------------------------------------

# convert array of SecurityContext (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{principal}}{\link{SecurityContextPrincipal} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SecurityContext-new}{\code{SecurityContext$new()}}
\item \href{#method-SecurityContext-toJSON}{\code{SecurityContext$toJSON()}}
\item \href{#method-SecurityContext-toList}{\code{SecurityContext$toList()}}
\item \href{#method-SecurityContext-toSimpleType}{\code{SecurityContext$toSimpleType()}}
\item \href{#method-SecurityContext-fromJSON}{\code{SecurityContext$fromJSON()}}
\item \href{#method-SecurityContext-toJSONString}{\code{SecurityContext$toJSONString()}}
\item \href{#method-SecurityContext-fromJSONString}{\code{SecurityContext$fromJSONString()}}
\item \href{#method-SecurityContext-validateJSON}{\code{SecurityContext$validateJSON()}}
\item \href{#method-SecurityContext-toString}{\code{SecurityContext$toString()}}
\item \href{#method-SecurityContext-isValid}{\code{SecurityContext$isValid()}}
\item \href{#method-SecurityContext-getInvalidFields}{\code{SecurityContext$getInvalidFields()}}
\item \href{#method-SecurityContext-print}{\code{SecurityContext$print()}}
\item \href{#method-SecurityContext-clone}{\code{SecurityContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContext-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SecurityContext class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContext$new(principal = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{principal}}{principal}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContext-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContext-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContext$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContext-toList"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContext-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContext$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SecurityContext as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of SecurityContext (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContext-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContext-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert SecurityContext to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContext$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContext-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContext-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of SecurityContext
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContext$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SecurityContext
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContext-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContext-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContext$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SecurityContext in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContext-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContext-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of SecurityContext
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContext$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SecurityContext
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContext-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContext-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to SecurityContext and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContext$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContext-toString"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContext-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContext$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of SecurityContext
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContext-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContext-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContext$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContext-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContext-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContext$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContext-print"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContext-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContext$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
