% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_context_principal.R
\docType{class}
\name{SecurityContextPrincipal}
\alias{SecurityContextPrincipal}
\title{SecurityContextPrincipal}
\format{
An \code{R6Class} generator object
}
\description{
SecurityContextPrincipal Class

SecurityContextPrincipal Class
}
\details{
Create a new SecurityContextPrincipal
}
\examples{

## ------------------------------------------------
## Method `SecurityContextPrincipal$toList`
## ------------------------------------------------

# convert array of SecurityContextPrincipal (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SecurityContextPrincipal-new}{\code{SecurityContextPrincipal$new()}}
\item \href{#method-SecurityContextPrincipal-toJSON}{\code{SecurityContextPrincipal$toJSON()}}
\item \href{#method-SecurityContextPrincipal-toList}{\code{SecurityContextPrincipal$toList()}}
\item \href{#method-SecurityContextPrincipal-toSimpleType}{\code{SecurityContextPrincipal$toSimpleType()}}
\item \href{#method-SecurityContextPrincipal-fromJSON}{\code{SecurityContextPrincipal$fromJSON()}}
\item \href{#method-SecurityContextPrincipal-toJSONString}{\code{SecurityContextPrincipal$toJSONString()}}
\item \href{#method-SecurityContextPrincipal-fromJSONString}{\code{SecurityContextPrincipal$fromJSONString()}}
\item \href{#method-SecurityContextPrincipal-validateJSON}{\code{SecurityContextPrincipal$validateJSON()}}
\item \href{#method-SecurityContextPrincipal-toString}{\code{SecurityContextPrincipal$toString()}}
\item \href{#method-SecurityContextPrincipal-isValid}{\code{SecurityContextPrincipal$isValid()}}
\item \href{#method-SecurityContextPrincipal-getInvalidFields}{\code{SecurityContextPrincipal$getInvalidFields()}}
\item \href{#method-SecurityContextPrincipal-print}{\code{SecurityContextPrincipal$print()}}
\item \href{#method-SecurityContextPrincipal-clone}{\code{SecurityContextPrincipal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContextPrincipal-new"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContextPrincipal-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SecurityContextPrincipal class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContextPrincipal$new(name = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContextPrincipal-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContextPrincipal-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContextPrincipal$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContextPrincipal-toList"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContextPrincipal-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContextPrincipal$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SecurityContextPrincipal as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of SecurityContextPrincipal (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContextPrincipal-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContextPrincipal-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert SecurityContextPrincipal to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContextPrincipal$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContextPrincipal-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContextPrincipal-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of SecurityContextPrincipal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContextPrincipal$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SecurityContextPrincipal
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContextPrincipal-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContextPrincipal-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContextPrincipal$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SecurityContextPrincipal in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContextPrincipal-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContextPrincipal-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of SecurityContextPrincipal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContextPrincipal$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SecurityContextPrincipal
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContextPrincipal-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContextPrincipal-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to SecurityContextPrincipal and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContextPrincipal$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContextPrincipal-toString"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContextPrincipal-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContextPrincipal$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of SecurityContextPrincipal
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContextPrincipal-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContextPrincipal-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContextPrincipal$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContextPrincipal-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContextPrincipal-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContextPrincipal$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContextPrincipal-print"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContextPrincipal-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContextPrincipal$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityContextPrincipal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityContextPrincipal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityContextPrincipal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
