% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_object.R
\docType{class}
\name{SortObject}
\alias{SortObject}
\title{SortObject}
\format{
An \code{R6Class} generator object
}
\description{
SortObject Class

SortObject Class
}
\details{
Create a new SortObject
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{empty}}{character [optional]}

\item{\code{sorted}}{character [optional]}

\item{\code{unsorted}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SortObject-new}{\code{SortObject$new()}}
\item \href{#method-SortObject-toJSON}{\code{SortObject$toJSON()}}
\item \href{#method-SortObject-fromJSON}{\code{SortObject$fromJSON()}}
\item \href{#method-SortObject-toJSONString}{\code{SortObject$toJSONString()}}
\item \href{#method-SortObject-fromJSONString}{\code{SortObject$fromJSONString()}}
\item \href{#method-SortObject-validateJSON}{\code{SortObject$validateJSON()}}
\item \href{#method-SortObject-toString}{\code{SortObject$toString()}}
\item \href{#method-SortObject-isValid}{\code{SortObject$isValid()}}
\item \href{#method-SortObject-getInvalidFields}{\code{SortObject$getInvalidFields()}}
\item \href{#method-SortObject-print}{\code{SortObject$print()}}
\item \href{#method-SortObject-clone}{\code{SortObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SortObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-SortObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SortObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SortObject$new(empty = NULL, sorted = NULL, unsorted = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{empty}}{empty}

\item{\code{sorted}}{sorted}

\item{\code{unsorted}}{unsorted}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SortObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SortObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SortObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SortObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SortObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SortObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of SortObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SortObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SortObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SortObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SortObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SortObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SortObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SortObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SortObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of SortObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SortObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SortObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SortObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SortObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to SortObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SortObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SortObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-SortObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SortObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of SortObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SortObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-SortObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SortObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SortObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-SortObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SortObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SortObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-SortObject-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SortObject$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SortObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SortObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SortObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
