% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_library_match_opt_field.R
\docType{class}
\name{SpectralLibraryMatchOptField}
\alias{SpectralLibraryMatchOptField}
\title{SpectralLibraryMatchOptField}
\format{
An \code{R6Class} generator object
}
\description{
SpectralLibraryMatchOptField Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SpectralLibraryMatchOptField-new}{\code{SpectralLibraryMatchOptField$new()}}
\item \href{#method-SpectralLibraryMatchOptField-toJSON}{\code{SpectralLibraryMatchOptField$toJSON()}}
\item \href{#method-SpectralLibraryMatchOptField-fromJSON}{\code{SpectralLibraryMatchOptField$fromJSON()}}
\item \href{#method-SpectralLibraryMatchOptField-toJSONString}{\code{SpectralLibraryMatchOptField$toJSONString()}}
\item \href{#method-SpectralLibraryMatchOptField-fromJSONString}{\code{SpectralLibraryMatchOptField$fromJSONString()}}
\item \href{#method-SpectralLibraryMatchOptField-clone}{\code{SpectralLibraryMatchOptField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectralLibraryMatchOptField-new"></a>}}
\if{latex}{\out{\hypertarget{method-SpectralLibraryMatchOptField-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SpectralLibraryMatchOptField class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectralLibraryMatchOptField$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectralLibraryMatchOptField-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SpectralLibraryMatchOptField-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectralLibraryMatchOptField$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SpectralLibraryMatchOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectralLibraryMatchOptField-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SpectralLibraryMatchOptField-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of SpectralLibraryMatchOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectralLibraryMatchOptField$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SpectralLibraryMatchOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectralLibraryMatchOptField-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SpectralLibraryMatchOptField-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectralLibraryMatchOptField$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SpectralLibraryMatchOptField in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectralLibraryMatchOptField-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SpectralLibraryMatchOptField-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of SpectralLibraryMatchOptField
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectralLibraryMatchOptField$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SpectralLibraryMatchOptField
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectralLibraryMatchOptField-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpectralLibraryMatchOptField-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectralLibraryMatchOptField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
