% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_matching_type.R
\docType{class}
\name{SpectralMatchingType}
\alias{SpectralMatchingType}
\title{SpectralMatchingType}
\format{
An \code{R6Class} generator object
}
\description{
SpectralMatchingType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SpectralMatchingType-new}{\code{SpectralMatchingType$new()}}
\item \href{#method-SpectralMatchingType-toJSON}{\code{SpectralMatchingType$toJSON()}}
\item \href{#method-SpectralMatchingType-fromJSON}{\code{SpectralMatchingType$fromJSON()}}
\item \href{#method-SpectralMatchingType-toJSONString}{\code{SpectralMatchingType$toJSONString()}}
\item \href{#method-SpectralMatchingType-fromJSONString}{\code{SpectralMatchingType$fromJSONString()}}
\item \href{#method-SpectralMatchingType-clone}{\code{SpectralMatchingType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectralMatchingType-new"></a>}}
\if{latex}{\out{\hypertarget{method-SpectralMatchingType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SpectralMatchingType class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectralMatchingType$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectralMatchingType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SpectralMatchingType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectralMatchingType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SpectralMatchingType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectralMatchingType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SpectralMatchingType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of SpectralMatchingType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectralMatchingType$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SpectralMatchingType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectralMatchingType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SpectralMatchingType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectralMatchingType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SpectralMatchingType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectralMatchingType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SpectralMatchingType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of SpectralMatchingType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectralMatchingType$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SpectralMatchingType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectralMatchingType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpectralMatchingType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectralMatchingType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
