% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_table.R
\docType{class}
\name{StatisticsTable}
\alias{StatisticsTable}
\title{StatisticsTable}
\format{
An \code{R6Class} generator object
}
\description{
StatisticsTable Class

StatisticsTable Class
}
\details{
Create a new StatisticsTable
}
\examples{

## ------------------------------------------------
## Method `StatisticsTable$toList`
## ------------------------------------------------

# convert array of StatisticsTable (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{statisticsType}}{character [optional]}

\item{\code{aggregationType}}{character [optional]}

\item{\code{quantificationMeasure}}{character [optional]}

\item{\code{rowType}}{character [optional]}

\item{\code{rowIds}}{list(character) [optional]}

\item{\code{columnNames}}{list(character) [optional]}

\item{\code{columnLeftGroups}}{list(character) [optional]}

\item{\code{columnRightGroups}}{list(character) [optional]}

\item{\code{values}}{list(list(numeric)) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StatisticsTable-new}{\code{StatisticsTable$new()}}
\item \href{#method-StatisticsTable-toJSON}{\code{StatisticsTable$toJSON()}}
\item \href{#method-StatisticsTable-toList}{\code{StatisticsTable$toList()}}
\item \href{#method-StatisticsTable-toSimpleType}{\code{StatisticsTable$toSimpleType()}}
\item \href{#method-StatisticsTable-fromJSON}{\code{StatisticsTable$fromJSON()}}
\item \href{#method-StatisticsTable-toJSONString}{\code{StatisticsTable$toJSONString()}}
\item \href{#method-StatisticsTable-fromJSONString}{\code{StatisticsTable$fromJSONString()}}
\item \href{#method-StatisticsTable-validateJSON}{\code{StatisticsTable$validateJSON()}}
\item \href{#method-StatisticsTable-toString}{\code{StatisticsTable$toString()}}
\item \href{#method-StatisticsTable-isValid}{\code{StatisticsTable$isValid()}}
\item \href{#method-StatisticsTable-getInvalidFields}{\code{StatisticsTable$getInvalidFields()}}
\item \href{#method-StatisticsTable-print}{\code{StatisticsTable$print()}}
\item \href{#method-StatisticsTable-clone}{\code{StatisticsTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticsTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticsTable-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new StatisticsTable class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticsTable$new(
  statisticsType = NULL,
  aggregationType = NULL,
  quantificationMeasure = NULL,
  rowType = NULL,
  rowIds = NULL,
  columnNames = NULL,
  columnLeftGroups = NULL,
  columnRightGroups = NULL,
  values = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statisticsType}}{statisticsType}

\item{\code{aggregationType}}{aggregationType}

\item{\code{quantificationMeasure}}{quantificationMeasure}

\item{\code{rowType}}{rowType}

\item{\code{rowIds}}{rowIds}

\item{\code{columnNames}}{columnNames}

\item{\code{columnLeftGroups}}{columnLeftGroups}

\item{\code{columnRightGroups}}{columnRightGroups}

\item{\code{values}}{values}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticsTable-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticsTable-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticsTable$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticsTable-toList"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticsTable-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticsTable$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StatisticsTable as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of StatisticsTable (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticsTable-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticsTable-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert StatisticsTable to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticsTable$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticsTable-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticsTable-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of StatisticsTable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticsTable$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StatisticsTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticsTable-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticsTable-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticsTable$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
StatisticsTable in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticsTable-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticsTable-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of StatisticsTable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticsTable$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StatisticsTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticsTable-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticsTable-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to StatisticsTable and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticsTable$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticsTable-toString"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticsTable-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticsTable$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of StatisticsTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticsTable-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticsTable-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticsTable$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticsTable-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticsTable-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticsTable$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticsTable-print"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticsTable-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticsTable$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticsTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticsTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticsTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
