% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored_job_submission.R
\docType{class}
\name{StoredJobSubmission}
\alias{StoredJobSubmission}
\title{StoredJobSubmission}
\format{
An \code{R6Class} generator object
}
\description{
StoredJobSubmission Class

StoredJobSubmission Class
}
\details{
Create a new StoredJobSubmission
}
\examples{

## ------------------------------------------------
## Method `StoredJobSubmission$toList`
## ------------------------------------------------

# convert array of StoredJobSubmission (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Unique name to identify this JobSubmission (job config). character}

\item{\code{editable}}{False for predefined configs which are not editable and not removable. character}

\item{\code{jobSubmission}}{The JobSubmission identified by the name \link{JobSubmission}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StoredJobSubmission-new}{\code{StoredJobSubmission$new()}}
\item \href{#method-StoredJobSubmission-toJSON}{\code{StoredJobSubmission$toJSON()}}
\item \href{#method-StoredJobSubmission-toList}{\code{StoredJobSubmission$toList()}}
\item \href{#method-StoredJobSubmission-toSimpleType}{\code{StoredJobSubmission$toSimpleType()}}
\item \href{#method-StoredJobSubmission-fromJSON}{\code{StoredJobSubmission$fromJSON()}}
\item \href{#method-StoredJobSubmission-toJSONString}{\code{StoredJobSubmission$toJSONString()}}
\item \href{#method-StoredJobSubmission-fromJSONString}{\code{StoredJobSubmission$fromJSONString()}}
\item \href{#method-StoredJobSubmission-validateJSON}{\code{StoredJobSubmission$validateJSON()}}
\item \href{#method-StoredJobSubmission-toString}{\code{StoredJobSubmission$toString()}}
\item \href{#method-StoredJobSubmission-isValid}{\code{StoredJobSubmission$isValid()}}
\item \href{#method-StoredJobSubmission-getInvalidFields}{\code{StoredJobSubmission$getInvalidFields()}}
\item \href{#method-StoredJobSubmission-print}{\code{StoredJobSubmission$print()}}
\item \href{#method-StoredJobSubmission-clone}{\code{StoredJobSubmission$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoredJobSubmission-new"></a>}}
\if{latex}{\out{\hypertarget{method-StoredJobSubmission-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new StoredJobSubmission class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoredJobSubmission$new(name, editable, jobSubmission, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Unique name to identify this JobSubmission (job config).}

\item{\code{editable}}{False for predefined configs which are not editable and not removable.}

\item{\code{jobSubmission}}{The JobSubmission identified by the name}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoredJobSubmission-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StoredJobSubmission-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoredJobSubmission$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoredJobSubmission-toList"></a>}}
\if{latex}{\out{\hypertarget{method-StoredJobSubmission-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoredJobSubmission$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StoredJobSubmission as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of StoredJobSubmission (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoredJobSubmission-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-StoredJobSubmission-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert StoredJobSubmission to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoredJobSubmission$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoredJobSubmission-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StoredJobSubmission-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of StoredJobSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoredJobSubmission$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StoredJobSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoredJobSubmission-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StoredJobSubmission-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoredJobSubmission$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
StoredJobSubmission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoredJobSubmission-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StoredJobSubmission-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of StoredJobSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoredJobSubmission$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StoredJobSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoredJobSubmission-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StoredJobSubmission-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to StoredJobSubmission and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoredJobSubmission$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoredJobSubmission-toString"></a>}}
\if{latex}{\out{\hypertarget{method-StoredJobSubmission-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoredJobSubmission$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of StoredJobSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoredJobSubmission-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-StoredJobSubmission-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoredJobSubmission$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoredJobSubmission-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-StoredJobSubmission-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoredJobSubmission$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoredJobSubmission-print"></a>}}
\if{latex}{\out{\hypertarget{method-StoredJobSubmission-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoredJobSubmission$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoredJobSubmission-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StoredJobSubmission-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoredJobSubmission$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
