% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_candidate_scored.R
\docType{class}
\name{StructureCandidateScored}
\alias{StructureCandidateScored}
\title{StructureCandidateScored}
\format{
An \code{R6Class} generator object
}
\description{
StructureCandidateScored Class

StructureCandidateScored Class
}
\details{
Create a new StructureCandidateScored
}
\examples{

## ------------------------------------------------
## Method `StructureCandidateScored$toList`
## ------------------------------------------------

# convert array of StructureCandidateScored (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{inchiKey}}{InChIKey of the structure candidate.  Searching with a full 27 character key matches its 2D (skeleton) part, so stereoisomers of the  searched structure are found alike. character [optional]}

\item{\code{smiles}}{character [optional]}

\item{\code{structureName}}{Name of the structure candidate.  Searching also resolves the searched term as a PubChem synonym, so a structure is found by any  of its common names. character [optional]}

\item{\code{structureSvg}}{SVG graphics of the structure candidate  OPTIONAL: needs to be added by parameter character [optional]}

\item{\code{dbLinks}}{List of structure database links belonging to the structure candidate  OPTIONAL: needs to be added by parameter list(\link{DBLink}) [optional]}

\item{\code{spectralLibraryMatches}}{List of spectral library matches belonging to the structure candidate  OPTIONAL: needs to be added by parameter list(\link{SpectralLibraryMatch}) [optional]}

\item{\code{xlogP}}{numeric [optional]}

\item{\code{rank}}{The overall rank of this candidate among all candidates of this feature. integer [optional]}

\item{\code{csiScore}}{CSI:FingerID score of the fingerprint of this compound to the predicted fingerprint of CSI:FingerID  This is the score used for ranking structure candidates numeric [optional]}

\item{\code{tanimotoSimilarity}}{Tanimoto similarity of the fingerprint of this compound to the predicted fingerprint of CSI:FingerID numeric [optional]}

\item{\code{mcesDistToTopHit}}{Maximum Common Edge Subgraph (MCES) distance to the top scoring hit (CSI:FingerID) in a candidate list. numeric [optional]}

\item{\code{fingerprint}}{\link{BinaryFingerprint} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StructureCandidateScored-new}{\code{StructureCandidateScored$new()}}
\item \href{#method-StructureCandidateScored-toJSON}{\code{StructureCandidateScored$toJSON()}}
\item \href{#method-StructureCandidateScored-toList}{\code{StructureCandidateScored$toList()}}
\item \href{#method-StructureCandidateScored-toSimpleType}{\code{StructureCandidateScored$toSimpleType()}}
\item \href{#method-StructureCandidateScored-fromJSON}{\code{StructureCandidateScored$fromJSON()}}
\item \href{#method-StructureCandidateScored-toJSONString}{\code{StructureCandidateScored$toJSONString()}}
\item \href{#method-StructureCandidateScored-fromJSONString}{\code{StructureCandidateScored$fromJSONString()}}
\item \href{#method-StructureCandidateScored-validateJSON}{\code{StructureCandidateScored$validateJSON()}}
\item \href{#method-StructureCandidateScored-toString}{\code{StructureCandidateScored$toString()}}
\item \href{#method-StructureCandidateScored-isValid}{\code{StructureCandidateScored$isValid()}}
\item \href{#method-StructureCandidateScored-getInvalidFields}{\code{StructureCandidateScored$getInvalidFields()}}
\item \href{#method-StructureCandidateScored-print}{\code{StructureCandidateScored$print()}}
\item \href{#method-StructureCandidateScored-clone}{\code{StructureCandidateScored$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateScored-new"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateScored-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new StructureCandidateScored class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateScored$new(
  inchiKey = NULL,
  smiles = NULL,
  structureName = NULL,
  structureSvg = NULL,
  dbLinks = NULL,
  spectralLibraryMatches = NULL,
  xlogP = NULL,
  rank = NULL,
  csiScore = NULL,
  tanimotoSimilarity = NULL,
  mcesDistToTopHit = NULL,
  fingerprint = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inchiKey}}{InChIKey of the structure candidate.  Searching with a full 27 character key matches its 2D (skeleton) part, so stereoisomers of the  searched structure are found alike.}

\item{\code{smiles}}{smiles}

\item{\code{structureName}}{Name of the structure candidate.  Searching also resolves the searched term as a PubChem synonym, so a structure is found by any  of its common names.}

\item{\code{structureSvg}}{SVG graphics of the structure candidate  OPTIONAL: needs to be added by parameter}

\item{\code{dbLinks}}{List of structure database links belonging to the structure candidate  OPTIONAL: needs to be added by parameter}

\item{\code{spectralLibraryMatches}}{List of spectral library matches belonging to the structure candidate  OPTIONAL: needs to be added by parameter}

\item{\code{xlogP}}{xlogP}

\item{\code{rank}}{The overall rank of this candidate among all candidates of this feature.}

\item{\code{csiScore}}{CSI:FingerID score of the fingerprint of this compound to the predicted fingerprint of CSI:FingerID  This is the score used for ranking structure candidates}

\item{\code{tanimotoSimilarity}}{Tanimoto similarity of the fingerprint of this compound to the predicted fingerprint of CSI:FingerID}

\item{\code{mcesDistToTopHit}}{Maximum Common Edge Subgraph (MCES) distance to the top scoring hit (CSI:FingerID) in a candidate list.}

\item{\code{fingerprint}}{fingerprint}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateScored-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateScored-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateScored$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateScored-toList"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateScored-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateScored$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StructureCandidateScored as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of StructureCandidateScored (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateScored-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateScored-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert StructureCandidateScored to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateScored$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateScored-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateScored-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of StructureCandidateScored
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateScored$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StructureCandidateScored
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateScored-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateScored-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateScored$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
StructureCandidateScored in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateScored-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateScored-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of StructureCandidateScored
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateScored$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StructureCandidateScored
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateScored-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateScored-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to StructureCandidateScored and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateScored$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateScored-toString"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateScored-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateScored$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of StructureCandidateScored
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateScored-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateScored-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateScored$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateScored-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateScored-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateScored$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateScored-print"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateScored-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateScored$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateScored-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateScored-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateScored$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
