% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_group.R
\docType{class}
\name{TagGroup}
\alias{TagGroup}
\title{TagGroup}
\format{
An \code{R6Class} generator object
}
\description{
TagGroup Class

TagGroup Class
}
\details{
Create a new TagGroup
}
\examples{

## ------------------------------------------------
## Method `TagGroup$toList`
## ------------------------------------------------

# convert array of TagGroup (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{groupName}}{Name of this Grouping query. character [optional]}

\item{\code{luceneQuery}}{Query used to group the entities in lucene format. character [optional]}

\item{\code{groupType}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TagGroup-new}{\code{TagGroup$new()}}
\item \href{#method-TagGroup-toJSON}{\code{TagGroup$toJSON()}}
\item \href{#method-TagGroup-toList}{\code{TagGroup$toList()}}
\item \href{#method-TagGroup-toSimpleType}{\code{TagGroup$toSimpleType()}}
\item \href{#method-TagGroup-fromJSON}{\code{TagGroup$fromJSON()}}
\item \href{#method-TagGroup-toJSONString}{\code{TagGroup$toJSONString()}}
\item \href{#method-TagGroup-fromJSONString}{\code{TagGroup$fromJSONString()}}
\item \href{#method-TagGroup-validateJSON}{\code{TagGroup$validateJSON()}}
\item \href{#method-TagGroup-toString}{\code{TagGroup$toString()}}
\item \href{#method-TagGroup-isValid}{\code{TagGroup$isValid()}}
\item \href{#method-TagGroup-getInvalidFields}{\code{TagGroup$getInvalidFields()}}
\item \href{#method-TagGroup-print}{\code{TagGroup$print()}}
\item \href{#method-TagGroup-clone}{\code{TagGroup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagGroup-new"></a>}}
\if{latex}{\out{\hypertarget{method-TagGroup-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TagGroup class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagGroup$new(groupName = NULL, luceneQuery = NULL, groupType = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groupName}}{Name of this Grouping query.}

\item{\code{luceneQuery}}{Query used to group the entities in lucene format.}

\item{\code{groupType}}{groupType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagGroup-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TagGroup-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagGroup$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagGroup-toList"></a>}}
\if{latex}{\out{\hypertarget{method-TagGroup-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagGroup$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TagGroup as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of TagGroup (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagGroup-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-TagGroup-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert TagGroup to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagGroup$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagGroup-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TagGroup-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TagGroup
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagGroup$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TagGroup
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagGroup-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TagGroup-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagGroup$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TagGroup in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagGroup-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TagGroup-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TagGroup
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagGroup$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TagGroup
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagGroup-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TagGroup-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TagGroup and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagGroup$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagGroup-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TagGroup-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagGroup$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TagGroup
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagGroup-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TagGroup-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagGroup$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagGroup-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TagGroup-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagGroup$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagGroup-print"></a>}}
\if{latex}{\out{\hypertarget{method-TagGroup-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagGroup$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagGroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TagGroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
