% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_submission.R
\docType{class}
\name{TagSubmission}
\alias{TagSubmission}
\title{TagSubmission}
\format{
An \code{R6Class} generator object
}
\description{
TagSubmission Class

TagSubmission Class
}
\details{
Create a new TagSubmission
}
\examples{

## ------------------------------------------------
## Method `TagSubmission$toList`
## ------------------------------------------------

# convert array of TagSubmission (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tagName}}{Name of the tag as defined by the corresponding TagDefinition  Links tag object to their definition. character}

\item{\code{value}}{Optional value of the tag.  <p>  Generic value of the tag as defined by the corresponding TagDefinition.  Can be Integer, Double, Boolean and String, whereas String values can represent Text, Date (yyyy-MM-dd) or Time (HH:mm:ss). object [optional]}

\item{\code{taggedObjectId}}{ID of the object where the tag should be added.  Then Object type is taken from context of the API endpoint. character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TagSubmission-new}{\code{TagSubmission$new()}}
\item \href{#method-TagSubmission-toJSON}{\code{TagSubmission$toJSON()}}
\item \href{#method-TagSubmission-toList}{\code{TagSubmission$toList()}}
\item \href{#method-TagSubmission-toSimpleType}{\code{TagSubmission$toSimpleType()}}
\item \href{#method-TagSubmission-fromJSON}{\code{TagSubmission$fromJSON()}}
\item \href{#method-TagSubmission-toJSONString}{\code{TagSubmission$toJSONString()}}
\item \href{#method-TagSubmission-fromJSONString}{\code{TagSubmission$fromJSONString()}}
\item \href{#method-TagSubmission-validateJSON}{\code{TagSubmission$validateJSON()}}
\item \href{#method-TagSubmission-toString}{\code{TagSubmission$toString()}}
\item \href{#method-TagSubmission-isValid}{\code{TagSubmission$isValid()}}
\item \href{#method-TagSubmission-getInvalidFields}{\code{TagSubmission$getInvalidFields()}}
\item \href{#method-TagSubmission-print}{\code{TagSubmission$print()}}
\item \href{#method-TagSubmission-clone}{\code{TagSubmission$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagSubmission-new"></a>}}
\if{latex}{\out{\hypertarget{method-TagSubmission-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TagSubmission class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagSubmission$new(tagName, taggedObjectId, value = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tagName}}{Name of the tag as defined by the corresponding TagDefinition  Links tag object to their definition.}

\item{\code{taggedObjectId}}{ID of the object where the tag should be added.  Then Object type is taken from context of the API endpoint.}

\item{\code{value}}{Optional value of the tag.  <p>  Generic value of the tag as defined by the corresponding TagDefinition.  Can be Integer, Double, Boolean and String, whereas String values can represent Text, Date (yyyy-MM-dd) or Time (HH:mm:ss).}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagSubmission-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TagSubmission-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagSubmission$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagSubmission-toList"></a>}}
\if{latex}{\out{\hypertarget{method-TagSubmission-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagSubmission$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TagSubmission as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of TagSubmission (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagSubmission-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-TagSubmission-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert TagSubmission to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagSubmission$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagSubmission-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TagSubmission-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TagSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagSubmission$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TagSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagSubmission-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TagSubmission-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagSubmission$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TagSubmission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagSubmission-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TagSubmission-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TagSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagSubmission$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TagSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagSubmission-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TagSubmission-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TagSubmission and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagSubmission$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagSubmission-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TagSubmission-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagSubmission$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TagSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagSubmission-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TagSubmission-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagSubmission$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagSubmission-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TagSubmission-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagSubmission$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagSubmission-print"></a>}}
\if{latex}{\out{\hypertarget{method-TagSubmission-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagSubmission$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagSubmission-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TagSubmission-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagSubmission$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
