% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\docType{class}
\name{Term}
\alias{Term}
\title{Term}
\format{
An \code{R6Class} generator object
}
\description{
Term Class

Term Class
}
\details{
Create a new Term
}
\examples{

## ------------------------------------------------
## Method `Term$toList`
## ------------------------------------------------

# convert array of Term (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{link}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Term-new}{\code{Term$new()}}
\item \href{#method-Term-toJSON}{\code{Term$toJSON()}}
\item \href{#method-Term-toList}{\code{Term$toList()}}
\item \href{#method-Term-toSimpleType}{\code{Term$toSimpleType()}}
\item \href{#method-Term-fromJSON}{\code{Term$fromJSON()}}
\item \href{#method-Term-toJSONString}{\code{Term$toJSONString()}}
\item \href{#method-Term-fromJSONString}{\code{Term$fromJSONString()}}
\item \href{#method-Term-validateJSON}{\code{Term$validateJSON()}}
\item \href{#method-Term-toString}{\code{Term$toString()}}
\item \href{#method-Term-isValid}{\code{Term$isValid()}}
\item \href{#method-Term-getInvalidFields}{\code{Term$getInvalidFields()}}
\item \href{#method-Term-print}{\code{Term$print()}}
\item \href{#method-Term-clone}{\code{Term$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Term-new"></a>}}
\if{latex}{\out{\hypertarget{method-Term-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Term class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Term$new(name = NULL, link = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{link}}{link}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Term-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Term-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Term$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Term-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Term-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Term$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Term as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of Term (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Term-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-Term-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert Term to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Term$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Term-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Term-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Term
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Term$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Term
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Term-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Term-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Term$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Term in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Term-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Term-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Term
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Term$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Term
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Term-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Term-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Term and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Term$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Term-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Term-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Term$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Term
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Term-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Term-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Term$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Term-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Term-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Term$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Term-print"></a>}}
\if{latex}{\out{\hypertarget{method-Term-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Term$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Term-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Term-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Term$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
