% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeout.R
\docType{class}
\name{Timeout}
\alias{Timeout}
\title{Timeout}
\format{
An \code{R6Class} generator object
}
\description{
Timeout Class

Timeout Class
}
\details{
Create a new Timeout
}
\examples{

## ------------------------------------------------
## Method `Timeout$toList`
## ------------------------------------------------

# convert array of Timeout (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{numberOfSecondsPerDecomposition}}{integer [optional]}

\item{\code{numberOfSecondsPerInstance}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Timeout-new}{\code{Timeout$new()}}
\item \href{#method-Timeout-toJSON}{\code{Timeout$toJSON()}}
\item \href{#method-Timeout-toList}{\code{Timeout$toList()}}
\item \href{#method-Timeout-toSimpleType}{\code{Timeout$toSimpleType()}}
\item \href{#method-Timeout-fromJSON}{\code{Timeout$fromJSON()}}
\item \href{#method-Timeout-toJSONString}{\code{Timeout$toJSONString()}}
\item \href{#method-Timeout-fromJSONString}{\code{Timeout$fromJSONString()}}
\item \href{#method-Timeout-validateJSON}{\code{Timeout$validateJSON()}}
\item \href{#method-Timeout-toString}{\code{Timeout$toString()}}
\item \href{#method-Timeout-isValid}{\code{Timeout$isValid()}}
\item \href{#method-Timeout-getInvalidFields}{\code{Timeout$getInvalidFields()}}
\item \href{#method-Timeout-print}{\code{Timeout$print()}}
\item \href{#method-Timeout-clone}{\code{Timeout$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timeout-new"></a>}}
\if{latex}{\out{\hypertarget{method-Timeout-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Timeout class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timeout$new(
  numberOfSecondsPerDecomposition = NULL,
  numberOfSecondsPerInstance = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numberOfSecondsPerDecomposition}}{numberOfSecondsPerDecomposition}

\item{\code{numberOfSecondsPerInstance}}{numberOfSecondsPerInstance}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timeout-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Timeout-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timeout$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timeout-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Timeout-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timeout$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Timeout as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of Timeout (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timeout-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-Timeout-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert Timeout to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timeout$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timeout-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Timeout-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Timeout
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timeout$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Timeout
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timeout-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Timeout-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timeout$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Timeout in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timeout-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Timeout-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Timeout
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timeout$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Timeout
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timeout-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Timeout-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Timeout and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timeout$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timeout-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Timeout-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timeout$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Timeout
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timeout-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Timeout-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timeout$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timeout-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Timeout-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timeout$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timeout-print"></a>}}
\if{latex}{\out{\hypertarget{method-Timeout-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timeout$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timeout-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Timeout-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timeout$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
