% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_annotation_experimental.R
\docType{class}
\name{TraceAnnotationExperimental}
\alias{TraceAnnotationExperimental}
\title{TraceAnnotationExperimental}
\format{
An \code{R6Class} generator object
}
\description{
EXPERIMENTAL: This schema is experimental and may be changed (or even removed) without notice until it is declared stable.

TraceAnnotationExperimental Class
}
\details{
Create a new TraceAnnotationExperimental
}
\examples{

## ------------------------------------------------
## Method `TraceAnnotationExperimental$toList`
## ------------------------------------------------

# convert array of TraceAnnotationExperimental (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{EXPERIMENTAL: This schema is experimental and may be changed (or even removed) without notice until it is declared stable. character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{index}}{integer [optional]}

\item{\code{from}}{integer [optional]}

\item{\code{to}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TraceAnnotationExperimental-new}{\code{TraceAnnotationExperimental$new()}}
\item \href{#method-TraceAnnotationExperimental-toJSON}{\code{TraceAnnotationExperimental$toJSON()}}
\item \href{#method-TraceAnnotationExperimental-toList}{\code{TraceAnnotationExperimental$toList()}}
\item \href{#method-TraceAnnotationExperimental-toSimpleType}{\code{TraceAnnotationExperimental$toSimpleType()}}
\item \href{#method-TraceAnnotationExperimental-fromJSON}{\code{TraceAnnotationExperimental$fromJSON()}}
\item \href{#method-TraceAnnotationExperimental-toJSONString}{\code{TraceAnnotationExperimental$toJSONString()}}
\item \href{#method-TraceAnnotationExperimental-fromJSONString}{\code{TraceAnnotationExperimental$fromJSONString()}}
\item \href{#method-TraceAnnotationExperimental-validateJSON}{\code{TraceAnnotationExperimental$validateJSON()}}
\item \href{#method-TraceAnnotationExperimental-toString}{\code{TraceAnnotationExperimental$toString()}}
\item \href{#method-TraceAnnotationExperimental-isValid}{\code{TraceAnnotationExperimental$isValid()}}
\item \href{#method-TraceAnnotationExperimental-getInvalidFields}{\code{TraceAnnotationExperimental$getInvalidFields()}}
\item \href{#method-TraceAnnotationExperimental-print}{\code{TraceAnnotationExperimental$print()}}
\item \href{#method-TraceAnnotationExperimental-clone}{\code{TraceAnnotationExperimental$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceAnnotationExperimental-new"></a>}}
\if{latex}{\out{\hypertarget{method-TraceAnnotationExperimental-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TraceAnnotationExperimental class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceAnnotationExperimental$new(
  type = NULL,
  description = NULL,
  index = NULL,
  from = NULL,
  to = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{EXPERIMENTAL: This schema is experimental and may be changed (or even removed) without notice until it is declared stable.}

\item{\code{description}}{description}

\item{\code{index}}{index}

\item{\code{from}}{from}

\item{\code{to}}{to}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceAnnotationExperimental-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TraceAnnotationExperimental-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceAnnotationExperimental$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceAnnotationExperimental-toList"></a>}}
\if{latex}{\out{\hypertarget{method-TraceAnnotationExperimental-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceAnnotationExperimental$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TraceAnnotationExperimental as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of TraceAnnotationExperimental (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceAnnotationExperimental-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-TraceAnnotationExperimental-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert TraceAnnotationExperimental to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceAnnotationExperimental$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceAnnotationExperimental-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TraceAnnotationExperimental-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TraceAnnotationExperimental
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceAnnotationExperimental$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TraceAnnotationExperimental
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceAnnotationExperimental-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TraceAnnotationExperimental-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceAnnotationExperimental$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TraceAnnotationExperimental in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceAnnotationExperimental-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TraceAnnotationExperimental-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TraceAnnotationExperimental
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceAnnotationExperimental$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TraceAnnotationExperimental
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceAnnotationExperimental-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TraceAnnotationExperimental-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TraceAnnotationExperimental and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceAnnotationExperimental$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceAnnotationExperimental-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TraceAnnotationExperimental-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceAnnotationExperimental$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TraceAnnotationExperimental
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceAnnotationExperimental-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TraceAnnotationExperimental-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceAnnotationExperimental$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceAnnotationExperimental-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TraceAnnotationExperimental-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceAnnotationExperimental$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceAnnotationExperimental-print"></a>}}
\if{latex}{\out{\hypertarget{method-TraceAnnotationExperimental-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceAnnotationExperimental$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceAnnotationExperimental-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TraceAnnotationExperimental-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceAnnotationExperimental$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
