% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_set.R
\docType{class}
\name{TraceSet}
\alias{TraceSet}
\title{TraceSet}
\format{
An \code{R6Class} generator object
}
\description{
TraceSet Class

TraceSet Class
}
\details{
Create a new TraceSet
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sampleId}}{integer [optional]}

\item{\code{sampleName}}{character [optional]}

\item{\code{axes}}{\link{Axes} [optional]}

\item{\code{traces}}{list(\link{Trace}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TraceSet-new}{\code{TraceSet$new()}}
\item \href{#method-TraceSet-toJSON}{\code{TraceSet$toJSON()}}
\item \href{#method-TraceSet-fromJSON}{\code{TraceSet$fromJSON()}}
\item \href{#method-TraceSet-toJSONString}{\code{TraceSet$toJSONString()}}
\item \href{#method-TraceSet-fromJSONString}{\code{TraceSet$fromJSONString()}}
\item \href{#method-TraceSet-validateJSON}{\code{TraceSet$validateJSON()}}
\item \href{#method-TraceSet-toString}{\code{TraceSet$toString()}}
\item \href{#method-TraceSet-isValid}{\code{TraceSet$isValid()}}
\item \href{#method-TraceSet-getInvalidFields}{\code{TraceSet$getInvalidFields()}}
\item \href{#method-TraceSet-print}{\code{TraceSet$print()}}
\item \href{#method-TraceSet-clone}{\code{TraceSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-TraceSet-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TraceSet class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceSet$new(
  sampleId = NULL,
  sampleName = NULL,
  axes = NULL,
  traces = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampleId}}{sampleId}

\item{\code{sampleName}}{sampleName}

\item{\code{axes}}{axes}

\item{\code{traces}}{traces}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceSet-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TraceSet-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceSet$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TraceSet in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceSet-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TraceSet-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TraceSet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceSet$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TraceSet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceSet-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TraceSet-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceSet$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TraceSet in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceSet-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TraceSet-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TraceSet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceSet$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TraceSet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceSet-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TraceSet-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TraceSet and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceSet$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceSet-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TraceSet-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceSet$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TraceSet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceSet-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TraceSet-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceSet$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceSet-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TraceSet-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceSet$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceSet-print"></a>}}
\if{latex}{\out{\hypertarget{method-TraceSet-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceSet$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraceSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TraceSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraceSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
