% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_heuristic.R
\docType{class}
\name{UseHeuristic}
\alias{UseHeuristic}
\title{UseHeuristic}
\format{
An \code{R6Class} generator object
}
\description{
UseHeuristic Class

UseHeuristic Class
}
\details{
Create a new UseHeuristic
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{useHeuristicAboveMz}}{integer [optional]}

\item{\code{useOnlyHeuristicAboveMz}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UseHeuristic-new}{\code{UseHeuristic$new()}}
\item \href{#method-UseHeuristic-toJSON}{\code{UseHeuristic$toJSON()}}
\item \href{#method-UseHeuristic-fromJSON}{\code{UseHeuristic$fromJSON()}}
\item \href{#method-UseHeuristic-toJSONString}{\code{UseHeuristic$toJSONString()}}
\item \href{#method-UseHeuristic-fromJSONString}{\code{UseHeuristic$fromJSONString()}}
\item \href{#method-UseHeuristic-validateJSON}{\code{UseHeuristic$validateJSON()}}
\item \href{#method-UseHeuristic-toString}{\code{UseHeuristic$toString()}}
\item \href{#method-UseHeuristic-isValid}{\code{UseHeuristic$isValid()}}
\item \href{#method-UseHeuristic-getInvalidFields}{\code{UseHeuristic$getInvalidFields()}}
\item \href{#method-UseHeuristic-print}{\code{UseHeuristic$print()}}
\item \href{#method-UseHeuristic-clone}{\code{UseHeuristic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseHeuristic-new"></a>}}
\if{latex}{\out{\hypertarget{method-UseHeuristic-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UseHeuristic class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseHeuristic$new(
  useHeuristicAboveMz = NULL,
  useOnlyHeuristicAboveMz = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{useHeuristicAboveMz}}{useHeuristicAboveMz}

\item{\code{useOnlyHeuristicAboveMz}}{useOnlyHeuristicAboveMz}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseHeuristic-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseHeuristic-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseHeuristic$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UseHeuristic in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseHeuristic-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseHeuristic-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UseHeuristic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseHeuristic$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UseHeuristic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseHeuristic-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UseHeuristic-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseHeuristic$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UseHeuristic in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseHeuristic-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UseHeuristic-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UseHeuristic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseHeuristic$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UseHeuristic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseHeuristic-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseHeuristic-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UseHeuristic and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseHeuristic$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseHeuristic-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UseHeuristic-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseHeuristic$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UseHeuristic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseHeuristic-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UseHeuristic-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseHeuristic$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseHeuristic-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UseHeuristic-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseHeuristic$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseHeuristic-print"></a>}}
\if{latex}{\out{\hypertarget{method-UseHeuristic-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseHeuristic$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseHeuristic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UseHeuristic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseHeuristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
