% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_info_controller_api.R
\docType{class}
\name{VersionInfoControllerApi}
\alias{VersionInfoControllerApi}
\title{VersionInfoController operations}
\format{
An \code{R6Class} generator object
}
\description{
VersionInfoControllerApi
}
\details{
SIRIUS Nightsky API

REST API that provides the full functionality of SIRIUS and its web services as background service. It is intended as entry-point for scripting languages and software integration SDKs.

The version of the OpenAPI document: v0.9 on SIRIUS 5.6.4-SNAPSHOT
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ GetVersionInfo } \emph{  }

\itemize{


\item status code : 200 | OK

\item return type : character
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  GetVersionInfo  ####################

library(Rsirius)

api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetVersionInfo(data_file = "result.txt")
result <- api_instance$version_info_controller_api$GetVersionInfo()
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VersionInfoControllerApi-new}{\code{VersionInfoControllerApi$new()}}
\item \href{#method-VersionInfoControllerApi-GetVersionInfo}{\code{VersionInfoControllerApi$GetVersionInfo()}}
\item \href{#method-VersionInfoControllerApi-GetVersionInfoWithHttpInfo}{\code{VersionInfoControllerApi$GetVersionInfoWithHttpInfo()}}
\item \href{#method-VersionInfoControllerApi-clone}{\code{VersionInfoControllerApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VersionInfoControllerApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-VersionInfoControllerApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new VersionInfoControllerApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionInfoControllerApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VersionInfoControllerApi-GetVersionInfo"></a>}}
\if{latex}{\out{\hypertarget{method-VersionInfoControllerApi-GetVersionInfo}{}}}
\subsection{Method \code{GetVersionInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionInfoControllerApi$GetVersionInfo(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
character
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VersionInfoControllerApi-GetVersionInfoWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-VersionInfoControllerApi-GetVersionInfoWithHttpInfo}{}}}
\subsection{Method \code{GetVersionInfoWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionInfoControllerApi$GetVersionInfoWithHttpInfo(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (character) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VersionInfoControllerApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VersionInfoControllerApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionInfoControllerApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
