% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_server_namespace.R
\docType{class}
\name{WebServerNamespace}
\alias{WebServerNamespace}
\title{WebServerNamespace}
\format{
An \code{R6Class} generator object
}
\description{
WebServerNamespace Class

WebServerNamespace Class
}
\details{
Create a new WebServerNamespace
}
\examples{

## ------------------------------------------------
## Method `WebServerNamespace$toList`
## ------------------------------------------------

# convert array of WebServerNamespace (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WebServerNamespace-new}{\code{WebServerNamespace$new()}}
\item \href{#method-WebServerNamespace-toJSON}{\code{WebServerNamespace$toJSON()}}
\item \href{#method-WebServerNamespace-toList}{\code{WebServerNamespace$toList()}}
\item \href{#method-WebServerNamespace-toSimpleType}{\code{WebServerNamespace$toSimpleType()}}
\item \href{#method-WebServerNamespace-fromJSON}{\code{WebServerNamespace$fromJSON()}}
\item \href{#method-WebServerNamespace-toJSONString}{\code{WebServerNamespace$toJSONString()}}
\item \href{#method-WebServerNamespace-fromJSONString}{\code{WebServerNamespace$fromJSONString()}}
\item \href{#method-WebServerNamespace-validateJSON}{\code{WebServerNamespace$validateJSON()}}
\item \href{#method-WebServerNamespace-toString}{\code{WebServerNamespace$toString()}}
\item \href{#method-WebServerNamespace-isValid}{\code{WebServerNamespace$isValid()}}
\item \href{#method-WebServerNamespace-getInvalidFields}{\code{WebServerNamespace$getInvalidFields()}}
\item \href{#method-WebServerNamespace-print}{\code{WebServerNamespace$print()}}
\item \href{#method-WebServerNamespace-clone}{\code{WebServerNamespace$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebServerNamespace-new"></a>}}
\if{latex}{\out{\hypertarget{method-WebServerNamespace-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new WebServerNamespace class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebServerNamespace$new(value = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{value}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebServerNamespace-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-WebServerNamespace-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebServerNamespace$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebServerNamespace-toList"></a>}}
\if{latex}{\out{\hypertarget{method-WebServerNamespace-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebServerNamespace$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
WebServerNamespace as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of WebServerNamespace (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebServerNamespace-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-WebServerNamespace-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert WebServerNamespace to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebServerNamespace$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebServerNamespace-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-WebServerNamespace-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of WebServerNamespace
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebServerNamespace$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of WebServerNamespace
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebServerNamespace-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-WebServerNamespace-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebServerNamespace$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
WebServerNamespace in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebServerNamespace-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-WebServerNamespace-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of WebServerNamespace
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebServerNamespace$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of WebServerNamespace
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebServerNamespace-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-WebServerNamespace-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to WebServerNamespace and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebServerNamespace$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebServerNamespace-toString"></a>}}
\if{latex}{\out{\hypertarget{method-WebServerNamespace-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebServerNamespace$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of WebServerNamespace
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebServerNamespace-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-WebServerNamespace-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebServerNamespace$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebServerNamespace-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-WebServerNamespace-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebServerNamespace$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebServerNamespace-print"></a>}}
\if{latex}{\out{\hypertarget{method-WebServerNamespace-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebServerNamespace$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebServerNamespace-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WebServerNamespace-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebServerNamespace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
