% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_info.R
\docType{class}
\name{WorkerInfo}
\alias{WorkerInfo}
\title{WorkerInfo}
\format{
An \code{R6Class} generator object
}
\description{
WorkerInfo Class

WorkerInfo Class
}
\details{
Create a new WorkerInfo
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer}

\item{\code{type}}{character}

\item{\code{supportedPredictors}}{list(character)}

\item{\code{version}}{character [optional]}

\item{\code{host}}{character [optional]}

\item{\code{prefix}}{character [optional]}

\item{\code{state}}{integer}

\item{\code{alive}}{integer}

\item{\code{serverTime}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WorkerInfo-new}{\code{WorkerInfo$new()}}
\item \href{#method-WorkerInfo-toJSON}{\code{WorkerInfo$toJSON()}}
\item \href{#method-WorkerInfo-fromJSON}{\code{WorkerInfo$fromJSON()}}
\item \href{#method-WorkerInfo-toJSONString}{\code{WorkerInfo$toJSONString()}}
\item \href{#method-WorkerInfo-fromJSONString}{\code{WorkerInfo$fromJSONString()}}
\item \href{#method-WorkerInfo-validateJSON}{\code{WorkerInfo$validateJSON()}}
\item \href{#method-WorkerInfo-toString}{\code{WorkerInfo$toString()}}
\item \href{#method-WorkerInfo-isValid}{\code{WorkerInfo$isValid()}}
\item \href{#method-WorkerInfo-getInvalidFields}{\code{WorkerInfo$getInvalidFields()}}
\item \href{#method-WorkerInfo-print}{\code{WorkerInfo$print()}}
\item \href{#method-WorkerInfo-clone}{\code{WorkerInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerInfo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new WorkerInfo class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerInfo$new(
  id,
  type,
  supportedPredictors,
  state,
  alive,
  serverTime,
  version = NULL,
  host = NULL,
  prefix = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{type}}{type}

\item{\code{supportedPredictors}}{supportedPredictors}

\item{\code{state}}{state}

\item{\code{alive}}{alive}

\item{\code{serverTime}}{serverTime}

\item{\code{version}}{version}

\item{\code{host}}{host}

\item{\code{prefix}}{prefix}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerInfo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerInfo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerInfo$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
WorkerInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerInfo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerInfo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of WorkerInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerInfo$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of WorkerInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerInfo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerInfo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerInfo$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
WorkerInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerInfo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerInfo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of WorkerInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerInfo$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of WorkerInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerInfo-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerInfo-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to WorkerInfo and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerInfo$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerInfo-toString"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerInfo-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerInfo$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of WorkerInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerInfo-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerInfo-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerInfo$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerInfo-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerInfo-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerInfo$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerInfo-print"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerInfo-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerInfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkerInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WorkerInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
