% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zodiac_analogue_nodes.R
\docType{class}
\name{ZodiacAnalogueNodes}
\alias{ZodiacAnalogueNodes}
\title{ZodiacAnalogueNodes}
\format{
An \code{R6Class} generator object
}
\description{
ZodiacAnalogueNodes Class

ZodiacAnalogueNodes Class
}
\details{
Create a new ZodiacAnalogueNodes
}
\examples{

## ------------------------------------------------
## Method `ZodiacAnalogueNodes$toList`
## ------------------------------------------------

# convert array of ZodiacAnalogueNodes (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{enabled}}{character [optional]}

\item{\code{minSimilarity}}{numeric [optional]}

\item{\code{minSharedPeaks}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ZodiacAnalogueNodes-new}{\code{ZodiacAnalogueNodes$new()}}
\item \href{#method-ZodiacAnalogueNodes-toJSON}{\code{ZodiacAnalogueNodes$toJSON()}}
\item \href{#method-ZodiacAnalogueNodes-toList}{\code{ZodiacAnalogueNodes$toList()}}
\item \href{#method-ZodiacAnalogueNodes-toSimpleType}{\code{ZodiacAnalogueNodes$toSimpleType()}}
\item \href{#method-ZodiacAnalogueNodes-fromJSON}{\code{ZodiacAnalogueNodes$fromJSON()}}
\item \href{#method-ZodiacAnalogueNodes-toJSONString}{\code{ZodiacAnalogueNodes$toJSONString()}}
\item \href{#method-ZodiacAnalogueNodes-fromJSONString}{\code{ZodiacAnalogueNodes$fromJSONString()}}
\item \href{#method-ZodiacAnalogueNodes-validateJSON}{\code{ZodiacAnalogueNodes$validateJSON()}}
\item \href{#method-ZodiacAnalogueNodes-toString}{\code{ZodiacAnalogueNodes$toString()}}
\item \href{#method-ZodiacAnalogueNodes-isValid}{\code{ZodiacAnalogueNodes$isValid()}}
\item \href{#method-ZodiacAnalogueNodes-getInvalidFields}{\code{ZodiacAnalogueNodes$getInvalidFields()}}
\item \href{#method-ZodiacAnalogueNodes-print}{\code{ZodiacAnalogueNodes$print()}}
\item \href{#method-ZodiacAnalogueNodes-clone}{\code{ZodiacAnalogueNodes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacAnalogueNodes-new"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacAnalogueNodes-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ZodiacAnalogueNodes class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacAnalogueNodes$new(
  enabled = NULL,
  minSimilarity = NULL,
  minSharedPeaks = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enabled}}{enabled}

\item{\code{minSimilarity}}{minSimilarity}

\item{\code{minSharedPeaks}}{minSharedPeaks}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacAnalogueNodes-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacAnalogueNodes-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacAnalogueNodes$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacAnalogueNodes-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacAnalogueNodes-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacAnalogueNodes$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ZodiacAnalogueNodes as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of ZodiacAnalogueNodes (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacAnalogueNodes-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacAnalogueNodes-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert ZodiacAnalogueNodes to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacAnalogueNodes$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacAnalogueNodes-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacAnalogueNodes-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ZodiacAnalogueNodes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacAnalogueNodes$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ZodiacAnalogueNodes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacAnalogueNodes-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacAnalogueNodes-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacAnalogueNodes$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ZodiacAnalogueNodes in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacAnalogueNodes-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacAnalogueNodes-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ZodiacAnalogueNodes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacAnalogueNodes$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ZodiacAnalogueNodes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacAnalogueNodes-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacAnalogueNodes-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ZodiacAnalogueNodes and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacAnalogueNodes$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacAnalogueNodes-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacAnalogueNodes-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacAnalogueNodes$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ZodiacAnalogueNodes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacAnalogueNodes-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacAnalogueNodes-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacAnalogueNodes$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacAnalogueNodes-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacAnalogueNodes-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacAnalogueNodes$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacAnalogueNodes-print"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacAnalogueNodes-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacAnalogueNodes$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacAnalogueNodes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacAnalogueNodes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacAnalogueNodes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
