% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zodiac_epochs.R
\docType{class}
\name{ZodiacEpochs}
\alias{ZodiacEpochs}
\title{ZodiacEpochs}
\format{
An \code{R6Class} generator object
}
\description{
ZodiacEpochs Class
}
\details{
Create a new ZodiacEpochs
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{iterations}}{integer [optional]}

\item{\code{burnInPeriod}}{integer [optional]}

\item{\code{numberOfMarkovChains}}{integer [optional]}

\item{\code{identifier}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ZodiacEpochs-new}{\code{ZodiacEpochs$new()}}
\item \href{#method-ZodiacEpochs-toJSON}{\code{ZodiacEpochs$toJSON()}}
\item \href{#method-ZodiacEpochs-fromJSON}{\code{ZodiacEpochs$fromJSON()}}
\item \href{#method-ZodiacEpochs-toJSONString}{\code{ZodiacEpochs$toJSONString()}}
\item \href{#method-ZodiacEpochs-fromJSONString}{\code{ZodiacEpochs$fromJSONString()}}
\item \href{#method-ZodiacEpochs-validateJSON}{\code{ZodiacEpochs$validateJSON()}}
\item \href{#method-ZodiacEpochs-toString}{\code{ZodiacEpochs$toString()}}
\item \href{#method-ZodiacEpochs-isValid}{\code{ZodiacEpochs$isValid()}}
\item \href{#method-ZodiacEpochs-getInvalidFields}{\code{ZodiacEpochs$getInvalidFields()}}
\item \href{#method-ZodiacEpochs-print}{\code{ZodiacEpochs$print()}}
\item \href{#method-ZodiacEpochs-clone}{\code{ZodiacEpochs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacEpochs-new"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacEpochs-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ZodiacEpochs class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacEpochs$new(
  iterations = NULL,
  burnInPeriod = NULL,
  numberOfMarkovChains = NULL,
  identifier = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{iterations}

\item{\code{burnInPeriod}}{burnInPeriod}

\item{\code{numberOfMarkovChains}}{numberOfMarkovChains}

\item{\code{identifier}}{identifier}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacEpochs-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacEpochs-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacEpochs$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ZodiacEpochs in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacEpochs-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacEpochs-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ZodiacEpochs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacEpochs$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ZodiacEpochs
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacEpochs-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacEpochs-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacEpochs$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ZodiacEpochs in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacEpochs-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacEpochs-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ZodiacEpochs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacEpochs$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ZodiacEpochs
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacEpochs-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacEpochs-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ZodiacEpochs and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacEpochs$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacEpochs-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacEpochs-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacEpochs$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ZodiacEpochs
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacEpochs-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacEpochs-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacEpochs$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacEpochs-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacEpochs-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacEpochs$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacEpochs-print"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacEpochs-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacEpochs$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacEpochs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacEpochs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacEpochs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
