% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zodiac_library_scoring.R
\docType{class}
\name{ZodiacLibraryScoring}
\alias{ZodiacLibraryScoring}
\title{ZodiacLibraryScoring}
\format{
An \code{R6Class} generator object
}
\description{
ZodiacLibraryScoring Class

ZodiacLibraryScoring Class
}
\details{
Create a new ZodiacLibraryScoring
}
\examples{

## ------------------------------------------------
## Method `ZodiacLibraryScoring$toList`
## ------------------------------------------------

# convert array of ZodiacLibraryScoring (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{enabled}}{character [optional]}

\item{\code{minSimilarity}}{numeric [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ZodiacLibraryScoring-new}{\code{ZodiacLibraryScoring$new()}}
\item \href{#method-ZodiacLibraryScoring-toJSON}{\code{ZodiacLibraryScoring$toJSON()}}
\item \href{#method-ZodiacLibraryScoring-toList}{\code{ZodiacLibraryScoring$toList()}}
\item \href{#method-ZodiacLibraryScoring-toSimpleType}{\code{ZodiacLibraryScoring$toSimpleType()}}
\item \href{#method-ZodiacLibraryScoring-fromJSON}{\code{ZodiacLibraryScoring$fromJSON()}}
\item \href{#method-ZodiacLibraryScoring-toJSONString}{\code{ZodiacLibraryScoring$toJSONString()}}
\item \href{#method-ZodiacLibraryScoring-fromJSONString}{\code{ZodiacLibraryScoring$fromJSONString()}}
\item \href{#method-ZodiacLibraryScoring-validateJSON}{\code{ZodiacLibraryScoring$validateJSON()}}
\item \href{#method-ZodiacLibraryScoring-toString}{\code{ZodiacLibraryScoring$toString()}}
\item \href{#method-ZodiacLibraryScoring-isValid}{\code{ZodiacLibraryScoring$isValid()}}
\item \href{#method-ZodiacLibraryScoring-getInvalidFields}{\code{ZodiacLibraryScoring$getInvalidFields()}}
\item \href{#method-ZodiacLibraryScoring-print}{\code{ZodiacLibraryScoring$print()}}
\item \href{#method-ZodiacLibraryScoring-clone}{\code{ZodiacLibraryScoring$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacLibraryScoring-new"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacLibraryScoring-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ZodiacLibraryScoring class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacLibraryScoring$new(enabled = NULL, minSimilarity = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enabled}}{enabled}

\item{\code{minSimilarity}}{minSimilarity}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacLibraryScoring-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacLibraryScoring-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacLibraryScoring$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacLibraryScoring-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacLibraryScoring-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacLibraryScoring$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ZodiacLibraryScoring as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of ZodiacLibraryScoring (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacLibraryScoring-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacLibraryScoring-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert ZodiacLibraryScoring to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacLibraryScoring$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacLibraryScoring-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacLibraryScoring-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ZodiacLibraryScoring
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacLibraryScoring$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ZodiacLibraryScoring
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacLibraryScoring-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacLibraryScoring-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacLibraryScoring$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ZodiacLibraryScoring in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacLibraryScoring-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacLibraryScoring-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ZodiacLibraryScoring
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacLibraryScoring$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ZodiacLibraryScoring
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacLibraryScoring-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacLibraryScoring-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ZodiacLibraryScoring and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacLibraryScoring$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacLibraryScoring-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacLibraryScoring-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacLibraryScoring$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ZodiacLibraryScoring
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacLibraryScoring-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacLibraryScoring-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacLibraryScoring$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacLibraryScoring-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacLibraryScoring-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacLibraryScoring$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacLibraryScoring-print"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacLibraryScoring-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacLibraryScoring$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacLibraryScoring-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacLibraryScoring-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacLibraryScoring$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
