% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/zodiac_number_of_considered_candidates_at300_mz.R
\docType{class}
\name{ZodiacNumberOfConsideredCandidatesAt300Mz}
\alias{ZodiacNumberOfConsideredCandidatesAt300Mz}
\title{ZodiacNumberOfConsideredCandidatesAt300Mz}
\format{
An \code{R6Class} generator object
}
\description{
ZodiacNumberOfConsideredCandidatesAt300Mz Class
}
\details{
Create a new ZodiacNumberOfConsideredCandidatesAt300Mz
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{integer [optional]}

\item{\code{identifier}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt300Mz-new}{\code{ZodiacNumberOfConsideredCandidatesAt300Mz$new()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt300Mz-toJSON}{\code{ZodiacNumberOfConsideredCandidatesAt300Mz$toJSON()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt300Mz-fromJSON}{\code{ZodiacNumberOfConsideredCandidatesAt300Mz$fromJSON()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt300Mz-toJSONString}{\code{ZodiacNumberOfConsideredCandidatesAt300Mz$toJSONString()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt300Mz-fromJSONString}{\code{ZodiacNumberOfConsideredCandidatesAt300Mz$fromJSONString()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt300Mz-validateJSON}{\code{ZodiacNumberOfConsideredCandidatesAt300Mz$validateJSON()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt300Mz-toString}{\code{ZodiacNumberOfConsideredCandidatesAt300Mz$toString()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt300Mz-isValid}{\code{ZodiacNumberOfConsideredCandidatesAt300Mz$isValid()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt300Mz-getInvalidFields}{\code{ZodiacNumberOfConsideredCandidatesAt300Mz$getInvalidFields()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt300Mz-print}{\code{ZodiacNumberOfConsideredCandidatesAt300Mz$print()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt300Mz-clone}{\code{ZodiacNumberOfConsideredCandidatesAt300Mz$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt300Mz-new"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt300Mz-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ZodiacNumberOfConsideredCandidatesAt300Mz class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt300Mz$new(
  value = NULL,
  identifier = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{value}

\item{\code{identifier}}{identifier}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt300Mz-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt300Mz-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt300Mz$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ZodiacNumberOfConsideredCandidatesAt300Mz in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt300Mz-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt300Mz-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ZodiacNumberOfConsideredCandidatesAt300Mz
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt300Mz$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ZodiacNumberOfConsideredCandidatesAt300Mz
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt300Mz-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt300Mz-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt300Mz$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ZodiacNumberOfConsideredCandidatesAt300Mz in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt300Mz-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt300Mz-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ZodiacNumberOfConsideredCandidatesAt300Mz
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt300Mz$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ZodiacNumberOfConsideredCandidatesAt300Mz
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt300Mz-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt300Mz-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ZodiacNumberOfConsideredCandidatesAt300Mz and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt300Mz$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt300Mz-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt300Mz-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt300Mz$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ZodiacNumberOfConsideredCandidatesAt300Mz
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt300Mz-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt300Mz-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt300Mz$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt300Mz-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt300Mz-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt300Mz$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt300Mz-print"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt300Mz-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt300Mz$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt300Mz-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt300Mz-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt300Mz$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
