% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/zodiac_number_of_considered_candidates_at800_mz.R
\docType{class}
\name{ZodiacNumberOfConsideredCandidatesAt800Mz}
\alias{ZodiacNumberOfConsideredCandidatesAt800Mz}
\title{ZodiacNumberOfConsideredCandidatesAt800Mz}
\format{
An \code{R6Class} generator object
}
\description{
ZodiacNumberOfConsideredCandidatesAt800Mz Class
}
\details{
Create a new ZodiacNumberOfConsideredCandidatesAt800Mz
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{integer [optional]}

\item{\code{identifier}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt800Mz-new}{\code{ZodiacNumberOfConsideredCandidatesAt800Mz$new()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt800Mz-toJSON}{\code{ZodiacNumberOfConsideredCandidatesAt800Mz$toJSON()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt800Mz-fromJSON}{\code{ZodiacNumberOfConsideredCandidatesAt800Mz$fromJSON()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt800Mz-toJSONString}{\code{ZodiacNumberOfConsideredCandidatesAt800Mz$toJSONString()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt800Mz-fromJSONString}{\code{ZodiacNumberOfConsideredCandidatesAt800Mz$fromJSONString()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt800Mz-validateJSON}{\code{ZodiacNumberOfConsideredCandidatesAt800Mz$validateJSON()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt800Mz-toString}{\code{ZodiacNumberOfConsideredCandidatesAt800Mz$toString()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt800Mz-isValid}{\code{ZodiacNumberOfConsideredCandidatesAt800Mz$isValid()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt800Mz-getInvalidFields}{\code{ZodiacNumberOfConsideredCandidatesAt800Mz$getInvalidFields()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt800Mz-print}{\code{ZodiacNumberOfConsideredCandidatesAt800Mz$print()}}
\item \href{#method-ZodiacNumberOfConsideredCandidatesAt800Mz-clone}{\code{ZodiacNumberOfConsideredCandidatesAt800Mz$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt800Mz-new"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt800Mz-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ZodiacNumberOfConsideredCandidatesAt800Mz class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt800Mz$new(
  value = NULL,
  identifier = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{value}

\item{\code{identifier}}{identifier}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt800Mz-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt800Mz-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt800Mz$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ZodiacNumberOfConsideredCandidatesAt800Mz in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt800Mz-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt800Mz-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ZodiacNumberOfConsideredCandidatesAt800Mz
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt800Mz$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ZodiacNumberOfConsideredCandidatesAt800Mz
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt800Mz-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt800Mz-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt800Mz$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ZodiacNumberOfConsideredCandidatesAt800Mz in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt800Mz-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt800Mz-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ZodiacNumberOfConsideredCandidatesAt800Mz
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt800Mz$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ZodiacNumberOfConsideredCandidatesAt800Mz
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt800Mz-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt800Mz-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ZodiacNumberOfConsideredCandidatesAt800Mz and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt800Mz$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt800Mz-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt800Mz-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt800Mz$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ZodiacNumberOfConsideredCandidatesAt800Mz
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt800Mz-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt800Mz-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt800Mz$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt800Mz-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt800Mz-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt800Mz$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt800Mz-print"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt800Mz-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt800Mz$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacNumberOfConsideredCandidatesAt800Mz-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacNumberOfConsideredCandidatesAt800Mz-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacNumberOfConsideredCandidatesAt800Mz$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
