% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zodiac_run_in_two_steps.R
\docType{class}
\name{ZodiacRunInTwoSteps}
\alias{ZodiacRunInTwoSteps}
\title{ZodiacRunInTwoSteps}
\format{
An \code{R6Class} generator object
}
\description{
ZodiacRunInTwoSteps Class
}
\details{
Create a new ZodiacRunInTwoSteps
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{character [optional]}

\item{\code{identifier}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ZodiacRunInTwoSteps-new}{\code{ZodiacRunInTwoSteps$new()}}
\item \href{#method-ZodiacRunInTwoSteps-toJSON}{\code{ZodiacRunInTwoSteps$toJSON()}}
\item \href{#method-ZodiacRunInTwoSteps-fromJSON}{\code{ZodiacRunInTwoSteps$fromJSON()}}
\item \href{#method-ZodiacRunInTwoSteps-toJSONString}{\code{ZodiacRunInTwoSteps$toJSONString()}}
\item \href{#method-ZodiacRunInTwoSteps-fromJSONString}{\code{ZodiacRunInTwoSteps$fromJSONString()}}
\item \href{#method-ZodiacRunInTwoSteps-validateJSON}{\code{ZodiacRunInTwoSteps$validateJSON()}}
\item \href{#method-ZodiacRunInTwoSteps-toString}{\code{ZodiacRunInTwoSteps$toString()}}
\item \href{#method-ZodiacRunInTwoSteps-isValid}{\code{ZodiacRunInTwoSteps$isValid()}}
\item \href{#method-ZodiacRunInTwoSteps-getInvalidFields}{\code{ZodiacRunInTwoSteps$getInvalidFields()}}
\item \href{#method-ZodiacRunInTwoSteps-print}{\code{ZodiacRunInTwoSteps$print()}}
\item \href{#method-ZodiacRunInTwoSteps-clone}{\code{ZodiacRunInTwoSteps$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacRunInTwoSteps-new"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacRunInTwoSteps-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ZodiacRunInTwoSteps class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacRunInTwoSteps$new(value = NULL, identifier = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{value}

\item{\code{identifier}}{identifier}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacRunInTwoSteps-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacRunInTwoSteps-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacRunInTwoSteps$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ZodiacRunInTwoSteps in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacRunInTwoSteps-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacRunInTwoSteps-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ZodiacRunInTwoSteps
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacRunInTwoSteps$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ZodiacRunInTwoSteps
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacRunInTwoSteps-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacRunInTwoSteps-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacRunInTwoSteps$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ZodiacRunInTwoSteps in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacRunInTwoSteps-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacRunInTwoSteps-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ZodiacRunInTwoSteps
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacRunInTwoSteps$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ZodiacRunInTwoSteps
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacRunInTwoSteps-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacRunInTwoSteps-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ZodiacRunInTwoSteps and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacRunInTwoSteps$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacRunInTwoSteps-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacRunInTwoSteps-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacRunInTwoSteps$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ZodiacRunInTwoSteps
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacRunInTwoSteps-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacRunInTwoSteps-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacRunInTwoSteps$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacRunInTwoSteps-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacRunInTwoSteps-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacRunInTwoSteps$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacRunInTwoSteps-print"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacRunInTwoSteps-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacRunInTwoSteps$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZodiacRunInTwoSteps-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZodiacRunInTwoSteps-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZodiacRunInTwoSteps$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
