% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_ms_ms_data.R
\docType{class}
\name{AnnotatedMsMsData}
\alias{AnnotatedMsMsData}
\title{AnnotatedMsMsData}
\format{
An \code{R6Class} generator object
}
\description{
AnnotatedMsMsData Class

AnnotatedMsMsData Class
}
\details{
Create a new AnnotatedMsMsData
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mergedMs2}}{\link{AnnotatedSpectrum}}

\item{\code{ms2Spectra}}{list(\link{AnnotatedSpectrum})}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnnotatedMsMsData-new}{\code{AnnotatedMsMsData$new()}}
\item \href{#method-AnnotatedMsMsData-toJSON}{\code{AnnotatedMsMsData$toJSON()}}
\item \href{#method-AnnotatedMsMsData-fromJSON}{\code{AnnotatedMsMsData$fromJSON()}}
\item \href{#method-AnnotatedMsMsData-toJSONString}{\code{AnnotatedMsMsData$toJSONString()}}
\item \href{#method-AnnotatedMsMsData-fromJSONString}{\code{AnnotatedMsMsData$fromJSONString()}}
\item \href{#method-AnnotatedMsMsData-validateJSON}{\code{AnnotatedMsMsData$validateJSON()}}
\item \href{#method-AnnotatedMsMsData-toString}{\code{AnnotatedMsMsData$toString()}}
\item \href{#method-AnnotatedMsMsData-isValid}{\code{AnnotatedMsMsData$isValid()}}
\item \href{#method-AnnotatedMsMsData-getInvalidFields}{\code{AnnotatedMsMsData$getInvalidFields()}}
\item \href{#method-AnnotatedMsMsData-print}{\code{AnnotatedMsMsData$print()}}
\item \href{#method-AnnotatedMsMsData-clone}{\code{AnnotatedMsMsData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedMsMsData-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedMsMsData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AnnotatedMsMsData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedMsMsData$new(mergedMs2, ms2Spectra, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mergedMs2}}{mergedMs2}

\item{\code{ms2Spectra}}{ms2Spectra}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedMsMsData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedMsMsData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedMsMsData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AnnotatedMsMsData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedMsMsData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedMsMsData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of AnnotatedMsMsData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedMsMsData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AnnotatedMsMsData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedMsMsData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedMsMsData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedMsMsData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AnnotatedMsMsData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedMsMsData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedMsMsData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of AnnotatedMsMsData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedMsMsData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AnnotatedMsMsData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedMsMsData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedMsMsData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to AnnotatedMsMsData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedMsMsData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedMsMsData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedMsMsData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedMsMsData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of AnnotatedMsMsData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedMsMsData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedMsMsData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedMsMsData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedMsMsData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedMsMsData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedMsMsData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedMsMsData-print"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedMsMsData-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedMsMsData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotatedMsMsData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotatedMsMsData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatedMsMsData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
