% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopus.R
\docType{class}
\name{Canopus}
\alias{Canopus}
\title{Canopus}
\format{
An \code{R6Class} generator object
}
\description{
User/developer friendly parameter subset for the CANOPUS tool  CANOPUS is parameter free, so this Object is just a flag that canopus should be executed.

Canopus Class
}
\details{
Create a new Canopus
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{enabled}}{tags whether the tool is enabled character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Canopus-new}{\code{Canopus$new()}}
\item \href{#method-Canopus-toJSON}{\code{Canopus$toJSON()}}
\item \href{#method-Canopus-fromJSON}{\code{Canopus$fromJSON()}}
\item \href{#method-Canopus-toJSONString}{\code{Canopus$toJSONString()}}
\item \href{#method-Canopus-fromJSONString}{\code{Canopus$fromJSONString()}}
\item \href{#method-Canopus-validateJSON}{\code{Canopus$validateJSON()}}
\item \href{#method-Canopus-toString}{\code{Canopus$toString()}}
\item \href{#method-Canopus-isValid}{\code{Canopus$isValid()}}
\item \href{#method-Canopus-getInvalidFields}{\code{Canopus$getInvalidFields()}}
\item \href{#method-Canopus-print}{\code{Canopus$print()}}
\item \href{#method-Canopus-clone}{\code{Canopus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Canopus-new"></a>}}
\if{latex}{\out{\hypertarget{method-Canopus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Canopus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Canopus$new(enabled = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enabled}}{tags whether the tool is enabled}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Canopus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Canopus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Canopus$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Canopus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Canopus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Canopus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Canopus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Canopus$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Canopus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Canopus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Canopus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Canopus$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Canopus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Canopus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Canopus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Canopus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Canopus$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Canopus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Canopus-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Canopus-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Canopus and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Canopus$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Canopus-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Canopus-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Canopus$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Canopus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Canopus-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Canopus-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Canopus$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Canopus-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Canopus-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Canopus$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Canopus-print"></a>}}
\if{latex}{\out{\hypertarget{method-Canopus-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Canopus$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Canopus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Canopus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Canopus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
