% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computed_subtools.R
\docType{class}
\name{ComputedSubtools}
\alias{ComputedSubtools}
\title{ComputedSubtools}
\format{
An \code{R6Class} generator object
}
\description{
Specifies which tools have been executed for this feature. Can be used to estimate which results can be expected. Null if it was not requested und non-null otherwise.

ComputedSubtools Class
}
\details{
Create a new ComputedSubtools
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{librarySearch}}{character [optional]}

\item{\code{formulaSearch}}{character [optional]}

\item{\code{zodiac}}{character [optional]}

\item{\code{fingerprint}}{character [optional]}

\item{\code{canopus}}{character [optional]}

\item{\code{structureSearch}}{character [optional]}

\item{\code{deNovoSearch}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ComputedSubtools-new}{\code{ComputedSubtools$new()}}
\item \href{#method-ComputedSubtools-toJSON}{\code{ComputedSubtools$toJSON()}}
\item \href{#method-ComputedSubtools-fromJSON}{\code{ComputedSubtools$fromJSON()}}
\item \href{#method-ComputedSubtools-toJSONString}{\code{ComputedSubtools$toJSONString()}}
\item \href{#method-ComputedSubtools-fromJSONString}{\code{ComputedSubtools$fromJSONString()}}
\item \href{#method-ComputedSubtools-validateJSON}{\code{ComputedSubtools$validateJSON()}}
\item \href{#method-ComputedSubtools-toString}{\code{ComputedSubtools$toString()}}
\item \href{#method-ComputedSubtools-isValid}{\code{ComputedSubtools$isValid()}}
\item \href{#method-ComputedSubtools-getInvalidFields}{\code{ComputedSubtools$getInvalidFields()}}
\item \href{#method-ComputedSubtools-print}{\code{ComputedSubtools$print()}}
\item \href{#method-ComputedSubtools-clone}{\code{ComputedSubtools$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComputedSubtools-new"></a>}}
\if{latex}{\out{\hypertarget{method-ComputedSubtools-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ComputedSubtools class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedSubtools$new(
  librarySearch = NULL,
  formulaSearch = NULL,
  zodiac = NULL,
  fingerprint = NULL,
  canopus = NULL,
  structureSearch = NULL,
  deNovoSearch = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{librarySearch}}{librarySearch}

\item{\code{formulaSearch}}{formulaSearch}

\item{\code{zodiac}}{zodiac}

\item{\code{fingerprint}}{fingerprint}

\item{\code{canopus}}{canopus}

\item{\code{structureSearch}}{structureSearch}

\item{\code{deNovoSearch}}{deNovoSearch}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComputedSubtools-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComputedSubtools-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedSubtools$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ComputedSubtools in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComputedSubtools-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComputedSubtools-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ComputedSubtools
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedSubtools$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ComputedSubtools
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComputedSubtools-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ComputedSubtools-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedSubtools$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ComputedSubtools in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComputedSubtools-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ComputedSubtools-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ComputedSubtools
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedSubtools$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ComputedSubtools
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComputedSubtools-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComputedSubtools-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ComputedSubtools and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedSubtools$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComputedSubtools-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ComputedSubtools-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedSubtools$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ComputedSubtools
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComputedSubtools-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ComputedSubtools-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedSubtools$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComputedSubtools-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ComputedSubtools-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedSubtools$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComputedSubtools-print"></a>}}
\if{latex}{\out{\hypertarget{method-ComputedSubtools-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedSubtools$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComputedSubtools-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ComputedSubtools-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedSubtools$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
