% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\docType{class}
\name{Job}
\alias{Job}
\title{Job}
\format{
An \code{R6Class} generator object
}
\description{
Identifier created by the SIRIUS Nightsky API for a newly created Job.  Object can be enriched with Job status/progress information ({@link JobProgress JobProgress}) and/or Job command information.

Job Class
}
\details{
Create a new Job
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Unique identifier to access the job via the API character [optional]}

\item{\code{command}}{Command string of the executed Task character [optional]}

\item{\code{progress}}{\link{JobProgress} [optional]}

\item{\code{affectedCompoundIds}}{List of compoundIds that are affected by this job.  This lis will also contain compoundIds where not all features of the compound are affected by the job.  If this job is creating compounds (e.g. data import jobs) this value will be NULL until the jobs has finished list(character) [optional]}

\item{\code{affectedAlignedFeatureIds}}{List of alignedFeatureIds that are affected by this job.  If this job is creating features (e.g. data import jobs) this value will be NULL until the jobs has finished list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Job-new}{\code{Job$new()}}
\item \href{#method-Job-toJSON}{\code{Job$toJSON()}}
\item \href{#method-Job-fromJSON}{\code{Job$fromJSON()}}
\item \href{#method-Job-toJSONString}{\code{Job$toJSONString()}}
\item \href{#method-Job-fromJSONString}{\code{Job$fromJSONString()}}
\item \href{#method-Job-validateJSON}{\code{Job$validateJSON()}}
\item \href{#method-Job-toString}{\code{Job$toString()}}
\item \href{#method-Job-isValid}{\code{Job$isValid()}}
\item \href{#method-Job-getInvalidFields}{\code{Job$getInvalidFields()}}
\item \href{#method-Job-print}{\code{Job$print()}}
\item \href{#method-Job-clone}{\code{Job$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-new"></a>}}
\if{latex}{\out{\hypertarget{method-Job-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Job class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$new(
  id = NULL,
  command = NULL,
  progress = NULL,
  affectedCompoundIds = NULL,
  affectedAlignedFeatureIds = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Unique identifier to access the job via the API}

\item{\code{command}}{Command string of the executed Task}

\item{\code{progress}}{progress}

\item{\code{affectedCompoundIds}}{List of compoundIds that are affected by this job.  This lis will also contain compoundIds where not all features of the compound are affected by the job.  If this job is creating compounds (e.g. data import jobs) this value will be NULL until the jobs has finished}

\item{\code{affectedAlignedFeatureIds}}{List of alignedFeatureIds that are affected by this job.  If this job is creating features (e.g. data import jobs) this value will be NULL until the jobs has finished}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Job-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Job in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Job-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Job
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Job-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Job in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Job-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Job
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Job-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Job and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Job-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Job
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Job-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Job-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-print"></a>}}
\if{latex}{\out{\hypertarget{method-Job-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Job-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
