% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcms_submission_parameters.R
\docType{class}
\name{LcmsSubmissionParameters}
\alias{LcmsSubmissionParameters}
\title{LcmsSubmissionParameters}
\format{
An \code{R6Class} generator object
}
\description{
LcmsSubmissionParameters Class

LcmsSubmissionParameters Class
}
\details{
Create a new LcmsSubmissionParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alignLCMSRuns}}{Specifies whether LC/MS runs should be aligned character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LcmsSubmissionParameters-new}{\code{LcmsSubmissionParameters$new()}}
\item \href{#method-LcmsSubmissionParameters-toJSON}{\code{LcmsSubmissionParameters$toJSON()}}
\item \href{#method-LcmsSubmissionParameters-fromJSON}{\code{LcmsSubmissionParameters$fromJSON()}}
\item \href{#method-LcmsSubmissionParameters-toJSONString}{\code{LcmsSubmissionParameters$toJSONString()}}
\item \href{#method-LcmsSubmissionParameters-fromJSONString}{\code{LcmsSubmissionParameters$fromJSONString()}}
\item \href{#method-LcmsSubmissionParameters-validateJSON}{\code{LcmsSubmissionParameters$validateJSON()}}
\item \href{#method-LcmsSubmissionParameters-toString}{\code{LcmsSubmissionParameters$toString()}}
\item \href{#method-LcmsSubmissionParameters-isValid}{\code{LcmsSubmissionParameters$isValid()}}
\item \href{#method-LcmsSubmissionParameters-getInvalidFields}{\code{LcmsSubmissionParameters$getInvalidFields()}}
\item \href{#method-LcmsSubmissionParameters-print}{\code{LcmsSubmissionParameters$print()}}
\item \href{#method-LcmsSubmissionParameters-clone}{\code{LcmsSubmissionParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LcmsSubmissionParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-LcmsSubmissionParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new LcmsSubmissionParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LcmsSubmissionParameters$new(alignLCMSRuns = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alignLCMSRuns}}{Specifies whether LC/MS runs should be aligned. Default to TRUE.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LcmsSubmissionParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LcmsSubmissionParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LcmsSubmissionParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
LcmsSubmissionParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LcmsSubmissionParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LcmsSubmissionParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of LcmsSubmissionParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LcmsSubmissionParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of LcmsSubmissionParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LcmsSubmissionParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LcmsSubmissionParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LcmsSubmissionParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
LcmsSubmissionParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LcmsSubmissionParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LcmsSubmissionParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of LcmsSubmissionParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LcmsSubmissionParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of LcmsSubmissionParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LcmsSubmissionParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LcmsSubmissionParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to LcmsSubmissionParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LcmsSubmissionParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LcmsSubmissionParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-LcmsSubmissionParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LcmsSubmissionParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of LcmsSubmissionParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LcmsSubmissionParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-LcmsSubmissionParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LcmsSubmissionParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LcmsSubmissionParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-LcmsSubmissionParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LcmsSubmissionParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LcmsSubmissionParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-LcmsSubmissionParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LcmsSubmissionParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LcmsSubmissionParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LcmsSubmissionParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LcmsSubmissionParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
