% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid_annotation.R
\docType{class}
\name{LipidAnnotation}
\alias{LipidAnnotation}
\title{LipidAnnotation}
\format{
An \code{R6Class} generator object
}
\description{
LipidAnnotation Class

LipidAnnotation Class
}
\details{
Create a new LipidAnnotation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lipidSpecies}}{Predicted lipid species in LIPID MAPS notation.  NULL if not classified as lipid. character [optional]}

\item{\code{lipidMapsId}}{LIPID MAPS id of the predicted lipid class. character [optional]}

\item{\code{lipidClassName}}{Human-readable name of the predicted lipid class. character [optional]}

\item{\code{hypotheticalStructure}}{Hypothetical molecular structure of the predicted lipid species as SMILES.  NULL if hypothetical structure not available. character [optional]}

\item{\code{chainsUnknown}}{True of the formula composition of the chains could not be determined from the MS/MS. character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LipidAnnotation-new}{\code{LipidAnnotation$new()}}
\item \href{#method-LipidAnnotation-toJSON}{\code{LipidAnnotation$toJSON()}}
\item \href{#method-LipidAnnotation-fromJSON}{\code{LipidAnnotation$fromJSON()}}
\item \href{#method-LipidAnnotation-toJSONString}{\code{LipidAnnotation$toJSONString()}}
\item \href{#method-LipidAnnotation-fromJSONString}{\code{LipidAnnotation$fromJSONString()}}
\item \href{#method-LipidAnnotation-validateJSON}{\code{LipidAnnotation$validateJSON()}}
\item \href{#method-LipidAnnotation-toString}{\code{LipidAnnotation$toString()}}
\item \href{#method-LipidAnnotation-isValid}{\code{LipidAnnotation$isValid()}}
\item \href{#method-LipidAnnotation-getInvalidFields}{\code{LipidAnnotation$getInvalidFields()}}
\item \href{#method-LipidAnnotation-print}{\code{LipidAnnotation$print()}}
\item \href{#method-LipidAnnotation-clone}{\code{LipidAnnotation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidAnnotation-new"></a>}}
\if{latex}{\out{\hypertarget{method-LipidAnnotation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new LipidAnnotation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidAnnotation$new(
  lipidSpecies = NULL,
  lipidMapsId = NULL,
  lipidClassName = NULL,
  hypotheticalStructure = NULL,
  chainsUnknown = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lipidSpecies}}{Predicted lipid species in LIPID MAPS notation.  NULL if not classified as lipid.}

\item{\code{lipidMapsId}}{LIPID MAPS id of the predicted lipid class.}

\item{\code{lipidClassName}}{Human-readable name of the predicted lipid class.}

\item{\code{hypotheticalStructure}}{Hypothetical molecular structure of the predicted lipid species as SMILES.  NULL if hypothetical structure not available.}

\item{\code{chainsUnknown}}{True of the formula composition of the chains could not be determined from the MS/MS.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidAnnotation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LipidAnnotation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidAnnotation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
LipidAnnotation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidAnnotation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LipidAnnotation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of LipidAnnotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidAnnotation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of LipidAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidAnnotation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LipidAnnotation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidAnnotation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
LipidAnnotation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidAnnotation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LipidAnnotation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of LipidAnnotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidAnnotation$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of LipidAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidAnnotation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LipidAnnotation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to LipidAnnotation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidAnnotation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidAnnotation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-LipidAnnotation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidAnnotation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of LipidAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidAnnotation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-LipidAnnotation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidAnnotation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidAnnotation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-LipidAnnotation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidAnnotation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidAnnotation-print"></a>}}
\if{latex}{\out{\hypertarget{method-LipidAnnotation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidAnnotation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidAnnotation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LipidAnnotation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidAnnotation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
