% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_novelist.R
\docType{class}
\name{MsNovelist}
\alias{MsNovelist}
\title{MsNovelist}
\format{
An \code{R6Class} generator object
}
\description{
MsNovelist Class

MsNovelist Class
}
\details{
Create a new MsNovelist
}
\examples{

## ------------------------------------------------
## Method `MsNovelist$toList`
## ------------------------------------------------

# convert array of MsNovelist (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{enabled}}{tags whether the tool is enabled character [optional]}

\item{\code{numberOfCandidateToPredict}}{Number of structure candidates to be predicted by MsNovelist.  Max Value 128. Values > 128 will be set to 128.  Actual number of returned candidate might be lower du to duplicates being created by MsNovelist. integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MsNovelist-new}{\code{MsNovelist$new()}}
\item \href{#method-MsNovelist-toJSON}{\code{MsNovelist$toJSON()}}
\item \href{#method-MsNovelist-toList}{\code{MsNovelist$toList()}}
\item \href{#method-MsNovelist-toSimpleType}{\code{MsNovelist$toSimpleType()}}
\item \href{#method-MsNovelist-fromJSON}{\code{MsNovelist$fromJSON()}}
\item \href{#method-MsNovelist-toJSONString}{\code{MsNovelist$toJSONString()}}
\item \href{#method-MsNovelist-fromJSONString}{\code{MsNovelist$fromJSONString()}}
\item \href{#method-MsNovelist-validateJSON}{\code{MsNovelist$validateJSON()}}
\item \href{#method-MsNovelist-toString}{\code{MsNovelist$toString()}}
\item \href{#method-MsNovelist-isValid}{\code{MsNovelist$isValid()}}
\item \href{#method-MsNovelist-getInvalidFields}{\code{MsNovelist$getInvalidFields()}}
\item \href{#method-MsNovelist-print}{\code{MsNovelist$print()}}
\item \href{#method-MsNovelist-clone}{\code{MsNovelist$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsNovelist-new"></a>}}
\if{latex}{\out{\hypertarget{method-MsNovelist-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new MsNovelist class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsNovelist$new(enabled = NULL, numberOfCandidateToPredict = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enabled}}{tags whether the tool is enabled}

\item{\code{numberOfCandidateToPredict}}{Number of structure candidates to be predicted by MsNovelist.  Max Value 128. Values > 128 will be set to 128.  Actual number of returned candidate might be lower du to duplicates being created by MsNovelist.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsNovelist-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MsNovelist-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsNovelist$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsNovelist-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MsNovelist-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsNovelist$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MsNovelist as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of MsNovelist (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsNovelist-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-MsNovelist-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert MsNovelist to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsNovelist$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsNovelist-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MsNovelist-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of MsNovelist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsNovelist$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MsNovelist
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsNovelist-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MsNovelist-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsNovelist$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
MsNovelist in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsNovelist-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MsNovelist-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of MsNovelist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsNovelist$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MsNovelist
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsNovelist-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MsNovelist-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to MsNovelist and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsNovelist$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsNovelist-toString"></a>}}
\if{latex}{\out{\hypertarget{method-MsNovelist-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsNovelist$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of MsNovelist
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsNovelist-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-MsNovelist-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsNovelist$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsNovelist-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-MsNovelist-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsNovelist$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsNovelist-print"></a>}}
\if{latex}{\out{\hypertarget{method-MsNovelist-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsNovelist$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsNovelist-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MsNovelist-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsNovelist$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
