% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annotation.R
\docType{class}
\name{PeakAnnotation}
\alias{PeakAnnotation}
\title{PeakAnnotation}
\format{
An \code{R6Class} generator object
}
\description{
PeakAnnotation Class

PeakAnnotation Class
}
\details{
Create a new PeakAnnotation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fragmentId}}{Identifier of the peak/fragment. Can be used to map fragments and peaks  among fragmentation trees and spectra. integer}

\item{\code{molecularFormula}}{Molecular formula that has been annotated to this peak character [optional]}

\item{\code{ionization}}{Ionization that has been annotated to this peak character [optional]}

\item{\code{exactMass}}{Exact mass of the annotated molecular formula and ionization numeric [optional]}

\item{\code{massDeviationMz}}{Absolute mass deviation of the exact mass to the measured peak mass in mDa numeric [optional]}

\item{\code{massDeviationPpm}}{Relative mass deviation of the exact mass to the measured peak mass in ppm numeric [optional]}

\item{\code{recalibratedMassDeviationMz}}{Absolute mass deviation of the exact mass to the recalibrated peak mass in mDa numeric [optional]}

\item{\code{recalibratedMassDeviationPpm}}{Relative mass deviation of the exact mass to the recalibrated peak mass in ppm numeric [optional]}

\item{\code{parentPeak}}{\link{ParentPeak} [optional]}

\item{\code{substructureAtoms}}{Array/List of indices of the atoms of the structure candidate that are part of this fragments substructure  (highlighted atoms) list(integer) [optional]}

\item{\code{substructureBonds}}{Array/List of indices of the bonds of the structure candidate that are part of this fragments substructure  (highlighted bonds)   Null if substructure annotation not available or not requested. list(integer) [optional]}

\item{\code{substructureBondsCut}}{Array/List of indices of the bonds of the structure candidate that need to be cut to produce this fragments  substructure (highlighted cutted bonds).   Null if substructure annotation not available or not requested. list(integer) [optional]}

\item{\code{substructureScore}}{This score roughly reflects the probability of this fragment forming.   This is the score of the path from root to this node which has the maximal score or \"profit\".  The score of a path is equal to the sum of scores of its contained fragments and edges.  Note: Refers to 'totalScore' in CombinatorialNode   Null if substructure annotation not available or not requested. numeric [optional]}

\item{\code{hydrogenRearrangements}}{Number of hydrogens rearrangements needed to match the substructure to the fragment formula.   Null if substructure annotation not available or not requested. integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PeakAnnotation-new}{\code{PeakAnnotation$new()}}
\item \href{#method-PeakAnnotation-toJSON}{\code{PeakAnnotation$toJSON()}}
\item \href{#method-PeakAnnotation-fromJSON}{\code{PeakAnnotation$fromJSON()}}
\item \href{#method-PeakAnnotation-toJSONString}{\code{PeakAnnotation$toJSONString()}}
\item \href{#method-PeakAnnotation-fromJSONString}{\code{PeakAnnotation$fromJSONString()}}
\item \href{#method-PeakAnnotation-validateJSON}{\code{PeakAnnotation$validateJSON()}}
\item \href{#method-PeakAnnotation-toString}{\code{PeakAnnotation$toString()}}
\item \href{#method-PeakAnnotation-isValid}{\code{PeakAnnotation$isValid()}}
\item \href{#method-PeakAnnotation-getInvalidFields}{\code{PeakAnnotation$getInvalidFields()}}
\item \href{#method-PeakAnnotation-print}{\code{PeakAnnotation$print()}}
\item \href{#method-PeakAnnotation-clone}{\code{PeakAnnotation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakAnnotation-new"></a>}}
\if{latex}{\out{\hypertarget{method-PeakAnnotation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PeakAnnotation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakAnnotation$new(
  fragmentId,
  molecularFormula = NULL,
  ionization = NULL,
  exactMass = NULL,
  massDeviationMz = NULL,
  massDeviationPpm = NULL,
  recalibratedMassDeviationMz = NULL,
  recalibratedMassDeviationPpm = NULL,
  parentPeak = NULL,
  substructureAtoms = NULL,
  substructureBonds = NULL,
  substructureBondsCut = NULL,
  substructureScore = NULL,
  hydrogenRearrangements = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fragmentId}}{Identifier of the peak/fragment. Can be used to map fragments and peaks  among fragmentation trees and spectra.}

\item{\code{molecularFormula}}{Molecular formula that has been annotated to this peak}

\item{\code{ionization}}{Ionization that has been annotated to this peak}

\item{\code{exactMass}}{Exact mass of the annotated molecular formula and ionization}

\item{\code{massDeviationMz}}{Absolute mass deviation of the exact mass to the measured peak mass in mDa}

\item{\code{massDeviationPpm}}{Relative mass deviation of the exact mass to the measured peak mass in ppm}

\item{\code{recalibratedMassDeviationMz}}{Absolute mass deviation of the exact mass to the recalibrated peak mass in mDa}

\item{\code{recalibratedMassDeviationPpm}}{Relative mass deviation of the exact mass to the recalibrated peak mass in ppm}

\item{\code{parentPeak}}{parentPeak}

\item{\code{substructureAtoms}}{Array/List of indices of the atoms of the structure candidate that are part of this fragments substructure  (highlighted atoms)}

\item{\code{substructureBonds}}{Array/List of indices of the bonds of the structure candidate that are part of this fragments substructure  (highlighted bonds)   Null if substructure annotation not available or not requested.}

\item{\code{substructureBondsCut}}{Array/List of indices of the bonds of the structure candidate that need to be cut to produce this fragments  substructure (highlighted cutted bonds).   Null if substructure annotation not available or not requested.}

\item{\code{substructureScore}}{This score roughly reflects the probability of this fragment forming.   This is the score of the path from root to this node which has the maximal score or \"profit\".  The score of a path is equal to the sum of scores of its contained fragments and edges.  Note: Refers to 'totalScore' in CombinatorialNode   Null if substructure annotation not available or not requested.}

\item{\code{hydrogenRearrangements}}{Number of hydrogens rearrangements needed to match the substructure to the fragment formula.   Null if substructure annotation not available or not requested.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakAnnotation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PeakAnnotation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakAnnotation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PeakAnnotation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakAnnotation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PeakAnnotation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PeakAnnotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakAnnotation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PeakAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakAnnotation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PeakAnnotation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakAnnotation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PeakAnnotation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakAnnotation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PeakAnnotation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PeakAnnotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakAnnotation$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PeakAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakAnnotation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PeakAnnotation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PeakAnnotation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakAnnotation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakAnnotation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PeakAnnotation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakAnnotation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PeakAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakAnnotation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PeakAnnotation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakAnnotation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakAnnotation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PeakAnnotation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakAnnotation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakAnnotation-print"></a>}}
\if{latex}{\out{\hypertarget{method-PeakAnnotation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakAnnotation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeakAnnotation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PeakAnnotation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeakAnnotation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
