% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchable_database.R
\docType{class}
\name{SearchableDatabase}
\alias{SearchableDatabase}
\title{SearchableDatabase}
\format{
An \code{R6Class} generator object
}
\description{
SearchableDatabase Class

SearchableDatabase Class
}
\details{
Create a new SearchableDatabase
}
\examples{

## ------------------------------------------------
## Method `SearchableDatabase$toList`
## ------------------------------------------------

# convert array of SearchableDatabase (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{displayName}}{display name of the database  Should be short character [optional]}

\item{\code{location}}{Storage location of user database  Might be NULL for non-user databases or if default location is used. character [optional]}

\item{\code{matchRtOfReferenceSpectra}}{Indicates whether this database shall be used to use retention time information for library matching.  Typically used for in-house spectral libraries that have been measured on character [optional]}

\item{\code{databaseId}}{A unique identifier or name of the database.  Should only contain file path and url save characters  For user databases this is usually the file name. character}

\item{\code{customDb}}{Indicates whether the database is a user managed custom database or if it is a  database that is included in SIRIUS which cannot be modified. character}

\item{\code{searchable}}{True when this database can be used as a search parameter.  False if the database is just an additional filter that can be applied after search. character}

\item{\code{dbDate}}{Date on which the data was imported / database was created. character [optional]}

\item{\code{dbVersion}}{database schema version integer [optional]}

\item{\code{updateNeeded}}{If true the database version is outdated and the database needs to be updated or re-imported before it can be used. character}

\item{\code{numberOfStructures}}{Number of unique compounds available in this database. integer [optional]}

\item{\code{numberOfFormulas}}{Number of different molecular formulas available in this database. integer [optional]}

\item{\code{numberOfReferenceSpectra}}{Number of reference spectra available in this database integer [optional]}

\item{\code{errorMessage}}{Error message if the database could not be loaded character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SearchableDatabase-new}{\code{SearchableDatabase$new()}}
\item \href{#method-SearchableDatabase-toJSON}{\code{SearchableDatabase$toJSON()}}
\item \href{#method-SearchableDatabase-toList}{\code{SearchableDatabase$toList()}}
\item \href{#method-SearchableDatabase-toSimpleType}{\code{SearchableDatabase$toSimpleType()}}
\item \href{#method-SearchableDatabase-fromJSON}{\code{SearchableDatabase$fromJSON()}}
\item \href{#method-SearchableDatabase-toJSONString}{\code{SearchableDatabase$toJSONString()}}
\item \href{#method-SearchableDatabase-fromJSONString}{\code{SearchableDatabase$fromJSONString()}}
\item \href{#method-SearchableDatabase-validateJSON}{\code{SearchableDatabase$validateJSON()}}
\item \href{#method-SearchableDatabase-toString}{\code{SearchableDatabase$toString()}}
\item \href{#method-SearchableDatabase-isValid}{\code{SearchableDatabase$isValid()}}
\item \href{#method-SearchableDatabase-getInvalidFields}{\code{SearchableDatabase$getInvalidFields()}}
\item \href{#method-SearchableDatabase-print}{\code{SearchableDatabase$print()}}
\item \href{#method-SearchableDatabase-clone}{\code{SearchableDatabase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabase-new"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabase-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SearchableDatabase class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabase$new(
  databaseId,
  customDb,
  searchable,
  updateNeeded,
  displayName = NULL,
  location = NULL,
  matchRtOfReferenceSpectra = FALSE,
  dbDate = NULL,
  dbVersion = NULL,
  numberOfStructures = NULL,
  numberOfFormulas = NULL,
  numberOfReferenceSpectra = NULL,
  errorMessage = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{databaseId}}{A unique identifier or name of the database.  Should only contain file path and url save characters  For user databases this is usually the file name.}

\item{\code{customDb}}{Indicates whether the database is a user managed custom database or if it is a  database that is included in SIRIUS which cannot be modified.}

\item{\code{searchable}}{True when this database can be used as a search parameter.  False if the database is just an additional filter that can be applied after search.}

\item{\code{updateNeeded}}{If true the database version is outdated and the database needs to be updated or re-imported before it can be used.}

\item{\code{displayName}}{display name of the database  Should be short}

\item{\code{location}}{Storage location of user database  Might be NULL for non-user databases or if default location is used.}

\item{\code{matchRtOfReferenceSpectra}}{Indicates whether this database shall be used to use retention time information for library matching.  Typically used for in-house spectral libraries that have been measured on. Default to FALSE.}

\item{\code{dbDate}}{Date on which the data was imported / database was created.}

\item{\code{dbVersion}}{database schema version}

\item{\code{numberOfStructures}}{Number of unique compounds available in this database.}

\item{\code{numberOfFormulas}}{Number of different molecular formulas available in this database.}

\item{\code{numberOfReferenceSpectra}}{Number of reference spectra available in this database}

\item{\code{errorMessage}}{Error message if the database could not be loaded}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabase-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabase-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabase$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabase-toList"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabase-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabase$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SearchableDatabase as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of SearchableDatabase (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabase-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabase-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert SearchableDatabase to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabase$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabase-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabase-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of SearchableDatabase
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabase$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SearchableDatabase
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabase-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabase-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabase$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SearchableDatabase in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabase-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabase-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of SearchableDatabase
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabase$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SearchableDatabase
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabase-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabase-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to SearchableDatabase and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabase$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabase-toString"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabase-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabase$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of SearchableDatabase
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabase-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabase-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabase$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabase-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabase-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabase$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabase-print"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabase-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabase$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
