% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchable_databases_api.R
\docType{class}
\name{SearchableDatabasesApi}
\alias{SearchableDatabasesApi}
\title{SearchableDatabases operations}
\format{
An \code{R6Class} generator object
}
\description{
SearchableDatabasesApi
}
\details{
SIRIUS Nightsky API

REST API that provides the full functionality of SIRIUS and its web services as background service. It is intended as entry-point for scripting languages and software integration SDKs.This API is exposed by SIRIUS 6

The version of the OpenAPI document: 3.1
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  AddDatabases  ####################

library(Rsirius)
var_request_body <- c("property_example") # array[character] | 

#DEPRECATED: this endpoint is based on local file paths and will likely be replaced in future versions of this API.
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$AddDatabases(var_request_bodydata_file = "result.txt")
result <- api_instance$searchable_databases_api$AddDatabases(var_request_body)
dput(result)


####################  CreateDatabase  ####################

library(Rsirius)
var_database_id <- "database_id_example" # character | 
var_searchable_database_parameters <- SearchableDatabaseParameters$new("displayName_example", "location_example", "matchRtOfReferenceSpectra_example") # SearchableDatabaseParameters |  (Optional)

api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$CreateDatabase(var_database_id, searchable_database_parameters = var_searchable_database_parametersdata_file = "result.txt")
result <- api_instance$searchable_databases_api$CreateDatabase(var_database_id, searchable_database_parameters = var_searchable_database_parameters)
dput(result)


####################  GetCustomDatabases  ####################

library(Rsirius)
var_include_stats <- FALSE # character |  (Optional)
var_include_with_errors <- FALSE # character |  (Optional)

api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetCustomDatabases(include_stats = var_include_stats, include_with_errors = var_include_with_errorsdata_file = "result.txt")
result <- api_instance$searchable_databases_api$GetCustomDatabases(include_stats = var_include_stats, include_with_errors = var_include_with_errors)
dput(result)


####################  GetDatabase  ####################

library(Rsirius)
var_database_id <- "database_id_example" # character | 
var_include_stats <- TRUE # character |  (Optional)

api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetDatabase(var_database_id, include_stats = var_include_statsdata_file = "result.txt")
result <- api_instance$searchable_databases_api$GetDatabase(var_database_id, include_stats = var_include_stats)
dput(result)


####################  GetDatabases  ####################

library(Rsirius)
var_include_stats <- FALSE # character |  (Optional)
var_include_with_errors <- FALSE # character |  (Optional)

api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetDatabases(include_stats = var_include_stats, include_with_errors = var_include_with_errorsdata_file = "result.txt")
result <- api_instance$searchable_databases_api$GetDatabases(include_stats = var_include_stats, include_with_errors = var_include_with_errors)
dput(result)


####################  GetIncludedDatabases  ####################

library(Rsirius)
var_include_stats <- FALSE # character |  (Optional)

api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetIncludedDatabases(include_stats = var_include_statsdata_file = "result.txt")
result <- api_instance$searchable_databases_api$GetIncludedDatabases(include_stats = var_include_stats)
dput(result)


####################  ImportIntoDatabase  ####################

library(Rsirius)
var_database_id <- "database_id_example" # character | database to import into
var_buffer_size <- 1000 # integer |  (Optional)
var_input_files <- c(123) # array[data.frame] |  (Optional)

#Start import of structure and spectra files into the specified database.
api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$ImportIntoDatabase(var_database_id, buffer_size = var_buffer_size, input_files = var_input_filesdata_file = "result.txt")
result <- api_instance$searchable_databases_api$ImportIntoDatabase(var_database_id, buffer_size = var_buffer_size, input_files = var_input_files)
dput(result)


####################  RemoveDatabase  ####################

library(Rsirius)
var_database_id <- "database_id_example" # character | 
var_delete <- FALSE # character |  (Optional)

api_instance <- rsirius_api$new()

api_instance$searchable_databases_api$RemoveDatabase(var_database_id, delete = var_delete)


####################  UpdateDatabase  ####################

library(Rsirius)
var_database_id <- "database_id_example" # character | 
var_searchable_database_parameters <- SearchableDatabaseParameters$new("displayName_example", "location_example", "matchRtOfReferenceSpectra_example") # SearchableDatabaseParameters |  (Optional)

api_instance <- rsirius_api$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$UpdateDatabase(var_database_id, searchable_database_parameters = var_searchable_database_parametersdata_file = "result.txt")
result <- api_instance$searchable_databases_api$UpdateDatabase(var_database_id, searchable_database_parameters = var_searchable_database_parameters)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SearchableDatabasesApi-new}{\code{SearchableDatabasesApi$new()}}
\item \href{#method-SearchableDatabasesApi-AddDatabases}{\code{SearchableDatabasesApi$AddDatabases()}}
\item \href{#method-SearchableDatabasesApi-AddDatabasesWithHttpInfo}{\code{SearchableDatabasesApi$AddDatabasesWithHttpInfo()}}
\item \href{#method-SearchableDatabasesApi-CreateDatabase}{\code{SearchableDatabasesApi$CreateDatabase()}}
\item \href{#method-SearchableDatabasesApi-CreateDatabaseWithHttpInfo}{\code{SearchableDatabasesApi$CreateDatabaseWithHttpInfo()}}
\item \href{#method-SearchableDatabasesApi-GetCustomDatabases}{\code{SearchableDatabasesApi$GetCustomDatabases()}}
\item \href{#method-SearchableDatabasesApi-GetCustomDatabasesWithHttpInfo}{\code{SearchableDatabasesApi$GetCustomDatabasesWithHttpInfo()}}
\item \href{#method-SearchableDatabasesApi-GetDatabase}{\code{SearchableDatabasesApi$GetDatabase()}}
\item \href{#method-SearchableDatabasesApi-GetDatabaseWithHttpInfo}{\code{SearchableDatabasesApi$GetDatabaseWithHttpInfo()}}
\item \href{#method-SearchableDatabasesApi-GetDatabases}{\code{SearchableDatabasesApi$GetDatabases()}}
\item \href{#method-SearchableDatabasesApi-GetDatabasesWithHttpInfo}{\code{SearchableDatabasesApi$GetDatabasesWithHttpInfo()}}
\item \href{#method-SearchableDatabasesApi-GetIncludedDatabases}{\code{SearchableDatabasesApi$GetIncludedDatabases()}}
\item \href{#method-SearchableDatabasesApi-GetIncludedDatabasesWithHttpInfo}{\code{SearchableDatabasesApi$GetIncludedDatabasesWithHttpInfo()}}
\item \href{#method-SearchableDatabasesApi-ImportIntoDatabase}{\code{SearchableDatabasesApi$ImportIntoDatabase()}}
\item \href{#method-SearchableDatabasesApi-ImportIntoDatabaseWithHttpInfo}{\code{SearchableDatabasesApi$ImportIntoDatabaseWithHttpInfo()}}
\item \href{#method-SearchableDatabasesApi-RemoveDatabase}{\code{SearchableDatabasesApi$RemoveDatabase()}}
\item \href{#method-SearchableDatabasesApi-RemoveDatabaseWithHttpInfo}{\code{SearchableDatabasesApi$RemoveDatabaseWithHttpInfo()}}
\item \href{#method-SearchableDatabasesApi-UpdateDatabase}{\code{SearchableDatabasesApi$UpdateDatabase()}}
\item \href{#method-SearchableDatabasesApi-UpdateDatabaseWithHttpInfo}{\code{SearchableDatabasesApi$UpdateDatabaseWithHttpInfo()}}
\item \href{#method-SearchableDatabasesApi-clone}{\code{SearchableDatabasesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SearchableDatabasesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-AddDatabases"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-AddDatabases}{}}}
\subsection{Method \code{AddDatabases()}}{
DEPRECATED: this endpoint is based on local file paths and will likely be replaced in future versions of this API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$AddDatabases(request_body, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{request_body}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array[SearchableDatabase]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-AddDatabasesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-AddDatabasesWithHttpInfo}{}}}
\subsection{Method \code{AddDatabasesWithHttpInfo()}}{
DEPRECATED: this endpoint is based on local file paths and will likely be replaced in future versions of this API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$AddDatabasesWithHttpInfo(
  request_body,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{request_body}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (array[SearchableDatabase]) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-CreateDatabase"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-CreateDatabase}{}}}
\subsection{Method \code{CreateDatabase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$CreateDatabase(
  database_id,
  searchable_database_parameters = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database_id}}{}

\item{\code{searchable_database_parameters}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SearchableDatabase
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-CreateDatabaseWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-CreateDatabaseWithHttpInfo}{}}}
\subsection{Method \code{CreateDatabaseWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$CreateDatabaseWithHttpInfo(
  database_id,
  searchable_database_parameters = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database_id}}{}

\item{\code{searchable_database_parameters}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (SearchableDatabase) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-GetCustomDatabases"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-GetCustomDatabases}{}}}
\subsection{Method \code{GetCustomDatabases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$GetCustomDatabases(
  include_stats = FALSE,
  include_with_errors = FALSE,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_stats}}{(optional) No description (default value: FALSE)}

\item{\code{include_with_errors}}{(optional) No description (default value: FALSE)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array[SearchableDatabase]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-GetCustomDatabasesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-GetCustomDatabasesWithHttpInfo}{}}}
\subsection{Method \code{GetCustomDatabasesWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$GetCustomDatabasesWithHttpInfo(
  include_stats = FALSE,
  include_with_errors = FALSE,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_stats}}{(optional) No description (default value: FALSE)}

\item{\code{include_with_errors}}{(optional) No description (default value: FALSE)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (array[SearchableDatabase]) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-GetDatabase"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-GetDatabase}{}}}
\subsection{Method \code{GetDatabase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$GetDatabase(
  database_id,
  include_stats = TRUE,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database_id}}{}

\item{\code{include_stats}}{(optional) No description (default value: TRUE)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SearchableDatabase
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-GetDatabaseWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-GetDatabaseWithHttpInfo}{}}}
\subsection{Method \code{GetDatabaseWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$GetDatabaseWithHttpInfo(
  database_id,
  include_stats = TRUE,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database_id}}{}

\item{\code{include_stats}}{(optional) No description (default value: TRUE)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (SearchableDatabase) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-GetDatabases"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-GetDatabases}{}}}
\subsection{Method \code{GetDatabases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$GetDatabases(
  include_stats = FALSE,
  include_with_errors = FALSE,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_stats}}{(optional) No description (default value: FALSE)}

\item{\code{include_with_errors}}{(optional) No description (default value: FALSE)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array[SearchableDatabase]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-GetDatabasesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-GetDatabasesWithHttpInfo}{}}}
\subsection{Method \code{GetDatabasesWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$GetDatabasesWithHttpInfo(
  include_stats = FALSE,
  include_with_errors = FALSE,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_stats}}{(optional) No description (default value: FALSE)}

\item{\code{include_with_errors}}{(optional) No description (default value: FALSE)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (array[SearchableDatabase]) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-GetIncludedDatabases"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-GetIncludedDatabases}{}}}
\subsection{Method \code{GetIncludedDatabases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$GetIncludedDatabases(
  include_stats = FALSE,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_stats}}{(optional) No description (default value: FALSE)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array[SearchableDatabase]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-GetIncludedDatabasesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-GetIncludedDatabasesWithHttpInfo}{}}}
\subsection{Method \code{GetIncludedDatabasesWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$GetIncludedDatabasesWithHttpInfo(
  include_stats = FALSE,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_stats}}{(optional) No description (default value: FALSE)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (array[SearchableDatabase]) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-ImportIntoDatabase"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-ImportIntoDatabase}{}}}
\subsection{Method \code{ImportIntoDatabase()}}{
Start import of structure and spectra files into the specified database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$ImportIntoDatabase(
  database_id,
  buffer_size = 1000,
  input_files = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database_id}}{database to import into}

\item{\code{buffer_size}}{(optional) No description (default value: 1000)}

\item{\code{input_files}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SearchableDatabase
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-ImportIntoDatabaseWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-ImportIntoDatabaseWithHttpInfo}{}}}
\subsection{Method \code{ImportIntoDatabaseWithHttpInfo()}}{
Start import of structure and spectra files into the specified database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$ImportIntoDatabaseWithHttpInfo(
  database_id,
  buffer_size = 1000,
  input_files = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database_id}}{database to import into}

\item{\code{buffer_size}}{(optional) No description (default value: 1000)}

\item{\code{input_files}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (SearchableDatabase) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-RemoveDatabase"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-RemoveDatabase}{}}}
\subsection{Method \code{RemoveDatabase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$RemoveDatabase(database_id, delete = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database_id}}{}

\item{\code{delete}}{(optional) No description (default value: FALSE)}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-RemoveDatabaseWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-RemoveDatabaseWithHttpInfo}{}}}
\subsection{Method \code{RemoveDatabaseWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$RemoveDatabaseWithHttpInfo(
  database_id,
  delete = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database_id}}{}

\item{\code{delete}}{(optional) No description (default value: FALSE)}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-UpdateDatabase"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-UpdateDatabase}{}}}
\subsection{Method \code{UpdateDatabase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$UpdateDatabase(
  database_id,
  searchable_database_parameters = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database_id}}{}

\item{\code{searchable_database_parameters}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SearchableDatabase
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-UpdateDatabaseWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-UpdateDatabaseWithHttpInfo}{}}}
\subsection{Method \code{UpdateDatabaseWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$UpdateDatabaseWithHttpInfo(
  database_id,
  searchable_database_parameters = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database_id}}{}

\item{\code{searchable_database_parameters}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (SearchableDatabase) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchableDatabasesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SearchableDatabasesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchableDatabasesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
