% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_candidate_formula.R
\docType{class}
\name{StructureCandidateFormula}
\alias{StructureCandidateFormula}
\title{StructureCandidateFormula}
\format{
An \code{R6Class} generator object
}
\description{
StructureCandidateFormula Class

StructureCandidateFormula Class
}
\details{
Create a new StructureCandidateFormula
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{inchiKey}}{character [optional]}

\item{\code{smiles}}{character [optional]}

\item{\code{structureName}}{character [optional]}

\item{\code{xlogP}}{numeric [optional]}

\item{\code{dbLinks}}{List of structure database links belonging to this structure candidate  OPTIONAL: needs to be added by parameter list(\link{DBLink}) [optional]}

\item{\code{spectralLibraryMatches}}{List of spectral library matches belonging to this structure candidate  OPTIONAL: needs to be added by parameter list(\link{SpectralLibraryMatch}) [optional]}

\item{\code{rank}}{the overall rank of this candidate among all candidates of this feature integer [optional]}

\item{\code{csiScore}}{CSI:FingerID score of the fingerprint of this compound to the predicted fingerprint of CSI:FingerID  This is the score used for ranking structure candidates numeric [optional]}

\item{\code{tanimotoSimilarity}}{Tanimoto similarly of the fingerprint of this compound to the predicted fingerprint of CSI:FingerID numeric [optional]}

\item{\code{mcesDistToTopHit}}{Maximum Common Edge Subgraph (MCES) distance to the top scoring hit (CSI:FingerID) in a candidate list. numeric [optional]}

\item{\code{fingerprint}}{\link{BinaryFingerprint} [optional]}

\item{\code{molecularFormula}}{Molecular formula of this candidate character [optional]}

\item{\code{adduct}}{Adduct of this candidate character [optional]}

\item{\code{formulaId}}{Id of the corresponding Formula candidate character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StructureCandidateFormula-new}{\code{StructureCandidateFormula$new()}}
\item \href{#method-StructureCandidateFormula-toJSON}{\code{StructureCandidateFormula$toJSON()}}
\item \href{#method-StructureCandidateFormula-fromJSON}{\code{StructureCandidateFormula$fromJSON()}}
\item \href{#method-StructureCandidateFormula-toJSONString}{\code{StructureCandidateFormula$toJSONString()}}
\item \href{#method-StructureCandidateFormula-fromJSONString}{\code{StructureCandidateFormula$fromJSONString()}}
\item \href{#method-StructureCandidateFormula-validateJSON}{\code{StructureCandidateFormula$validateJSON()}}
\item \href{#method-StructureCandidateFormula-toString}{\code{StructureCandidateFormula$toString()}}
\item \href{#method-StructureCandidateFormula-isValid}{\code{StructureCandidateFormula$isValid()}}
\item \href{#method-StructureCandidateFormula-getInvalidFields}{\code{StructureCandidateFormula$getInvalidFields()}}
\item \href{#method-StructureCandidateFormula-print}{\code{StructureCandidateFormula$print()}}
\item \href{#method-StructureCandidateFormula-clone}{\code{StructureCandidateFormula$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateFormula-new"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateFormula-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new StructureCandidateFormula class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateFormula$new(
  inchiKey = NULL,
  smiles = NULL,
  structureName = NULL,
  xlogP = NULL,
  dbLinks = NULL,
  spectralLibraryMatches = NULL,
  rank = NULL,
  csiScore = NULL,
  tanimotoSimilarity = NULL,
  mcesDistToTopHit = NULL,
  fingerprint = NULL,
  molecularFormula = NULL,
  adduct = NULL,
  formulaId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inchiKey}}{inchiKey}

\item{\code{smiles}}{smiles}

\item{\code{structureName}}{structureName}

\item{\code{xlogP}}{xlogP}

\item{\code{dbLinks}}{List of structure database links belonging to this structure candidate  OPTIONAL: needs to be added by parameter}

\item{\code{spectralLibraryMatches}}{List of spectral library matches belonging to this structure candidate  OPTIONAL: needs to be added by parameter}

\item{\code{rank}}{the overall rank of this candidate among all candidates of this feature}

\item{\code{csiScore}}{CSI:FingerID score of the fingerprint of this compound to the predicted fingerprint of CSI:FingerID  This is the score used for ranking structure candidates}

\item{\code{tanimotoSimilarity}}{Tanimoto similarly of the fingerprint of this compound to the predicted fingerprint of CSI:FingerID}

\item{\code{mcesDistToTopHit}}{Maximum Common Edge Subgraph (MCES) distance to the top scoring hit (CSI:FingerID) in a candidate list.}

\item{\code{fingerprint}}{fingerprint}

\item{\code{molecularFormula}}{Molecular formula of this candidate}

\item{\code{adduct}}{Adduct of this candidate}

\item{\code{formulaId}}{Id of the corresponding Formula candidate}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateFormula-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateFormula-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateFormula$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StructureCandidateFormula in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateFormula-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateFormula-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of StructureCandidateFormula
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateFormula$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StructureCandidateFormula
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateFormula-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateFormula-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateFormula$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StructureCandidateFormula in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateFormula-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateFormula-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of StructureCandidateFormula
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateFormula$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StructureCandidateFormula
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateFormula-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateFormula-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to StructureCandidateFormula and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateFormula$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateFormula-toString"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateFormula-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateFormula$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of StructureCandidateFormula
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateFormula-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateFormula-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateFormula$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateFormula-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateFormula-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateFormula$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateFormula-print"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateFormula-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateFormula$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StructureCandidateFormula-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StructureCandidateFormula-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StructureCandidateFormula$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
